refactor(lsp-tests): split testFoamLSP.js into per-category modules
The entrypoint had grown to ~3.4k lines. Splitting keeps the aggregated
suite intact while letting each category be read and run in isolation.

- tools/tests/testFoamLSP.js — 48-line boot + require-in-order aggregator
- tools/tests/lsp/_harness.js — pmake boot + counters + shared instances
  (index, grammar, cache, typeTracker, analyzer, handlers, resolvers)
- tools/tests/lsp/{foamIndex,grammar,utilities,completion,hover,
  diagnostics,navigation,java,jrl}.js — per-feature sections pulled from
  the original file; each reads its own copy of the shared harness

Category files declare only their local instances; every identifier
that used to be defined in one section and referenced hundreds of lines
later (cache, typeTracker, analyzer, defHandler, semanticHandler, SFV)
now lives on the harness. __dirname-relative paths inside moved blocks
get an extra '../' so they still resolve against tools/tests/.

Aggregate run: 567 passed, 0 failed — same as the pre-split suite.
