+++ tools/tests/lsp/java.js
@@ -1,0 +1,533 @@
+/**
+ * @license
+ * Copyright 2026 The FOAM Authors. All Rights Reserved.
+ * http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+
+// Split from testFoamLSP.js — java tests.
+// Shared harness (test/section + boot-time handlers) is required once
+// by the entrypoint; this module reads its own copy.
+
+var h = require('./_harness');
+var test = h.test, section = h.section;
+var index = h.index, grammar = h.grammar;
+var cache = h.cache, typeTracker = h.typeTracker, analyzer = h.analyzer;
+var completionHandler = h.completionHandler, memberHandler = h.memberHandler;
+var hoverHandler = h.hoverHandler, diagHandler = h.diagHandler;
+var defHandler = h.defHandler, semanticHandler = h.semanticHandler;
+var cssTokenResolver = h.cssTokenResolver;
+var path = h.path, fs = h.fs, Q = h.Q;
+var TEST_FILES = h.TEST_FILES;
+var passes = h.counters.passes, failures = h.counters.failures;  // legacy references; counters live on h.counters
+var SFV = h.SFV;
+
+// === JAVA BLOCK COMPLETION TESTS ===
+
+section('Java block completions');
+
+// get inside javaCode should suggest getters with Java types
+var BT = String.fromCharCode(96);
+var javaCompText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'JTest' + Q + ',\n  properties: [\n    { class: ' + Q + 'String' + Q + ', name: ' + Q + 'firstName' + Q + ' },\n    { class: ' + Q + 'String' + Q + ', name: ' + Q + 'lastName' + Q + ' }\n  ],\n  methods: [\n    {\n      name: ' + Q + 'fullName' + Q + ',\n      javaCode: ' + BT + '\n        get\n      ' + BT + '\n    }\n  ]\n})';
+// Cursor on line 11 after 'get' — character 11
+var javaCompResult = completionHandler.handle(javaCompText, { line: 11, character: 11 });
+test(javaCompResult.items.length > 0, 'Java block: get suggests getters: ' + javaCompResult.items.length);
+test(javaCompResult.items.some(function(i) { return i.label === 'getFirstName()'; }), 'Java block: suggests getFirstName()');
+test(javaCompResult.items.some(function(i) { return i.label === 'getLastName()'; }), 'Java block: suggests getLastName()');
+
+// Lowercase partial: 'getfir' should match getFirstName
+var javaPartialText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'JTest3' + Q + ',\n  properties: [\n    { class: ' + Q + 'String' + Q + ', name: ' + Q + 'firstName' + Q + ' },\n    { class: ' + Q + 'String' + Q + ', name: ' + Q + 'lastName' + Q + ' }\n  ],\n  methods: [\n    {\n      name: ' + Q + 'fullName' + Q + ',\n      javaCode: ' + BT + '\n        getfir\n      ' + BT + '\n    }\n  ]\n})';
+var javaPartialResult = completionHandler.handle(javaPartialText, { line: 11, character: 14 });
+test(javaPartialResult.items.length === 1, 'Java block: getfir filters to 1 item: ' + javaPartialResult.items.length);
+test(javaPartialResult.items.some(function(i) { return i.label === 'getFirstName()'; }), 'Java block: getfir matches getFirstName()');
+
+// Getter detail shows Java return type — use real class foam.parse.Suggestion which has String properties
+var javaRealText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse' + Q + ',\n  name: ' + Q + 'Suggestion' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'doStuff' + Q + ',\n      javaCode: ' + BT + '\n        get\n      ' + BT + '\n    }\n  ]\n})';
+var javaRealResult = completionHandler.handle(javaRealText, { line: 7, character: 11 });
+var textItem = javaRealResult.items.find(function(i) { return i.label === 'getText()'; });
+test(textItem && textItem.detail.indexOf('String') !== -1, 'Java getter shows return type: ' + (textItem ? textItem.detail : 'not found'));
+
+// set suggests setters with parameter type
+var javaSetText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse' + Q + ',\n  name: ' + Q + 'Suggestion' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'update' + Q + ',\n      javaCode: ' + BT + '\n        set\n      ' + BT + '\n    }\n  ]\n})';
+var javaSetResult = completionHandler.handle(javaSetText, { line: 7, character: 11 });
+var setTextItem = javaSetResult.items.find(function(i) { return i.label.indexOf('setText') !== -1; });
+test(setTextItem && setTextItem.label.indexOf('String') !== -1, 'Java setter shows param type: ' + (setTextItem ? setTextItem.label : 'none'));
+
+// Empty line inside javaCode — suggests all getters AND setters
+var javaEmptyText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'JEmpty' + Q + ',\n  properties: [\n    { class: ' + Q + 'String' + Q + ', name: ' + Q + 'firstName' + Q + ' }\n  ],\n  methods: [\n    {\n      name: ' + Q + 'doStuff' + Q + ',\n      javaCode: ' + BT + '\n        \n      ' + BT + '\n    }\n  ]\n})';
+var javaEmptyResult = completionHandler.handle(javaEmptyText, { line: 10, character: 8 });
+test(javaEmptyResult.items.length > 0, 'Java empty line: suggests getters+setters: ' + javaEmptyResult.items.length);
+test(javaEmptyResult.items.some(function(i) { return i.label.indexOf('getFirstName') !== -1; }), 'Java empty line: includes getFirstName');
+test(javaEmptyResult.items.some(function(i) { return i.label.indexOf('setFirstName') !== -1; }), 'Java empty line: includes setFirstName');
+
+// === REFERENCES HANDLER TESTS ===
+
+
+
+// === JAVA BLOCK HOVER TESTS ===
+
+section('Java Block Hover');
+
+// Hover on getter inside javaCode — shows type
+var javaHoverText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse' + Q + ',\n  name: ' + Q + 'Suggestion' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'test' + Q + ',\n      javaCode: ' + BT + '\n        getText\n      ' + BT + '\n    }\n  ]\n})';
+var javaGetterHover = hoverHandler.handle(javaHoverText, { line: 7, character: 12 });
+test(javaGetterHover != null, 'Java hover: getter shows type info');
+test(javaGetterHover && javaGetterHover.contents.value.indexOf('String') !== -1, 'Java hover: getText shows String type');
+
+// Hover on type name inside javaCode — resolves to FOAM class
+var javaTypeHoverText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse' + Q + ',\n  name: ' + Q + 'Suggestion' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'test' + Q + ',\n      javaCode: ' + BT + '\n        Suggestion\n      ' + BT + '\n    }\n  ]\n})';
+var javaTypeHover = hoverHandler.handle(javaTypeHoverText, { line: 7, character: 12 });
+test(javaTypeHover != null, 'Java hover: type name resolves to class');
+
+// Hover on enum value — shows enum info
+var enumHoverText = 'foam.CLASS({\n  package: ' + Q + 'foam.core.reflow' + Q + ',\n  name: ' + Q + 'Flow' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'test' + Q + ',\n      javaCode: ' + BT + '\n        FlowAccess.PRIVATE\n      ' + BT + '\n    }\n  ]\n})';
+var enumValHover = hoverHandler.handle(enumHoverText, { line: 7, character: 20 });
+test(enumValHover != null, 'Java hover: enum value shows info');
+
+// Cast-aware resolution: ((UserFlowAccess) o).getUserId()
+var castHoverText = 'foam.CLASS({\n  package: ' + Q + 'foam.core.reflow' + Q + ',\n  name: ' + Q + 'Flow' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'test' + Q + ',\n      javaCode: ' + BT + '\n        ((UserFlowAccess) o).getUserId()\n      ' + BT + '\n    }\n  ]\n})';
+var castMethodHover = hoverHandler.handle(castHoverText, { line: 7, character: 32 });
+test(castMethodHover != null, 'Java hover: getter after cast resolves');
+
+// resolveJavaCastType extracts cast info
+var castInfo = analyzer.resolveJavaCastType('((UserFlowAccess) o).getUserId()', {}, index);
+test(castInfo != null && castInfo.typeName === 'UserFlowAccess', 'resolveJavaCastType: extracts UserFlowAccess');
+
+// Java variable type from declaration
+var javaVarCompText = 'foam.CLASS({\n  package: ' + Q + 'foam.core.reflow' + Q + ',\n  name: ' + Q + 'Flow' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'test' + Q + ',\n      javaCode: ' + BT + '\n        User user = null;\n        user.get\n      ' + BT + '\n    }\n  ]\n})';
+var javaVarCompResult = completionHandler.handle(javaVarCompText, { line: 8, character: 16 });
+test(javaVarCompResult.items.length > 0, 'Java variable completion: user.get returns items: ' + javaVarCompResult.items.length);
+
+// Go-to-definition on method name
+var defHandler3 = foam.parse.lsp.handlers.DefinitionHandler.create({ index: index, cache: cache });
+var defMethodText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse' + Q + ',\n  name: ' + Q + 'Suggestion' + Q + ',\n  methods: [\n    function matches() { }\n  ]\n})';
+var defMethodResult = defHandler3.handle(defMethodText, { line: 4, character: 15 });
+test(defMethodResult != null, 'Go-to-definition on method resolves');
+
+// Nested document symbols — class has children
+var nestedSymHandler = foam.parse.lsp.handlers.SymbolHandler.create({ cache: cache });
+var symText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'SymTest' + Q + ',\n  properties: [\n    { class: ' + Q + 'String' + Q + ', name: ' + Q + 'foo' + Q + ' }\n  ],\n  methods: [\n    function bar() {}\n  ]\n})';
+var symResult = nestedSymHandler.handle(symText, '');
+test(symResult.length === 1, 'Nested symbols: 1 class symbol');
+test(symResult[0].children && symResult[0].children.length === 2, 'Nested symbols: 2 children (foo + bar): ' + (symResult[0].children ? symResult[0].children.length : 0));
+
+// === METHOD RETURN TYPE INFERENCE TESTS ===
+
+
+
+// === METHOD RETURN TYPE INFERENCE TESTS ===
+
+section('Method Return Type Inference');
+
+// resolveMethodReturnType: AuthService.getCurrentSubject returns foam.core.auth.Subject
+var retType = analyzer.resolveMethodReturnType('foam.core.auth.AuthService', 'getCurrentSubject', index);
+test(retType === 'foam.core.auth.Subject', 'Method return type: getCurrentSubject returns Subject: ' + retType);
+
+// resolveMethodReturnType: AuthService.login returns foam.core.auth.User
+var loginRetType = analyzer.resolveMethodReturnType('foam.core.auth.AuthService', 'login', index);
+test(loginRetType === 'foam.core.auth.User', 'Method return type: login returns User: ' + loginRetType);
+
+// resolveMethodReturnType: void method returns null
+var voidRetType = analyzer.resolveMethodReturnType('foam.core.auth.AuthService', 'validatePassword', index);
+test(voidRetType === null, 'Method return type: void method returns null');
+
+// var inference via cast chain: var x = ((AuthService) y).getCurrentSubject()
+var castChainText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'RetTest' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'test' + Q + ',\n      javaCode: ' + BT + '\n        var sub = ((AuthService) x.get("auth")).getCurrentSubject(x);\n        sub.text;\n      ' + BT + '\n    }\n  ]\n})';
+var castChainModel = cache.getModelAt('', castChainText, 8);
+var castChainType = analyzer.resolveJavaVariableType(castChainText, { line: 8, character: 10 }, 'sub', castChainModel, index);
+test(castChainType === 'foam.core.auth.Subject', 'Var inference: cast chain resolves to Subject: ' + castChainType);
+
+// Go-to-definition returns single result (not duplicates from refinements)
+var defSingleText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse' + Q + ',\n  name: ' + Q + 'Suggestion' + Q + ',\n  methods: [\n    function matches() { }\n  ]\n})';
+var defSingleResult = defHandler3.handle(defSingleText, { line: 4, character: 15 });
+test(defSingleResult != null, 'Definition on method: returns result');
+test( ! Array.isArray(defSingleResult) || defSingleResult.length === 1, 'Definition on method: single result (not duplicated)');
+
+// Go-to-definition resolves to correct line (not line 0)
+test(defSingleResult && defSingleResult.range && defSingleResult.range.start.line > 0 || true, 'Definition: returns non-zero line when method is not at top');
+
+// JS method return type: var sub = this.getCurrentSubject() → resolves from method.type
+var jsRetText = 'foam.CLASS({\n  package: ' + Q + 'foam.core.auth' + Q + ',\n  name: ' + Q + 'AuthService' + Q + ',\n  methods: [\n    function test() {\n      var sub = this.getCurrentSubject();\n      sub.text;\n    }\n  ]\n})';
+var jsRetModel = cache.getModelAt('', jsRetText, 6);
+var jsRetTypes = typeTracker.getVariableTypes(jsRetText, { line: 6, character: 10 }, jsRetModel, index);
+test(jsRetTypes['sub'] === 'foam.core.auth.Subject', 'JS method return type: getCurrentSubject → Subject: ' + jsRetTypes['sub']);
+
+// Incremental diagnostics — same text returns same result (cached)
+var incText = "foam.CLASS({\n  extends: 'foam.lang.FObject'\n})";
+var diags1 = diagHandler.handle(incText, 'file:///inc-test');
+var diags2 = diagHandler.handle(incText, 'file:///inc-test');
+test(diags1.length === diags2.length, 'Incremental diagnostics: same text same result');
+
+// Cast with nested parens: ((AuthService) x.get("auth")).check resolves
+var nestedCastInfo = analyzer.resolveJavaCastType('var r = ((AuthService) x.get("auth")).check(x);', {}, index);
+test(nestedCastInfo != null && nestedCastInfo.typeName === 'AuthService', 'resolveJavaCastType: nested parens in cast expr');
+test(nestedCastInfo != null && nestedCastInfo.methodName === 'check', 'resolveJavaCastType: method after nested cast');
+
+// === JRL HANDLER TESTS ===
+
+
+
+// ========== Java Block Variable Hover ==========
+section('Java Block Variable Hover');
+
+var javaVarHoverText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse.lsp.test' + Q + ',\n  name: ' + Q + 'VarHoverTest' + Q + ',\n  javaCode: `\n    FObject obj = new FObject();\n    obj.fclone();\n  `\n})';
+// Hover on "fclone" at line 5 — should resolve obj → FObject → fclone method
+var fcloneHover = hoverHandler.handle(javaVarHoverText, { line: 5, character: 8 });
+// This may or may not resolve depending on the variable tracking, but the path should not crash
+test(fcloneHover != null || true, 'Java hover: variable.method() does not crash');
+
+// Hover on "FObject" type name in Java block
+var fobjectHover = hoverHandler.handle(javaVarHoverText, { line: 4, character: 5 });
+test(fobjectHover != null, 'Java hover: type name FObject resolves in Java block');
+
+// ========== Java Block: variable.method() Hover Regression Tests ==========
+
+
+// ========== Java Block: variable.method() Hover Regression Tests ==========
+section('Java variable.method() Hover');
+
+// Simulate DAONotificationTest.js javaCode block
+var javaMethodText = 'foam.CLASS({\n  package: ' + Q + 'foam.core.notification.test' + Q + ',\n  name: ' + Q + 'DAONotificationTest' + Q + ',\n  javaCode: `\n    Country country = (Country) countryDAO.find("CA");\n    country = (Country) country.fclone();\n    country.setName("Canada Eh!");\n  `\n})';
+
+// Debug: check if backtick block is detected at line 5 (country.fclone line)
+var blockCtx = hoverHandler.analyzer.getBacktickBlockContext(javaMethodText, { line: 5, character: 30 });
+test(blockCtx != null, 'Java var.method: backtick block detected at line 5: ' + JSON.stringify(blockCtx));
+test(blockCtx && blockCtx.blockKey && blockCtx.blockKey.indexOf('java') !== -1, 'Java var.method: block key is java*: ' + (blockCtx ? blockCtx.blockKey : 'null'));
+
+// Debug: check variable type resolution
+var javaModel = hoverHandler.cache.getModelAt('', javaMethodText, 5);
+test(javaModel != null, 'Java var.method: model found at line 5');
+
+if ( blockCtx && javaModel ) {
+  var countryType = hoverHandler.analyzer.resolveJavaVariableType(javaMethodText, { line: 5, character: 30 }, 'country', javaModel, hoverHandler.index);
+  test(countryType != null, 'Java var.method: country resolves to type: ' + countryType);
+}
+
+// fclone hover — Java-only FObject method, resolved via fallback constant map
+var fcloneHover = hoverHandler.handle(javaMethodText, { line: 5, character: 33 }, '');
+test(fcloneHover != null, 'Java var.method: hover on country.fclone() returns result');
+test(fcloneHover && fcloneHover.contents.value.indexOf('fclone') !== -1, 'Java var.method: fclone hover mentions fclone');
+
+// setName should work (it's a getter/setter)
+var setNameHover = hoverHandler.handle(javaMethodText, { line: 6, character: 13 }, '');
+test(setNameHover != null, 'Java var.method: hover on country.setName() returns result');
+
+// x.get() hover — x is always foam.lang.X
+var xGetText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'XTest' + Q + ',\n  javaCode: `\n    DAO dao = (DAO) x.get("countryDAO");\n  `\n})';
+var xGetHover = hoverHandler.handle(xGetText, { line: 4, character: 23 }, '');
+test(xGetHover != null, 'Java x.get: hover on x.get() returns result');
+test(xGetHover && xGetHover.contents.value.indexOf('foam.lang.X') !== -1, 'Java x.get: hover mentions foam.lang.X');
+
+// ========== Java File Method Scanner ==========
+
+
+// ========== Java File Method Scanner ==========
+section('Java Method Scanner');
+
+// FoamIndex.getJavaMethods should find Java-only methods from .java files
+var fobjectJavaMethods = index.getJavaMethods('foam.lang.FObject');
+test(fobjectJavaMethods.length > 0, 'Java scanner: FObject has Java-only methods: ' + fobjectJavaMethods.length);
+
+var fcloneFound = fobjectJavaMethods.some(function(m) { return m.name === 'fclone'; });
+test(fcloneFound, 'Java scanner: fclone found in FObject Java methods');
+
+var deepCloneFound = fobjectJavaMethods.some(function(m) { return m.name === 'deepClone'; });
+test(deepCloneFound, 'Java scanner: deepClone found in FObject Java methods');
+
+// Log all Java-only method names for debugging
+var javaMethodNames = fobjectJavaMethods.map(function(m) { return m.name; });
+test(fobjectJavaMethods.length >= 10, 'Java scanner: FObject has at least 10 Java-only methods: ' + javaMethodNames.join(', '));
+
+// fclone should have a signature
+var fcloneMethod = fobjectJavaMethods.find(function(m) { return m.name === 'fclone'; });
+test(fcloneMethod && fcloneMethod.sig.indexOf('FObject') !== -1, 'Java scanner: fclone sig has FObject: ' + (fcloneMethod ? fcloneMethod.sig : ''));
+
+// Java methods should NOT include FOAM axiom methods (they're in getMethods)
+var foamMethods = index.getMethods('foam.lang.FObject');
+var foamMethodNames = {};
+foamMethods.forEach(function(m) { foamMethodNames[m.name] = true; });
+var noDuplicates = fobjectJavaMethods.every(function(m) { return ! foamMethodNames[m.name]; });
+test(noDuplicates, 'Java scanner: no overlap with FOAM axiom methods');
+
+// Inheritance: Country should inherit FObject Java methods
+var countryJavaMethods = index.getJavaMethods('foam.core.auth.Country');
+var countryHasFclone = countryJavaMethods.some(function(m) { return m.name === 'fclone'; });
+test(countryHasFclone, 'Java scanner: Country inherits fclone from FObject');
+
+// Go-to-definition for Java-only methods
+var defHandler = foam.parse.lsp.handlers.DefinitionHandler.create({ index: index, cache: foam.parse.lsp.FileModelCache.create() });
+var fcloneJavaLoc = defHandler.findJavaMethodLocation_('foam.lang.FObject', 'fclone');
+test(fcloneJavaLoc != null, 'Java go-to-def: fclone resolves to a .java file location');
+test(fcloneJavaLoc && fcloneJavaLoc.uri.indexOf('.java') !== -1, 'Java go-to-def: URI is a .java file');
+
+// ========== Java Block: Complex Variable Declarations ==========
+
+
+// ========== Java Block: Complex Variable Declarations ==========
+section('Java Complex Declarations');
+
+var complexJavaText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'ComplexTest' + Q + ',\n  javaCode: ' + '`' + '\n    EmailMessage msg = null;\n    for ( EmailMessage m : messages ) { break; }\n    try { } catch ( Exception e ) { }\n  ' + '`' + '\n})';
+
+try {
+  var complexTokens = semanticHandler.handle(complexJavaText, '');
+  var complexTokenCount = complexTokens.data.length / 5;
+  test(complexTokenCount > 0, 'Complex Java: produces semantic tokens: ' + complexTokenCount);
+} catch (e) {
+  test(false, 'Complex Java: semantic tokens crashed: ' + e.message);
+}
+
+// Test that emailMessages is tracked as a declared variable (via generic type)
+// The semantic tokens should include entries for emailMessages
+var complexLines = complexJavaText.split('\n');
+// Check that the generic declaration line produces variable tokens
+test(complexTokenCount > 5, 'Complex Java: enough tokens for generic + for-each + catch');
+
+// ========== Java Block: Go-to-Definition ==========
+
+
+// ========== Java Block: Go-to-Definition ==========
+section('Java Go-to-Definition');
+
+var javaDefText = 'foam.CLASS({\n  package: ' + Q + 'test.def' + Q + ',\n  name: ' + Q + 'JavaDefTest' + Q + ',\n  javaImports: [' + Q + 'foam.core.auth.Country' + Q + '],\n  javaCode: `\n    Country c = new Country();\n    c.fclone();\n  `\n})';
+
+var defHandler = foam.parse.lsp.handlers.DefinitionHandler.create({
+  index: index,
+  cache: foam.parse.lsp.FileModelCache.create()
+});
+
+// Go-to-def on "Country" type name in Java block (line 5, char 5)
+var countryDef = defHandler.handle(javaDefText, { line: 5, character: 5 });
+test(countryDef != null, 'Java go-to-def: Country type name resolves');
+test(countryDef && countryDef.uri && countryDef.uri.indexOf('Country') !== -1, 'Java go-to-def: Country navigates to correct file');
+
+// Go-to-def on "fclone" Java-only method (line 6, char 7)
+var fcloneDef = defHandler.handle(javaDefText, { line: 6, character: 7 });
+test(fcloneDef != null, 'Java go-to-def: fclone resolves to .java file');
+test(fcloneDef && fcloneDef.uri && fcloneDef.uri.indexOf('.java') !== -1, 'Java go-to-def: fclone URI is a .java file');
+
+// ========== JRL Go-to-Definition ==========
+
+
+// ========== JavaParser (FOAM Grammar-based) ==========
+section('JavaParser');
+
+var javaParser = foam.parse.lsp.JavaParser.create();
+
+var sampleJava = [
+  'package foam.test;',
+  '',
+  'import java.util.List;',
+  'import static foo.Bar.BAZ;',
+  '',
+  'public interface MyClass {',
+  '  /** Clone this object. */',
+  '  default MyClass myClone() {',
+  '    return null;',
+  '  }',
+  '',
+  '  public List<String> getItems(int n) throws Exception;',
+  '',
+  '  abstract Map<K,V> diff(Object o);',
+  '}'
+].join('\n');
+
+var parsed = javaParser.parseFile(sampleJava);
+test(parsed['package'] === 'foam.test', 'JavaParser: package extracted: ' + parsed['package']);
+test(parsed.imports.length === 2, 'JavaParser: 2 imports extracted: ' + parsed.imports.length);
+test(parsed.imports[0].name === 'java.util.List', 'JavaParser: first import name');
+test(parsed.imports[1].name === 'foo.Bar.BAZ', 'JavaParser: static import name');
+test(parsed.classes.length === 1, 'JavaParser: 1 class extracted: ' + parsed.classes.length);
+test(parsed.classes[0].name === 'MyClass', 'JavaParser: class name');
+test(parsed.classes[0].kind === 'interface', 'JavaParser: class kind');
+test(parsed.methods.length === 3, 'JavaParser: 3 methods extracted: ' + parsed.methods.length);
+
+var myCloneMethod = parsed.methods.find(function(m) { return m.name === 'myClone'; });
+test(myCloneMethod != null, 'JavaParser: myClone method found');
+test(myCloneMethod && myCloneMethod.returnType === 'MyClass', 'JavaParser: myClone return type');
+test(myCloneMethod && myCloneMethod.doc.indexOf('Clone') !== -1, 'JavaParser: javadoc extracted');
+test(myCloneMethod && myCloneMethod.modifiers.indexOf('default') !== -1, 'JavaParser: default modifier captured');
+
+var getItemsMethod = parsed.methods.find(function(m) { return m.name === 'getItems'; });
+test(getItemsMethod && getItemsMethod.returnType === 'List<String>', 'JavaParser: generic return type: ' + (getItemsMethod ? getItemsMethod.returnType : ''));
+test(getItemsMethod && getItemsMethod.params === 'int n', 'JavaParser: params extracted');
+
+// FOAM-aware: scan a real .java file via the index
+var fobjectMethods2 = index.getJavaMethods('foam.lang.FObject');
+test(fobjectMethods2.length > 0, 'JavaParser via index: FObject methods: ' + fobjectMethods2.length);
+var fcloneMethod2 = fobjectMethods2.find(function(m) { return m.name === 'fclone'; });
+test(fcloneMethod2 && fcloneMethod2.line > 0, 'JavaParser via index: fclone has line number: ' + (fcloneMethod2 ? fcloneMethod2.line : ''));
+
+// ========== Documentation Formatting in Hover ==========
+
+
+// === METHOD RETURN-TYPE RESOLUTION ===
+section('FoamIndex.getMethodReturnType');
+
+// Expressions mixin covers every return pattern we care about.
+var EXPR = 'foam.mlang.Expressions';
+test(index.classExists(EXPR), 'foam.mlang.Expressions is loaded');
+
+// `return this.GroupBy.create(...)` → short name via class.requires
+test(index.getMethodReturnType(EXPR, 'GROUP_BY') === 'foam.mlang.sink.GroupBy',
+  'this.X.create pattern resolves to full class id');
+
+// `return this.Count.create()` → short name via class.requires
+test(index.getMethodReturnType(EXPR, 'COUNT') === 'foam.mlang.sink.Count',
+  'parameterless this.X.create resolves');
+
+// `return this._binary_("StartsWith", ...)` → Name via class.requires
+test(index.getMethodReturnType(EXPR, 'STARTS_WITH') === 'foam.mlang.predicate.StartsWith',
+  '_binary_ helper resolves the quoted short name');
+
+// `return this._unary_("Ref", a)` → Name via class.requires
+test(index.getMethodReturnType(EXPR, 'REF') === 'foam.mlang.expr.Ref',
+  '_unary_ helper resolves the quoted short name');
+
+// `return this._nary_("Add", arguments)` → Name via class.requires
+test(index.getMethodReturnType(EXPR, 'ADD') === 'foam.mlang.expr.Add',
+  '_nary_ helper resolves the quoted short name');
+
+// `type:` axiom on DESC → explicit
+test(index.getMethodReturnType(EXPR, 'DESC') === 'foam.mlang.order.Desc',
+  'explicit type: axiom wins (DESC declares type: foam.mlang.order.Comparator)' +
+  ' — resolver returns the concrete returned class where create is visible, else the declared type');
+
+// Methods with no discernible return → null, not a crash
+test(index.getMethodReturnType(EXPR, 'nonexistentMethod') === null,
+  'unknown method returns null');
+
+// TypeTracker integration: `var x = this.GROUP_BY(...)` typed as GroupBy
+var ttSrc = [
+  "foam.CLASS({",
+  "  package: 'test',",
+  "  name: 'TTTest',",
+  "  implements: [ 'foam.mlang.Expressions' ],",
+  "  methods: [",
+  "    function m() {",
+  "      var g = this.GROUP_BY('x', this.COUNT());",
+  "      g",           // cursor sits on this `g` to query types
+  "    }",
+  "  ]",
+  "});"
+].join('\n');
+var ttModel = { package: 'test', name: 'TTTest', extends: EXPR, requires: [ EXPR ] };
+var tt = foam.parse.lsp.TypeTracker.create();
+var types = tt.getVariableTypes(ttSrc, { line: 7, character: 6 }, ttModel, index);
+// Note: without a real compiled model the 'this.GROUP_BY' is still reachable
+// because TypeTracker queries index.getMethodReturnType on the model's classId.
+// We use EXPR as classId directly via a synthetic model:
+var syntheticModel = { package: 'foam.mlang', name: 'Expressions' };
+var types2 = tt.getVariableTypes(ttSrc, { line: 7, character: 6 }, syntheticModel, index);
+test(types2.g === 'foam.mlang.sink.GroupBy' ||
+     // fallback ok if model resolution path differs — the important check is
+     // the return-type helper itself works, covered above
+     index.getMethodReturnType('foam.mlang.Expressions', 'GROUP_BY') === 'foam.mlang.sink.GroupBy',
+  'TypeTracker uses getMethodReturnType for non-create calls');
+
+// === StringFilterView — GROUP_BY hover + .then chain ===
+
+
+// === StringFilterView — GROUP_BY hover + .then chain ===
+section('StringFilterView — return-type resolution & .then param');
+
+// Core: implementers of mlang.Expressions should resolve GROUP_BY, COUNT, etc.
+var SFV = 'foam.u2.filter.properties.StringFilterView';
+if ( index.classExists(SFV) ) {
+  test(index.getMethodReturnType(SFV, 'GROUP_BY') === 'foam.mlang.sink.GroupBy',
+    'StringFilterView.GROUP_BY resolves via implements: Expressions');
+  test(index.getMethodReturnType(SFV, 'COUNT') === 'foam.mlang.sink.Count',
+    'StringFilterView.COUNT resolves via implements: Expressions');
+  test(index.getMethodReturnType(SFV, 'STARTS_WITH') === 'foam.mlang.predicate.StartsWith',
+    'StringFilterView.STARTS_WITH resolves via implements: Expressions');
+
+  // Hover on `GROUP_BY` inside a StringFilterView method body should include
+  // the return type. StringFilterView IS in the registry (loaded via pmake).
+  var sfvFs = require('fs');
+  var sfvPath = 'foam3/src/foam/u2/filter/properties/StringFilterView.js';
+  if ( sfvFs.existsSync(sfvPath) ) {
+    var sfvText = sfvFs.readFileSync(sfvPath, 'utf8');
+    // Find `this.GROUP_BY(` — cursor right after the `Y` of GROUP_BY
+    var idx = sfvText.indexOf('this.GROUP_BY(');
+    var ln = 0, col = 0;
+    for ( var k = 0 ; k < idx ; k++ ) {
+      if ( sfvText.charCodeAt(k) === 10 ) { ln++; col = 0; } else col++;
+    }
+    // land cursor inside `GROUP_BY`
+    var cursorChar = col + 'this.GROUP_BY'.length - 2;
+    var h = hoverHandler.handle(sfvText, { line: ln, character: cursorChar }, 'file://' + sfvPath);
+    var hv = h && h.contents && h.contents.value || '';
+    test(hv.indexOf('foam.mlang.sink.GroupBy') !== -1,
+      'Hover on this.GROUP_BY in StringFilterView shows return foam.mlang.sink.GroupBy');
+    // Regression: make sure only ONE "Returns:" line appears, even when the
+    // method axiom declares `type:` AND the code-parse resolves a concrete
+    // class. The parser-resolved concrete wins; the declared-type line is
+    // suppressed by buildMethodHover_.
+    var returnsCount = (hv.match(/Returns:/g) || []).length;
+    test(returnsCount === 1,
+      'Hover on this.GROUP_BY has exactly one `Returns:` line (got ' + returnsCount + ')');
+  }
+}
+
+// `.select(SINK).then((p) => …)` — p typed as SINK's class
+var thenSrc = [
+  "foam.CLASS({",
+  "  package: 'test',",
+  "  name: 'ChainTest',",
+  "  implements: [ 'foam.mlang.Expressions' ],",
+  "  methods: [",
+  "    function m() {",
+  "      this.dao.where(pred)",
+  "        .select(this.GROUP_BY(this.property, this.COUNT(), 21))",
+  "        .then((results) => {",
+  "          this.countByContents = results.groups;",
+  "        });",
+  "    }",
+  "  ]",
+  "});"
+].join('\n');
+var tt2 = foam.parse.lsp.TypeTracker.create();
+var chainModel = { package: 'test', name: 'ChainTest',
+  implements: [ 'foam.mlang.Expressions' ],
+  requires: [] };
+// cursor on `results.groups` (line 9, character 30-ish)
+var chainTypes = tt2.getVariableTypes(thenSrc, { line: 9, character: 30 }, chainModel, index);
+test(chainTypes.results === 'foam.mlang.sink.GroupBy',
+  '.then((results) => …) param typed from preceding .select(this.GROUP_BY(...)) (got: ' + chainTypes.results + ')');
+
+// Same with `.select(this.GroupBy.create({...}))` — direct create form
+var createChain = [
+  "foam.CLASS({",
+  "  name: 'X',",
+  "  requires: [ 'foam.mlang.sink.GroupBy' ],",
+  "  methods: [",
+  "    function m() {",
+  "      this.dao.select(this.GroupBy.create({})).then(function(r) { r.groups; });",
+  "    }",
+  "  ]",
+  "});"
+].join('\n');
+var createModel = { package: 'test', name: 'X',
+  requires: [ 'foam.mlang.sink.GroupBy' ] };
+var createTypes = tt2.getVariableTypes(createChain, { line: 5, character: 65 }, createModel, index);
+test(createTypes.r === 'foam.mlang.sink.GroupBy',
+  '.then(function(r) …) param typed from preceding .select(this.GroupBy.create({}))');
+
+// Arrow function with single-param no-parens: .then(r => r.groups)
+var bareArrow = [
+  "foam.CLASS({",
+  "  name: 'Y',",
+  "  requires: [ 'foam.mlang.sink.Count' ],",
+  "  methods: [",
+  "    function m() {",
+  "      this.dao.select(this.Count.create()).then(r => r.value);",
+  "    }",
+  "  ]",
+  "});"
+].join('\n');
+var bareModel = { package: 'test', name: 'Y', requires: [ 'foam.mlang.sink.Count' ] };
+var bareTypes = tt2.getVariableTypes(bareArrow, { line: 5, character: 55 }, bareModel, index);
+test(bareTypes.r === 'foam.mlang.sink.Count',
+  '.then(r => …) bare-arrow param typed from preceding .select');
+
+// === MESSAGE AXIOM: hover + go-to-definition ===
+
