+++ tools/tests/testFoamLSP.js
@@ -6,19 +6,12 @@
  * http://www.apache.org/licenses/LICENSE-2.0
  */
 
-// Quick test for FOAM LSP grammar — runs in seconds without full build.
+// Entrypoint for the FOAM LSP test suite. The actual tests live under
+// foam3/tools/tests/lsp/<category>.js — this file just boots the shared
+// harness and loads each category in order so the aggregate run produces a
+// single pass/fail tally.
+//
 // Usage: cd <your-project> && node foam3/tools/tests/testFoamLSP.js
-
-console.log = function() { console.error.apply(console, arguments); };
-console.warn = function() { console.error.apply(console, arguments); };
-globalThis.SILENT = false; globalThis.VERBOSE = false;
-globalThis.DRY_RUN = false; globalThis.HELP = false; globalThis.NOP = '';
-
-process.on('unhandledRejection', function(e) {});
-process.on('uncaughtException', function(e) {
-  if ( e.message && ( e.message.includes('document') || e.message.includes('window') ) ) return;
-  if ( e instanceof SyntaxError ) return;
-});
 
 // Hard watchdog: fail fast if any single test infinite-loops. 30s is generous
 // given the whole suite normally completes in ~2s after pmake boot.
@@ -35,3385 +28,21 @@ if ( require.main === module ) {
   }, 30000).unref();
 }
 
-var path = require('path');
-var fs = require('fs');
-var pmake = require(path.resolve(__dirname, '../pmake'));
-var buildlib = require(path.resolve(__dirname, '../buildlib'));
-buildlib.error = function() { /* suppress fatal errors during boot */ };
-
-var pomPath = path.resolve(process.cwd(), 'pom');
-pmake.bind(buildlib, '-makers=LSP -pom=' + pomPath)();
-
-// === TEST HARNESS ===
-
-var failures = 0;
-var passes = 0;
-
-function test(condition, message) {
-  if ( condition ) {
-    passes++;
-    console.error('  \x1b[32m✓\x1b[0m ' + message);
-  } else {
-    failures++;
-    console.error('  \x1b[31m✘ FAIL:\x1b[0m ' + message);
-  }
-}
-
-function section(name) {
-  console.error('\n\x1b[1m=== ' + name + ' ===\x1b[0m');
-}
-
-// === TEST FILES ===
-
-var TEST_FILES = [
-  'foam3/src/foam/lang/types.js',
-  'foam3/src/foam/parse/parse.js',
-  'foam3/src/foam/core/controller/ApplicationController.js',
-  'foam3/src/foam/lang/Enum.js',
-  'foam3/src/foam/parse/SimpleQueryParser.js'
-];
-
-// === FOAMINDEX TESTS ===
-
-section('FoamIndex');
-var index = foam.parse.lsp.FoamIndex.create();
-test(index.getAllClassIds().length > 100, 'getAllClassIds returns many classes: ' + index.getAllClassIds().length);
-test(index.classExists('foam.lang.FObject'), 'FObject exists');
-test(index.getPropertyTypes().length > 50, 'Many property types: ' + index.getPropertyTypes().length);
-test(index.getPropertyTypes().some(function(t) { return t.name === 'String'; }), 'Includes String type');
-test(index.getPropertyTypes().some(function(t) { return t.name === 'Boolean'; }), 'Includes Boolean type');
-test(index.getPropertyTypes().some(function(t) { return t.name === 'FObjectProperty'; }), 'Includes FObjectProperty type');
-
-// === GRAMMAR TESTS ===
-
-section('FoamClassGrammar — symbol check');
-var grammar = foam.parse.lsp.FoamClassGrammar.create({ index: index });
-test(Object.keys(grammar.symbolMap_).length > 20, 'Grammar has symbols: ' + Object.keys(grammar.symbolMap_).length);
-test('START' in grammar.symbolMap_, 'Has START symbol');
-
-section('FoamClassGrammar — parse real files');
-TEST_FILES.forEach(function(filePath) {
-  var absPath = path.resolve(process.cwd(), filePath);
-  if ( ! fs.existsSync(absPath) ) {
-    console.error('  Skipping (not found): ' + filePath);
-    return;
-  }
-
-  var text = fs.readFileSync(absPath, 'utf8');
-  console.error('\n  File: ' + filePath + ' (' + text.split('\n').length + ' lines)');
-
-  var ps = foam.parse.StringPStream.create({ str: text + String.fromCharCode(26) });
-  try {
-    var result = grammar.parse(ps);
-    test(result !== undefined, 'Parses without error');
-  } catch (e) {
-    test(false, 'Parse threw: ' + e.message);
-  }
-});
-
-// === COMPLETION TESTS ===
-
-section('CompletionHandler — property types');
-var completionHandler = foam.parse.lsp.handlers.CompletionHandler.create({ index: index, grammar: grammar });
-
-// Test completion when user is TYPING (empty class value) — the real use case
-var Q = String.fromCharCode(39); // single quote
-var compText = 'foam.CLASS({\n  properties: [\n    { class: ' + Q + Q + ', name: ' + Q + 'x' + Q + ' }\n  ]\n})';
-var compLines = compText.split('\n');
-var compCharPos = compLines[2].indexOf(Q) + 1;
-var result = completionHandler.handle(compText, { line: 2, character: compCharPos });
-test(result.items.length > 0, 'Property type completions (empty value): ' + result.items.length + ' items');
-test(result.items.some(function(i) { return i.label === 'String'; }), 'Includes String');
-test(result.items.some(function(i) { return i.label === 'Boolean'; }), 'Includes Boolean');
-test(result.items.some(function(i) { return i.label === 'FObjectProperty'; }), 'Includes FObjectProperty');
-
-// Test completion for extends
-var extendsText = 'foam.CLASS({\n  extends: ' + Q + Q + '\n})';
-var extendsResult = completionHandler.handle(extendsText, { line: 1, character: 13 });
-test(extendsResult.items.length > 0, 'Class completions for extends (empty): ' + extendsResult.items.length + ' items');
-
-// Test completion for partial extends value (typing 'foam.')
-var partialExtendsText = 'foam.CLASS({\n  extends: ' + Q + 'foam.' + Q + '\n})';
-var partialExtendsResult = completionHandler.handle(partialExtendsText, { line: 1, character: 17 });
-test(partialExtendsResult.items.length > 0, 'Class completions for extends (partial foam.): ' + partialExtendsResult.items.length + ' items');
-
-// Test completion for partial class type (typing 'S')
-var partialClassText = 'foam.CLASS({\n  properties: [\n    { class: ' + Q + 'S' + Q + ' }\n  ]\n})';
-var partialClassResult = completionHandler.handle(partialClassText, { line: 2, character: 15 });
-test(partialClassResult.items.length > 0, 'Property type completions (partial S): ' + partialClassResult.items.length + ' items');
-
-// Test completion with factory function before the class property (regression test)
-var factoryText = 'foam.CLASS({\n  properties: [\n    { name: ' + Q + 'y' + Q + ', factory: function() { return {}; } },\n    { class: ' + Q + Q + ', name: ' + Q + 'x' + Q + ' }\n  ]\n})';
-var factoryResult = completionHandler.handle(factoryText, { line: 3, character: 14 });
-test(factoryResult.items.length > 0, 'Completions after factory property: ' + factoryResult.items.length + ' items');
-
-// Completion inside existing quoted value — extends: 'f' with closing quote present
-var existingQuoteText = 'foam.CLASS({\n  extends: ' + Q + 'f' + Q + ',\n  name: ' + Q + 'Test' + Q + '\n})';
-var existingQuoteResult = completionHandler.handle(existingQuoteText, { line: 1, character: 13 });
-test(existingQuoteResult.items.length > 0, 'Completion inside existing quoted value: ' + existingQuoteResult.items.length + ' items');
-
-// Completion inside existing class: 'S' with closing quote
-var existingClassText = 'foam.CLASS({\n  properties: [\n    { class: ' + Q + 'S' + Q + ', name: ' + Q + 'x' + Q + ' }\n  ]\n})';
-var existingClassResult = completionHandler.handle(existingClassText, { line: 2, character: 15 });
-test(existingClassResult.items.length > 0, 'Completion inside existing class value: ' + existingClassResult.items.length + ' items');
-
-// === MEMBER COMPLETION TESTS ===
-
-section('MemberCompletionHandler — this. + requires + create');
-var memberHandler = foam.parse.lsp.handlers.MemberCompletionHandler.create({ index: index });
-
-// this. suggests properties + methods + required classes + imports
-var memberText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'Foo' + Q + ',\n  requires: [\n    ' + Q + 'foam.parse.Suggestion' + Q + '\n  ],\n  imports: [\n    ' + Q + 'userDAO' + Q + '\n  ],\n  properties: [\n    { class: ' + Q + 'String' + Q + ', name: ' + Q + 'bar' + Q + ' }\n  ],\n  methods: [\n    function doStuff() {\n      this.\n    }\n  ]\n})';
-var memberResult = memberHandler.handle(memberText, { line: 14, character: 11 });
-test(memberResult.items.length > 0, 'this. returns items: ' + memberResult.items.length);
-test(memberResult.items.some(function(i) { return i.label === 'Suggestion'; }), 'this. includes required class Suggestion');
-test(memberResult.items.some(function(i) { return i.label === 'userDAO'; }), 'this. includes imported userDAO');
-
-// this.Suggestion.create({ ▊ }) suggests Suggestion properties
-var createText = 'foam.CLASS({\n  requires: [\n    ' + Q + 'foam.parse.Suggestion' + Q + '\n  ],\n  methods: [\n    function go() {\n      this.Suggestion.create({\n    }\n  ]\n})';
-var createResult = memberHandler.handle(createText, { line: 6, character: 38 });
-test(createResult.items.length > 0, 'this.X.create({ suggests properties: ' + createResult.items.length);
-test(createResult.items.some(function(i) { return i.label === 'text'; }), 'create({}) includes text property');
-test(createResult.items.some(function(i) { return i.label === 'category'; }), 'create({}) includes category property');
-
-// this.Suggestion.create({ ... multi-line ... }) — cursor inside block on separate line
-var multiCreateText = 'foam.CLASS({\n  requires: [\n    ' + Q + 'foam.parse.Suggestion' + Q + '\n  ],\n  methods: [\n    function go() {\n      this.Suggestion.create({\n        \n      })\n    }\n  ]\n})';
-var multiCreateResult = memberHandler.handle(multiCreateText, { line: 7, character: 8 });
-test(multiCreateResult.items.length > 0, 'Multi-line create({}) suggests properties: ' + multiCreateResult.items.length);
-
-// Multi-line create with { on separate line from .create(
-var separateBraceText = 'foam.CLASS({\n  requires: [\n    ' + Q + 'foam.parse.Suggestion' + Q + '\n  ],\n  methods: [\n    function go() {\n      this.Suggestion.create(\n        {\n          \n        }\n      )\n    }\n  ]\n})';
-var separateBraceResult = memberHandler.handle(separateBraceText, { line: 8, character: 10 });
-test(separateBraceResult.items.length > 0, 'create( + { on separate lines suggests: ' + separateBraceResult.items.length);
-
-// Method signature has params in detail — test with a real class
-var fs = require('fs');
-var realText = fs.readFileSync(path.resolve(process.cwd(), 'foam3/src/foam/u2/CitationView.js'), 'utf8');
-var realResult = memberHandler.handle(realText, { line: 79, character: 11 });
-var methodItems = realResult.items.filter(function(i) { return i.kind === 2 && i.detail && i.detail.indexOf('(') !== -1; });
-test(methodItems.length > 0, 'Method completions have param signatures: ' + methodItems.length);
-var myClassItem = realResult.items.find(function(i) { return i.label === 'myClass'; });
-test(myClassItem && myClassItem.detail === 'myClass(opt_extra)', 'myClass detail shows params: ' + (myClassItem ? myClassItem.detail : 'not found'));
-
-// === HOVER TESTS ===
-
-// Shared CSSTokenResolver so CSS-aware hovers (e.g. ^selector) work below.
-var cssTokenResolver = foam.parse.lsp.CSSTokenResolver.create();
-cssTokenResolver.loadFromRegistry();
-
-section('HoverHandler — class hover');
-var hoverHandler = foam.parse.lsp.handlers.HoverHandler.create({ index: index, cssTokenResolver: cssTokenResolver });
-
-var hoverText = "foam.CLASS({\n  requires: ['foam.parse.Suggestion']\n})";
-var hoverResult = hoverHandler.handle(hoverText, { line: 1, character: 20 });
-test(hoverResult != null, 'Hover returns result for class name');
-test(hoverResult && hoverResult.contents.value.indexOf('foam.parse.Suggestion') !== -1, 'Hover contains class name');
-
-// Hover on property type
-var propTypeHover = hoverHandler.handle("foam.CLASS({\n  properties: [\n    { class: 'FObjectProperty' }\n  ]\n})", { line: 2, character: 18 });
-test(propTypeHover != null, 'Hover returns result for property type');
-
-// Hover on short name from requires (e.g., 'Suggestion' resolves to foam.parse.Suggestion)
-var requiresHoverText = 'foam.CLASS({\n  requires: [\n    ' + Q + 'foam.parse.Suggestion' + Q + '\n  ],\n  methods: [\n    function go() {\n      this.Suggestion.create();\n    }\n  ]\n})';
-var shortNameHover = hoverHandler.handle(requiresHoverText, { line: 6, character: 12 });
-test(shortNameHover != null, 'Hover on required short name resolves');
-test(shortNameHover && shortNameHover.contents.value.indexOf('foam.parse.Suggestion') !== -1, 'Short name hover shows full class info');
-
-// Hover on 'create' shows class properties
-var createHover = hoverHandler.handle(requiresHoverText, { line: 6, character: 22 });
-test(createHover != null, 'Hover on create shows class info');
-test(createHover && createHover.contents.value.indexOf('create') !== -1, 'Create hover mentions create');
-
-// Hover on method name in synthetic text (avoids file line number issues)
-var methodHoverText2 = 'foam.CLASS({\n  package: ' + Q + 'foam.parse' + Q + ',\n  name: ' + Q + 'Suggestion' + Q + ',\n  methods: [\n    function matches() {}\n  ]\n})';
-var methodHover2 = hoverHandler.handle(methodHoverText2, { line: 4, character: 15 });
-test(methodHover2 != null, 'Hover on method name shows info');
-
-// === DIAGNOSTICS TESTS ===
-
-section('DiagnosticsHandler');
-var diagHandler = foam.parse.lsp.handlers.DiagnosticsHandler.create({ index: index });
-
-// Valid file — no errors
-var validText = "foam.CLASS({\n  package: 'test',\n  name: 'Valid',\n  extends: 'foam.lang.FObject'\n})";
-var diags = diagHandler.handle(validText);
-var errors = diags.filter(function(d) { return d.severity <= 2; });
-test(errors.length === 0, 'Valid file has no errors/warnings');
-
-// Invalid extends
-var invalidText = "foam.CLASS({\n  extends: 'foo.bar.Missing'\n})";
-var diags2 = diagHandler.handle(invalidText);
-test(diags2.some(function(d) { return d.message.indexOf('Missing') !== -1; }), 'Flags unknown extends class');
-
-// Valid property type (full path) — should NOT be flagged
-var validPropText = "foam.CLASS({\n  properties: [\n    { class: 'foam.lang.FObjectProperty', name: 'x' }\n  ]\n})";
-var diags3 = diagHandler.handle(validPropText);
-var propErrors = diags3.filter(function(d) { return d.message.indexOf('FObjectProperty') !== -1; });
-test(propErrors.length === 0, 'foam.lang.FObjectProperty NOT flagged as unknown');
-
-// Valid property type (short name) — should NOT be flagged
-var validShortPropText = "foam.CLASS({\n  properties: [\n    { class: 'String', name: 'x' }\n  ]\n})";
-var diags3b = diagHandler.handle(validShortPropText);
-var shortPropErrors = diags3b.filter(function(d) { return d.message.indexOf('String') !== -1; });
-test(shortPropErrors.length === 0, 'String NOT flagged as unknown');
-
-// Constants strings — should NOT be flagged
-var constantsText = "foam.CLASS({\n  constants: [\n    { name: 'MACROS', value: ['DisplayWidth.XS', 'primary1'] }\n  ]\n})";
-var diags4 = diagHandler.handle(constantsText);
-test(diags4.filter(function(d) { return d.message.indexOf('DisplayWidth') !== -1; }).length === 0,
-  'DisplayWidth.XS in constants NOT flagged');
-
-// Requires with 'as' alias — should NOT flag the aliased class as unknown
-var aliasText = "foam.CLASS({\n  requires: [\n    'foam.parse.Suggestion as Sug'\n  ]\n})";
-var aliasDiags = diagHandler.handle(aliasText);
-var aliasErrors = aliasDiags.filter(function(d) { return d.message.indexOf('Suggestion') !== -1; });
-test(aliasErrors.length === 0, 'Requires with as alias NOT flagged as unknown');
-
-// Requires with 'as' alias for unknown class — SHOULD flag it
-var unknownAliasText = "foam.CLASS({\n  requires: [\n    'foo.bar.Missing as M'\n  ]\n})";
-var unknownAliasDiags = diagHandler.handle(unknownAliasText);
-test(unknownAliasDiags.some(function(d) { return d.message.indexOf('Missing') !== -1; }), 'Unknown class with as alias IS flagged');
-
-// Interface javaGetter referencing own property — should NOT be flagged
-var ifaceText = 'foam.INTERFACE({ package: ' + Q + 'test' + Q + ', name: ' + Q + 'MyAware' + Q + ', properties: [{ class: ' + Q + 'String' + Q + ', name: ' + Q + 'fileDate' + Q + ', javaGetter: ' + Q + 'return getFileDate();' + Q + ' }] })';
-var ifaceDiags = diagHandler.handle(ifaceText);
-var ifaceGetterErrors = ifaceDiags.filter(function(d) { return d.message.indexOf('fileDate') !== -1; });
-test(ifaceGetterErrors.length === 0, 'Interface own property getter NOT flagged');
-
-// Test getImplementors
-var implementors = index.getImplementors('foam.core.auth.CreatedByAware');
-test(implementors.length > 0, 'getImplementors finds classes implementing CreatedByAware: ' + implementors.length);
-
-// === LSP #4993 Fix 3: user-defined cssTokens not flagged as unknown ===
-section('DiagnosticsHandler — local cssTokens (issue #4993)');
-var diagWithTokens = foam.parse.lsp.handlers.DiagnosticsHandler.create({
-  index: index,
-  cssTokenResolver: cssTokenResolver
-});
-
-var localTokenSrc =
-  "foam.CLASS({\n" +
-  "  package: 'test',\n  name: 'LocalTokenUser',\n" +
-  "  cssTokens: [\n    { name: 'tooltipBackground', value: '#eeeeee' }\n  ],\n" +
-  "  css: `\n    ^ { background: $tooltipBackground; color: $white; }\n  `\n" +
-  "})";
-var localDiags = diagWithTokens.handle(localTokenSrc);
-test(localDiags.filter(function(d) { return d.message.indexOf('tooltipBackground') !== -1; }).length === 0,
-  'Local cssTokens: $tooltipBackground NOT flagged as unknown');
-test(localDiags.filter(function(d) { return d.message.indexOf('Unknown CSS token') !== -1 && d.message.indexOf('nonExistent') !== -1; }).length === 0,
-  'Local cssTokens: unrelated tokens untouched');
-
-var unknownTokenSrc =
-  "foam.CLASS({\n  package: 'test',\n  name: 'UnknownTokenUser',\n" +
-  "  cssTokens: [\n    { name: 'fooBg', value: '#ffffff' }\n  ],\n" +
-  "  css: `\n    ^ { background: $nonExistent; }\n  `\n})";
-var unknownDiags = diagWithTokens.handle(unknownTokenSrc);
-test(unknownDiags.some(function(d) { return d.message.indexOf('nonExistent') !== -1 && d.message.indexOf('Unknown CSS token') !== -1; }),
-  'Local cssTokens: $nonExistent still flagged');
-
-// === LSP #4993 Fix 4: unused ^classname in css: ===
-section('DiagnosticsHandler — unused ^classname (issue #4993)');
-var unusedSrc =
-  "foam.CLASS({\n  package: 'test',\n  name: 'UnusedCss',\n" +
-  "  css: `\n    ^foo { color: red; }\n    ^bar { color: blue; }\n  `,\n" +
-  "  methods: [\n" +
-  "    function render() { this.addClass(this.myClass('foo')); }\n" +
-  "  ]\n})";
-var unusedDiags = diagWithTokens.handle(unusedSrc);
-var unusedWarns = unusedDiags.filter(function(d) { return /Unused CSS class/.test(d.message); });
-test(unusedWarns.length === 1, 'Unused ^classname: exactly one unused class flagged');
-test(unusedWarns.some(function(d) { return d.message.indexOf("'^bar'") !== -1; }),
-  'Unused ^classname: ^bar (unused) is flagged');
-test(! unusedWarns.some(function(d) { return d.message.indexOf("'^foo'") !== -1; }),
-  'Unused ^classname: ^foo (applied via myClass) is NOT flagged');
-
-// Dynamic myClass(var) → suppress unused-class diagnostics entirely
-var dynamicSrc =
-  "foam.CLASS({\n  package: 'test',\n  name: 'DynamicMyClass',\n" +
-  "  css: `\n    ^alpha { color: red; }\n    ^beta { color: blue; }\n  `,\n" +
-  "  methods: [\n" +
-  "    function render(tag) { this.addClass(this.myClass(tag)); }\n" +
-  "  ]\n})";
-var dynamicDiags = diagWithTokens.handle(dynamicSrc);
-test(dynamicDiags.filter(function(d) { return /Unused CSS class/.test(d.message); }).length === 0,
-  'Unused ^classname: dynamic myClass(arg) suppresses all unused-class warnings');
-
-// === LSP #4993 Fix 1: go-to-definition follows FObjectProperty of: ===
-section('DefinitionHandler — property-chain navigation (issue #4993)');
-// Uses existing indexed classes: FObjectProperty with of: is pervasive in foam3.
-// foam.u2.Element has `tooltip: { class: 'FObjectProperty', of: 'foam.u2.Tooltip' }` in many versions —
-// pick any real example. We use the FoamIndex resolver directly as the core check:
-var resolvedClassId = index.resolvePropertyTypeClassId('foam.core.auth.User', 'group');
-test(resolvedClassId === null || typeof resolvedClassId === 'string',
-  'resolvePropertyTypeClassId: returns string or null for foam.core.auth.User.group');
-
-// Chain-walk in DefinitionHandler: synthesize a minimal pair of classes and check the walk.
-var chainClassA = index.getAllClassIds().find(function(id) {
-  var cls = index.getClass(id);
-  if ( ! cls ) return false;
-  var props = cls.getAxiomsByClass(foam.lang.Property);
-  for ( var i = 0 ; i < props.length ; i++ ) {
-    if ( index.resolvePropertyTypeClassId(id, props[i].name) ) return true;
-  }
-  return false;
-});
-test(typeof chainClassA === 'string',
-  'Found at least one class with an FObjectProperty-typed property for chain walking');
-
-// === LSP #4993 Fix 2: foam.LIB indexing ===
-section('FoamIndex — foam.LIB registry (issue #4993)');
-test(index.getAllLibNames().length > 0,
-  'LIB registry: at least one foam.LIB indexed (got ' + index.getAllLibNames().length + ')');
-var colorEntry = index.getLibEntry('foam.Color');
-test(colorEntry !== null, 'LIB registry: foam.Color has an entry');
-test(colorEntry && (colorEntry.methods || []).indexOf('adjustAlpha') !== -1,
-  'LIB registry: foam.Color.adjustAlpha indexed as method');
-
-// Go-to-definition for foam.Color.adjustAlpha — must be inside a foam.CLASS
-// so isFoamFile() passes. Cursor lands on 'adjustAlpha'.
-var libDefHandler = foam.parse.lsp.handlers.DefinitionHandler.create({ index: index });
-var libCallText =
-  "foam.CLASS({\n  package: 'test',\n  name: 'LibCaller',\n" +
-  "  methods: [\n    function f() {\n      var c = foam.Color.adjustAlpha(x, 0.5);\n    }\n  ]\n})";
-// Line 5 is "      var c = foam.Color.adjustAlpha(x, 0.5);"
-// 'adjustAlpha' starts at character 25; land cursor on 'adjustAlpha'.
-var libDef = libDefHandler.handle(libCallText, { line: 5, character: 33 });
-test(libDef && libDef.uri && libDef.uri.indexOf('colorlib.js') !== -1,
-  'LIB definition: foam.Color.adjustAlpha navigates to colorlib.js');
-
-// Completion after 'foam.Color.' — inside a method body of a foam.CLASS
-var memberCompletion = foam.parse.lsp.handlers.MemberCompletionHandler.create({ index: index });
-var compSrc =
-  "foam.CLASS({\n  package: 'test',\n  name: 'LibCompletion',\n" +
-  "  methods: [\n    function f() {\n      var c = foam.Color.\n    }\n  ]\n})";
-// Line 5 ends with 'foam.Color.'; cursor positioned right after the trailing dot.
-var compResult = memberCompletion.handle(compSrc, { line: 5, character: 25 });
-test(compResult && compResult.items && compResult.items.some(function(it) { return it.label === 'adjustAlpha'; }),
-  'LIB completion: foam.Color. suggests adjustAlpha');
-
-// Hover on a LIB member
-var libHoverHandler = foam.parse.lsp.handlers.HoverHandler.create({
-  index: index,
-  cssTokenResolver: cssTokenResolver
-});
-var hoverResult = libHoverHandler.handle(libCallText, { line: 5, character: 33 });
-test(hoverResult && hoverResult.contents && /foam\.Color/.test(hoverResult.contents.value),
-  'LIB hover: foam.Color.adjustAlpha shows hover with lib name');
-
-// === DEFINITION TESTS ===
-
-section('DefinitionHandler');
-var defHandler = foam.parse.lsp.handlers.DefinitionHandler.create({ index: index });
-
-// Definition on class in extends
-var defText = "foam.CLASS({\n  extends: 'foam.lang.FObject'\n})";
-var defResult = defHandler.handle(defText, { line: 1, character: 20 });
-// May return null if file index not built yet — that's OK for now
-test(defResult == null || (defResult.uri && defResult.uri.indexOf('FObject') !== -1),
-  'Definition returns null or correct path');
-
-// === CURSOR ANALYZER TESTS ===
-
-section('CursorAnalyzer');
-var analyzer = foam.parse.lsp.CursorAnalyzer.create();
-
-// offsetToPosition
-var testText = 'line0\nline1\nline2';
-var pos = analyzer.offsetToPosition(testText, 6); // start of line1
-test(pos.line === 1 && pos.character === 0, 'offsetToPosition: line 1 start');
-
-// positionToOffset
-var offset = analyzer.positionToOffset(testText, { line: 1, character: 3 });
-test(offset === 9, 'positionToOffset: line1 char3 = offset 9, got: ' + offset);
-
-// resolveClassId
-var classText = 'foam.CLASS({ package: ' + Q + 'foam.parse' + Q + ', name: ' + Q + 'Suggestion' + Q + ' })';
-test(analyzer.resolveClassId(classText) === 'foam.parse.Suggestion', 'resolveClassId extracts class ID');
-
-// parseRequires
-var reqText = 'foam.CLASS({ requires: [' + Q + 'foam.u2.DetailView' + Q + ', ' + Q + 'foam.u2.Element as El' + Q + '] })';
-var reqMap = analyzer.parseRequires(reqText);
-test(reqMap['DetailView'] === 'foam.u2.DetailView', 'parseRequires: DetailView');
-test(reqMap['El'] === 'foam.u2.Element', 'parseRequires: alias El');
-
-// getMethodSignature
-var mockMethod = { name: 'start', code: function start(spec, args) {} };
-test(analyzer.getMethodSignature(mockMethod) === 'start(spec, args)', 'getMethodSignature from code');
-
-// getAllPropertiesForFile — includes implements interfaces
-// FOAM JS doesn't merge interface props into class — we need to check separately
-var implText = 'foam.CLASS({ package: ' + Q + 'foam.core.auth' + Q + ', name: ' + Q + 'User' + Q + ', implements: [' + Q + 'foam.core.auth.CreatedByAware' + Q + '] })';
-var allProps = index.getAllPropertiesForFile('foam.core.auth.User', implText);
-test(allProps['createdby'] != null, 'getAllPropertiesForFile includes createdBy from interface');
-
-// === REAL FILE COVERAGE ===
-
-section('Real file coverage');
-
-// Go-to-definition on real file
-var defHandler2 = foam.parse.lsp.handlers.DefinitionHandler.create({ index: index });
-index.buildFileIndex();
-var defText2 = 'foam.CLASS({ extends: ' + Q + 'foam.core.auth.User' + Q + ' })';
-var defResult2 = defHandler2.handle(defText2, { line: 0, character: 25 });
-test(defResult2 != null && defResult2.uri.indexOf('User.js') !== -1, 'Go-to-definition finds User.js');
-
-// this. on real class shows 200+ items
-var realMemberText = fs.readFileSync(path.resolve(process.cwd(), 'foam3/src/foam/u2/CitationView.js'), 'utf8');
-var realMemberHandler = foam.parse.lsp.handlers.MemberCompletionHandler.create({ index: index });
-var realMemberResult = realMemberHandler.handle(realMemberText, { line: 79, character: 11 });
-test(realMemberResult.items.length > 100, 'this. on real file: ' + realMemberResult.items.length + ' items');
-
-// === FLAG-AWARE FILE INDEX TESTS ===
-
-section('Flag-aware file index');
-index.buildFileIndex();
-test(Object.keys(index.fileIndex_).length > 3000, 'File index includes 3000+ classes: ' + Object.keys(index.fileIndex_).length);
-
-// Test classes are in the index with correct flags
-var testEntry = index.fileIndex_['foam.core.test.Test'];
-test(testEntry != null, 'foam.core.test.Test found in file index');
-test(testEntry && testEntry.flags && testEntry.flags.indexOf('test') !== -1, 'Test class has test flag');
-
-// Swift classes are in the index
-var swiftEntry = index.fileIndex_['foam.swift.SwiftClass'];
-test(swiftEntry != null || true, 'Swift class in file index (may not exist in all projects)');
-
-// classKnown_ via diagnostics should not flag test classes
-var diagHandler2 = foam.parse.lsp.handlers.DiagnosticsHandler.create({ index: index });
-var testExtendsText = 'foam.CLASS({\n  extends: ' + Q + 'foam.core.test.Test' + Q + '\n})';
-var testDiags = diagHandler2.handle(testExtendsText);
-var testWarnings = testDiags.filter(function(d) { return d.message.indexOf('foam.core.test.Test') !== -1; });
-test(testWarnings.length === 0, 'extends foam.core.test.Test NOT flagged as unknown');
-
-// === WORKSPACE ANALYZER TESTS ===
-
-section('WorkspaceAnalyzer');
-var wsAnalyzer = foam.parse.lsp.handlers.WorkspaceAnalyzer.create({ index: index });
-
-// Test single file analysis
-var singleResult = wsAnalyzer.analyzeSingleFile(path.resolve(process.cwd(), 'foam3/src/foam/parse/parse.js'));
-test(singleResult != null, 'WorkspaceAnalyzer can analyze a single file');
-test(Array.isArray(singleResult), 'Single file result is an array');
-
-// Test message generalization
-var gen1 = wsAnalyzer.generalizeMessage("Unknown class in requires: 'foam.core.auth.User'");
-test(gen1.indexOf('*') !== -1, 'generalizeMessage replaces class name with wildcard: ' + gen1);
-
-var gen2 = wsAnalyzer.generalizeMessage("Unknown property type: 'FooBar'");
-test(gen2 === "Unknown property type: 'FooBar'", 'generalizeMessage leaves short names alone');
-
-// === FOLDING RANGE TESTS ===
-
-section('Folding Ranges');
-
-// Load getFoldingRanges from server module via inline test
-var foldText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'Fold' + Q + ',\n  properties: [\n    { class: ' + Q + 'String' + Q + ', name: ' + Q + 'x' + Q + ' },\n    { class: ' + Q + 'Int' + Q + ', name: ' + Q + 'y' + Q + ' }\n  ],\n  methods: [\n    function foo() {},\n    function bar() {}\n  ]\n})';
-
-// Manual fold range detection (same algorithm as server)
-function testGetFoldingRanges(text) {
-  var ranges = [];
-  var keywords = ['properties', 'methods', 'requires', 'imports', 'exports', 'javaImports', 'actions', 'listeners'];
-  var lines = text.split('\n');
-
-  for ( var k = 0 ; k < keywords.length ; k++ ) {
-    var kw = keywords[k];
-    var pattern = new RegExp(kw + '\\s*:\\s*\\[');
-    for ( var i = 0 ; i < lines.length ; i++ ) {
-      if ( ! pattern.test(lines[i]) ) continue;
-      var depth = 0;
-      var foundOpen = false;
-      var endLine = -1;
-      for ( var j = i ; j < lines.length ; j++ ) {
-        var line = lines[j];
-        for ( var c = 0 ; c < line.length ; c++ ) {
-          if ( line[c] === '[' ) { depth++; foundOpen = true; }
-          else if ( line[c] === ']' ) {
-            depth--;
-            if ( foundOpen && depth === 0 ) { endLine = j; break; }
-          }
-        }
-        if ( endLine !== -1 ) break;
-      }
-      if ( endLine > i ) ranges.push({ startLine: i, endLine: endLine, kind: 'region' });
-    }
-  }
-  return ranges;
-}
-
-var foldRanges = testGetFoldingRanges(foldText);
-test(foldRanges.length === 2, 'Fold ranges found properties and methods: ' + foldRanges.length);
-test(foldRanges[0].startLine === 3, 'Properties fold starts at line 3');
-test(foldRanges[1].startLine === 7, 'Methods fold starts at line 7');
-
-// Test with requires
-var foldText2 = 'foam.CLASS({\n  requires: [\n    ' + Q + 'foam.u2.Element' + Q + '\n  ],\n  properties: [\n    ' + Q + 'x' + Q + '\n  ]\n})';
-var foldRanges2 = testGetFoldingRanges(foldText2);
-test(foldRanges2.length === 2, 'Fold ranges found requires and properties');
-
-// === CODE ACTION TESTS ===
-
-section('Code Actions');
-
-// Test findSimilarClasses (same algorithm as server)
-function testFindSimilarClasses(target, idx, maxResults) {
-  var targetShort = target.split('.').pop().toLowerCase();
-  var ids = idx.getAllClassIds();
-  var scored = [];
-  for ( var i = 0 ; i < ids.length ; i++ ) {
-    var shortName = ids[i].split('.').pop().toLowerCase();
-    if ( shortName === targetShort ) {
-      scored.push({ id: ids[i], score: 100 });
-    } else if ( shortName.indexOf(targetShort) !== -1 || targetShort.indexOf(shortName) !== -1 ) {
-      scored.push({ id: ids[i], score: 50 });
-    }
-  }
-  scored.sort(function(a, b) { return b.score - a.score; });
-  var results = [];
-  for ( var i = 0 ; i < Math.min(scored.length, maxResults) ; i++ ) results.push(scored[i].id);
-  return results;
-}
-
-// 'foam.core.FObject' should suggest 'foam.lang.FObject'
-var suggestions = testFindSimilarClasses('foam.core.FObject', index, 3);
-test(suggestions.some(function(s) { return s === 'foam.lang.FObject'; }), 'Suggests foam.lang.FObject for foam.core.FObject');
-
-// === WORKSPACE SYMBOL TESTS ===
-
-section('Workspace Symbols');
-var allIds = index.getAllClassIds();
-var symbolQuery = 'fobject';
-var matchCount = 0;
-for ( var i = 0 ; i < allIds.length ; i++ ) {
-  if ( allIds[i].toLowerCase().indexOf(symbolQuery) !== -1 ) matchCount++;
-}
-test(matchCount > 0, 'Workspace symbol query "fobject" finds matches: ' + matchCount);
-
-// === FILE MODEL CACHE TESTS ===
-
-section('FileModelCache');
-var cache = foam.parse.lsp.FileModelCache.create();
-
-// Single class file
-var singleText = 'foam.CLASS({ package: ' + Q + 'test' + Q + ', name: ' + Q + 'Foo' + Q + ', extends: ' + Q + 'foam.lang.FObject' + Q + ', properties: [{ class: ' + Q + 'String' + Q + ', name: ' + Q + 'bar' + Q + ' }] })';
-var singleModels = cache.parseFileModels(singleText);
-test(singleModels.length === 1, 'Single class: 1 model');
-test(singleModels[0].package === 'test', 'Single class: package');
-test(singleModels[0].name === 'Foo', 'Single class: name');
-test(singleModels[0].extends === 'foam.lang.FObject', 'Single class: extends');
-test(singleModels[0].properties.length === 1, 'Single class: 1 property');
-test(singleModels[0].properties[0].name === 'bar', 'Single class: property name');
-
-// Multi-class file
-var multiText = 'foam.CLASS({ package: ' + Q + 'test' + Q + ', name: ' + Q + 'A' + Q + ' });\nfoam.CLASS({ package: ' + Q + 'test' + Q + ', name: ' + Q + 'B' + Q + ' });';
-var multiModels = cache.parseFileModels(multiText);
-test(multiModels.length === 2, 'Multi-class: 2 models');
-test(multiModels[0].name === 'A', 'Multi-class: first is A');
-test(multiModels[1].name === 'B', 'Multi-class: second is B');
-
-// Multi-refines file
-var refinesText = 'foam.CLASS({ refines: ' + Q + 'foam.core.reflow.TableDAOAgent' + Q + ', properties: [{ name: ' + Q + 'x' + Q + ' }] });\nfoam.CLASS({ refines: ' + Q + 'foam.core.reflow.Flow' + Q + ', properties: [{ name: ' + Q + 'y' + Q + ' }] });';
-var refinesModels = cache.parseFileModels(refinesText);
-test(refinesModels.length === 2, 'Multi-refines: 2 models');
-test(refinesModels[0].refines === 'foam.core.reflow.TableDAOAgent', 'Refines: first target');
-test(refinesModels[1].refines === 'foam.core.reflow.Flow', 'Refines: second target');
-
-// ENUM
-var enumText = 'foam.ENUM({ package: ' + Q + 'test' + Q + ', name: ' + Q + 'Status' + Q + ', values: [{ name: ' + Q + 'ACTIVE' + Q + ' }] })';
-var enumModels = cache.parseFileModels(enumText);
-test(enumModels.length === 1, 'ENUM: 1 model');
-test(enumModels[0].type_ === 'ENUM', 'ENUM: type is ENUM');
-
-// Implements array
-var implText2 = 'foam.CLASS({ package: ' + Q + 'test' + Q + ', name: ' + Q + 'Impl' + Q + ', implements: [' + Q + 'foam.core.auth.CreatedByAware' + Q + '] })';
-var implModels = cache.parseFileModels(implText2);
-test(implModels[0].implements.length === 1, 'Implements: 1 interface');
-test(implModels[0].implements[0] === 'foam.core.auth.CreatedByAware', 'Implements: correct interface');
-
-// Broken file (user typing) — returns partial results
-var brokenText = 'foam.CLASS({ package: ' + Q + 'test' + Q + ', name: ' + Q + 'Broken' + Q + ' });\nfoam.CLASS({ package: ' + Q + 'test' + Q + ', name: ';
-var brokenModels = cache.parseFileModels(brokenText);
-test(brokenModels.length >= 1, 'Broken file: at least 1 model recovered');
-
-// Caching
-var cached1 = cache.getModels('file:///test.js', singleText);
-var cached2 = cache.getModels('file:///test.js', singleText);
-test(cached1 === cached2, 'Cache hit: same reference returned');
-
-// Cache invalidation
-cache.invalidate('file:///test.js');
-var cached3 = cache.getModels('file:///test.js', singleText);
-test(cached3 !== cached1, 'Cache invalidated: new reference');
-
-// Real file
-var realText2 = fs.readFileSync(path.resolve(process.cwd(), 'foam3/src/foam/core/controller/ApplicationController.js'), 'utf8');
-var realModels = cache.parseFileModels(realText2);
-test(realModels.length >= 1, 'Real file: ' + realModels.length + ' models');
-test(realModels[0].package === 'foam.core.controller', 'Real file: correct package');
-test(realModels[0].name === 'ApplicationController', 'Real file: correct name');
-test(realModels[0].requires && realModels[0].requires.length > 10, 'Real file: has requires');
-test(realModels[0].properties && realModels[0].properties.length > 5, 'Real file: has properties');
-
-// === TYPE TRACKER TESTS ===
-
-section('TypeTracker');
-var typeTracker = foam.parse.lsp.TypeTracker.create({ cache: cache });
-
-// var sug = this.Suggestion.create() → sug has type foam.parse.Suggestion
-var typeText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'TypeTest' + Q + ',\n  requires: [\n    ' + Q + 'foam.parse.Suggestion' + Q + '\n  ],\n  methods: [\n    function go() {\n      var sug = this.Suggestion.create({});\n      sug.text;\n    }\n  ]\n})';
-var typeModel = cache.getModelAt('', typeText, 9);
-var varTypes = typeTracker.getVariableTypes(typeText, { line: 9, character: 10 }, typeModel, index);
-test(varTypes['sug'] === 'foam.parse.Suggestion', 'TypeTracker: sug resolved to foam.parse.Suggestion');
-
-// Unknown variable — not tracked
-test(varTypes['unknown'] === undefined, 'TypeTracker: unknown variable returns undefined');
-
-// Variable type completion: sug. suggests Suggestion properties
-var typeMemberHandler = foam.parse.lsp.handlers.MemberCompletionHandler.create({ index: index, cache: cache, typeTracker: typeTracker });
-var typeCompText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'TypeTest2' + Q + ',\n  requires: [\n    ' + Q + 'foam.parse.Suggestion' + Q + '\n  ],\n  methods: [\n    function go() {\n      var sug = this.Suggestion.create({});\n      sug.text;\n    }\n  ]\n})';
-var typeCompResult = typeMemberHandler.handle(typeCompText, { line: 9, character: 10 });
-test(typeCompResult.items.length > 0, 'Variable type completion: sug. returns items: ' + typeCompResult.items.length);
-test(typeCompResult.items.some(function(i) { return i.label === 'text'; }), 'Variable type completion includes text property');
-
-// Hover on variable.property — sug.text should show property info from Suggestion
-var typeHoverHandler = foam.parse.lsp.handlers.HoverHandler.create({ index: index, cache: cache, typeTracker: typeTracker });
-var typeHoverText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'HoverTest' + Q + ',\n  requires: [\n    ' + Q + 'foam.parse.Suggestion' + Q + '\n  ],\n  methods: [\n    function go() {\n      var sug = this.Suggestion.create({});\n      sug.text;\n    }\n  ]\n})';
-var varPropHover = typeHoverHandler.handle(typeHoverText, { line: 9, character: 10 });
-test(varPropHover != null, 'Hover on variable.property resolves: ' + (varPropHover ? 'yes' : 'null'));
-test(varPropHover && varPropHover.contents.value.indexOf('text') !== -1, 'variable.property hover shows property name');
-
-// === SEMANTIC TOKEN HANDLER TESTS ===
-
-section('SemanticTokenHandler');
-var semanticHandler = foam.parse.lsp.handlers.SemanticTokenHandler.create({ index: index, cache: cache, typeTracker: typeTracker });
-
-// File with requires — this.Suggestion should get semantic token
-var semText = 'foam.CLASS({\n  requires: [\n    ' + Q + 'foam.parse.Suggestion' + Q + '\n  ],\n  methods: [\n    function go() {\n      var s = this.Suggestion.create({});\n      s.text;\n    }\n  ]\n})';
-var semResult = semanticHandler.handle(semText);
-test(semResult.data.length > 0, 'Semantic tokens: has token data: ' + semResult.data.length + ' values');
-// Each token is 5 values: deltaLine, deltaChar, length, type, modifiers
-test(semResult.data.length % 5 === 0, 'Semantic tokens: data length is multiple of 5');
-
-// Token for 'Suggestion' in this.Suggestion — should be type 0 (type)
-var tokenCount = semResult.data.length / 5;
-var hasTypeToken = false;
-for ( var t = 0 ; t < tokenCount ; t++ ) {
-  if ( semResult.data[t * 5 + 3] === 0 ) { hasTypeToken = true; break; }
-}
-test(hasTypeToken, 'Semantic tokens: includes type token for requires alias');
-
-// === JAVA BLOCK COMPLETION TESTS ===
-
-section('Java block completions');
-
-// get inside javaCode should suggest getters with Java types
-var BT = String.fromCharCode(96);
-var javaCompText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'JTest' + Q + ',\n  properties: [\n    { class: ' + Q + 'String' + Q + ', name: ' + Q + 'firstName' + Q + ' },\n    { class: ' + Q + 'String' + Q + ', name: ' + Q + 'lastName' + Q + ' }\n  ],\n  methods: [\n    {\n      name: ' + Q + 'fullName' + Q + ',\n      javaCode: ' + BT + '\n        get\n      ' + BT + '\n    }\n  ]\n})';
-// Cursor on line 11 after 'get' — character 11
-var javaCompResult = completionHandler.handle(javaCompText, { line: 11, character: 11 });
-test(javaCompResult.items.length > 0, 'Java block: get suggests getters: ' + javaCompResult.items.length);
-test(javaCompResult.items.some(function(i) { return i.label === 'getFirstName()'; }), 'Java block: suggests getFirstName()');
-test(javaCompResult.items.some(function(i) { return i.label === 'getLastName()'; }), 'Java block: suggests getLastName()');
-
-// Lowercase partial: 'getfir' should match getFirstName
-var javaPartialText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'JTest3' + Q + ',\n  properties: [\n    { class: ' + Q + 'String' + Q + ', name: ' + Q + 'firstName' + Q + ' },\n    { class: ' + Q + 'String' + Q + ', name: ' + Q + 'lastName' + Q + ' }\n  ],\n  methods: [\n    {\n      name: ' + Q + 'fullName' + Q + ',\n      javaCode: ' + BT + '\n        getfir\n      ' + BT + '\n    }\n  ]\n})';
-var javaPartialResult = completionHandler.handle(javaPartialText, { line: 11, character: 14 });
-test(javaPartialResult.items.length === 1, 'Java block: getfir filters to 1 item: ' + javaPartialResult.items.length);
-test(javaPartialResult.items.some(function(i) { return i.label === 'getFirstName()'; }), 'Java block: getfir matches getFirstName()');
-
-// Getter detail shows Java return type — use real class foam.parse.Suggestion which has String properties
-var javaRealText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse' + Q + ',\n  name: ' + Q + 'Suggestion' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'doStuff' + Q + ',\n      javaCode: ' + BT + '\n        get\n      ' + BT + '\n    }\n  ]\n})';
-var javaRealResult = completionHandler.handle(javaRealText, { line: 7, character: 11 });
-var textItem = javaRealResult.items.find(function(i) { return i.label === 'getText()'; });
-test(textItem && textItem.detail.indexOf('String') !== -1, 'Java getter shows return type: ' + (textItem ? textItem.detail : 'not found'));
-
-// set suggests setters with parameter type
-var javaSetText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse' + Q + ',\n  name: ' + Q + 'Suggestion' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'update' + Q + ',\n      javaCode: ' + BT + '\n        set\n      ' + BT + '\n    }\n  ]\n})';
-var javaSetResult = completionHandler.handle(javaSetText, { line: 7, character: 11 });
-var setTextItem = javaSetResult.items.find(function(i) { return i.label.indexOf('setText') !== -1; });
-test(setTextItem && setTextItem.label.indexOf('String') !== -1, 'Java setter shows param type: ' + (setTextItem ? setTextItem.label : 'none'));
-
-// Empty line inside javaCode — suggests all getters AND setters
-var javaEmptyText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'JEmpty' + Q + ',\n  properties: [\n    { class: ' + Q + 'String' + Q + ', name: ' + Q + 'firstName' + Q + ' }\n  ],\n  methods: [\n    {\n      name: ' + Q + 'doStuff' + Q + ',\n      javaCode: ' + BT + '\n        \n      ' + BT + '\n    }\n  ]\n})';
-var javaEmptyResult = completionHandler.handle(javaEmptyText, { line: 10, character: 8 });
-test(javaEmptyResult.items.length > 0, 'Java empty line: suggests getters+setters: ' + javaEmptyResult.items.length);
-test(javaEmptyResult.items.some(function(i) { return i.label.indexOf('getFirstName') !== -1; }), 'Java empty line: includes getFirstName');
-test(javaEmptyResult.items.some(function(i) { return i.label.indexOf('setFirstName') !== -1; }), 'Java empty line: includes setFirstName');
-
-// === REFERENCES HANDLER TESTS ===
-
-section('ReferencesHandler');
-var refsHandler = foam.parse.lsp.handlers.ReferencesHandler.create({ index: index });
-
-// Find references to foam.u2.Element — should have many subclasses
-var refsText = 'foam.CLASS({ extends: ' + Q + 'foam.u2.Element' + Q + ' })';
-var refsResult = refsHandler.handle(refsText, { line: 0, character: 25 });
-test(refsResult.length > 10, 'References: Element has many subclasses: ' + refsResult.length);
-
-// Find references to CreatedByAware — should have implementors
-var implRefsText = 'foam.CLASS({ implements: [' + Q + 'foam.core.auth.CreatedByAware' + Q + '] })';
-var implRefsResult = refsHandler.handle(implRefsText, { line: 0, character: 30 });
-test(implRefsResult.length > 0, 'References: CreatedByAware has implementors: ' + implRefsResult.length);
-
-// === JAVA BLOCK HOVER TESTS ===
-
-section('Java Block Hover');
-
-// Hover on getter inside javaCode — shows type
-var javaHoverText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse' + Q + ',\n  name: ' + Q + 'Suggestion' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'test' + Q + ',\n      javaCode: ' + BT + '\n        getText\n      ' + BT + '\n    }\n  ]\n})';
-var javaGetterHover = hoverHandler.handle(javaHoverText, { line: 7, character: 12 });
-test(javaGetterHover != null, 'Java hover: getter shows type info');
-test(javaGetterHover && javaGetterHover.contents.value.indexOf('String') !== -1, 'Java hover: getText shows String type');
-
-// Hover on type name inside javaCode — resolves to FOAM class
-var javaTypeHoverText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse' + Q + ',\n  name: ' + Q + 'Suggestion' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'test' + Q + ',\n      javaCode: ' + BT + '\n        Suggestion\n      ' + BT + '\n    }\n  ]\n})';
-var javaTypeHover = hoverHandler.handle(javaTypeHoverText, { line: 7, character: 12 });
-test(javaTypeHover != null, 'Java hover: type name resolves to class');
-
-// Hover on enum value — shows enum info
-var enumHoverText = 'foam.CLASS({\n  package: ' + Q + 'foam.core.reflow' + Q + ',\n  name: ' + Q + 'Flow' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'test' + Q + ',\n      javaCode: ' + BT + '\n        FlowAccess.PRIVATE\n      ' + BT + '\n    }\n  ]\n})';
-var enumValHover = hoverHandler.handle(enumHoverText, { line: 7, character: 20 });
-test(enumValHover != null, 'Java hover: enum value shows info');
-
-// Cast-aware resolution: ((UserFlowAccess) o).getUserId()
-var castHoverText = 'foam.CLASS({\n  package: ' + Q + 'foam.core.reflow' + Q + ',\n  name: ' + Q + 'Flow' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'test' + Q + ',\n      javaCode: ' + BT + '\n        ((UserFlowAccess) o).getUserId()\n      ' + BT + '\n    }\n  ]\n})';
-var castMethodHover = hoverHandler.handle(castHoverText, { line: 7, character: 32 });
-test(castMethodHover != null, 'Java hover: getter after cast resolves');
-
-// resolveJavaCastType extracts cast info
-var castInfo = analyzer.resolveJavaCastType('((UserFlowAccess) o).getUserId()', {}, index);
-test(castInfo != null && castInfo.typeName === 'UserFlowAccess', 'resolveJavaCastType: extracts UserFlowAccess');
-
-// Java variable type from declaration
-var javaVarCompText = 'foam.CLASS({\n  package: ' + Q + 'foam.core.reflow' + Q + ',\n  name: ' + Q + 'Flow' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'test' + Q + ',\n      javaCode: ' + BT + '\n        User user = null;\n        user.get\n      ' + BT + '\n    }\n  ]\n})';
-var javaVarCompResult = completionHandler.handle(javaVarCompText, { line: 8, character: 16 });
-test(javaVarCompResult.items.length > 0, 'Java variable completion: user.get returns items: ' + javaVarCompResult.items.length);
-
-// Go-to-definition on method name
-var defHandler3 = foam.parse.lsp.handlers.DefinitionHandler.create({ index: index, cache: cache });
-var defMethodText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse' + Q + ',\n  name: ' + Q + 'Suggestion' + Q + ',\n  methods: [\n    function matches() { }\n  ]\n})';
-var defMethodResult = defHandler3.handle(defMethodText, { line: 4, character: 15 });
-test(defMethodResult != null, 'Go-to-definition on method resolves');
-
-// Nested document symbols — class has children
-var nestedSymHandler = foam.parse.lsp.handlers.SymbolHandler.create({ cache: cache });
-var symText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'SymTest' + Q + ',\n  properties: [\n    { class: ' + Q + 'String' + Q + ', name: ' + Q + 'foo' + Q + ' }\n  ],\n  methods: [\n    function bar() {}\n  ]\n})';
-var symResult = nestedSymHandler.handle(symText, '');
-test(symResult.length === 1, 'Nested symbols: 1 class symbol');
-test(symResult[0].children && symResult[0].children.length === 2, 'Nested symbols: 2 children (foo + bar): ' + (symResult[0].children ? symResult[0].children.length : 0));
-
-// === METHOD RETURN TYPE INFERENCE TESTS ===
-
-section('Method Return Type Inference');
-
-// resolveMethodReturnType: AuthService.getCurrentSubject returns foam.core.auth.Subject
-var retType = analyzer.resolveMethodReturnType('foam.core.auth.AuthService', 'getCurrentSubject', index);
-test(retType === 'foam.core.auth.Subject', 'Method return type: getCurrentSubject returns Subject: ' + retType);
-
-// resolveMethodReturnType: AuthService.login returns foam.core.auth.User
-var loginRetType = analyzer.resolveMethodReturnType('foam.core.auth.AuthService', 'login', index);
-test(loginRetType === 'foam.core.auth.User', 'Method return type: login returns User: ' + loginRetType);
-
-// resolveMethodReturnType: void method returns null
-var voidRetType = analyzer.resolveMethodReturnType('foam.core.auth.AuthService', 'validatePassword', index);
-test(voidRetType === null, 'Method return type: void method returns null');
-
-// var inference via cast chain: var x = ((AuthService) y).getCurrentSubject()
-var castChainText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'RetTest' + Q + ',\n  methods: [\n    {\n      name: ' + Q + 'test' + Q + ',\n      javaCode: ' + BT + '\n        var sub = ((AuthService) x.get("auth")).getCurrentSubject(x);\n        sub.text;\n      ' + BT + '\n    }\n  ]\n})';
-var castChainModel = cache.getModelAt('', castChainText, 8);
-var castChainType = analyzer.resolveJavaVariableType(castChainText, { line: 8, character: 10 }, 'sub', castChainModel, index);
-test(castChainType === 'foam.core.auth.Subject', 'Var inference: cast chain resolves to Subject: ' + castChainType);
-
-// Go-to-definition returns single result (not duplicates from refinements)
-var defSingleText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse' + Q + ',\n  name: ' + Q + 'Suggestion' + Q + ',\n  methods: [\n    function matches() { }\n  ]\n})';
-var defSingleResult = defHandler3.handle(defSingleText, { line: 4, character: 15 });
-test(defSingleResult != null, 'Definition on method: returns result');
-test( ! Array.isArray(defSingleResult) || defSingleResult.length === 1, 'Definition on method: single result (not duplicated)');
-
-// Go-to-definition resolves to correct line (not line 0)
-test(defSingleResult && defSingleResult.range && defSingleResult.range.start.line > 0 || true, 'Definition: returns non-zero line when method is not at top');
-
-// JS method return type: var sub = this.getCurrentSubject() → resolves from method.type
-var jsRetText = 'foam.CLASS({\n  package: ' + Q + 'foam.core.auth' + Q + ',\n  name: ' + Q + 'AuthService' + Q + ',\n  methods: [\n    function test() {\n      var sub = this.getCurrentSubject();\n      sub.text;\n    }\n  ]\n})';
-var jsRetModel = cache.getModelAt('', jsRetText, 6);
-var jsRetTypes = typeTracker.getVariableTypes(jsRetText, { line: 6, character: 10 }, jsRetModel, index);
-test(jsRetTypes['sub'] === 'foam.core.auth.Subject', 'JS method return type: getCurrentSubject → Subject: ' + jsRetTypes['sub']);
-
-// Incremental diagnostics — same text returns same result (cached)
-var incText = "foam.CLASS({\n  extends: 'foam.lang.FObject'\n})";
-var diags1 = diagHandler.handle(incText, 'file:///inc-test');
-var diags2 = diagHandler.handle(incText, 'file:///inc-test');
-test(diags1.length === diags2.length, 'Incremental diagnostics: same text same result');
-
-// Cast with nested parens: ((AuthService) x.get("auth")).check resolves
-var nestedCastInfo = analyzer.resolveJavaCastType('var r = ((AuthService) x.get("auth")).check(x);', {}, index);
-test(nestedCastInfo != null && nestedCastInfo.typeName === 'AuthService', 'resolveJavaCastType: nested parens in cast expr');
-test(nestedCastInfo != null && nestedCastInfo.methodName === 'check', 'resolveJavaCastType: method after nested cast');
-
-// === JRL HANDLER TESTS ===
-
-section('JrlHandler');
-var jrlHandler = foam.parse.lsp.handlers.JrlHandler.create({ index: index });
-
-// JRL hover on class value
-var jrlLine = 'p({"class":"foam.parse.Suggestion","id":1,"text":"hello"})';
-var jrlClassHover = jrlHandler.handleHover(jrlLine, { line: 0, character: 16 });
-test(jrlClassHover != null, 'JRL hover: class value shows class info');
-test(jrlClassHover && jrlClassHover.contents.value.indexOf('foam.parse.Suggestion') !== -1, 'JRL hover: class value contains class name');
-
-// JRL hover on property name — "text" starts at col 43
-var jrlPropHover = jrlHandler.handleHover(jrlLine, { line: 0, character: 43 });
-test(jrlPropHover != null, 'JRL hover: property name shows type: ' + (jrlPropHover ? 'yes' : 'null'));
-
-// JRL hover on timestamp → formatted date
-var jrlDateLine = 'p({"class":"foam.core.auth.User","id":1,"lastLogin":1735689600000})';
-var jrlDateHover = jrlHandler.handleHover(jrlDateLine, { line: 0, character: 55 });
-test(jrlDateHover != null, 'JRL hover: timestamp shows date');
-test(jrlDateHover && jrlDateHover.contents.value.indexOf('2025') !== -1, 'JRL hover: date contains year 2025');
-
-// JRL semantic tokens
-var jrlText = 'p({"class":"foam.parse.Suggestion","id":1,"text":"hello","active":true})\nc({"class":"foam.parse.Suggestion","id":2})';
-var jrlTokens = jrlHandler.handleSemanticTokens(jrlText);
-test(jrlTokens.data.length > 0, 'JRL semantic tokens: has data: ' + jrlTokens.data.length);
-test(jrlTokens.data.length % 5 === 0, 'JRL semantic tokens: multiple of 5');
-
-// JRL journal class map — resolves filename to class without "class" field
-index.buildFileIndex();
-jrlHandler.buildJournalClassMap();
-var mapSize = Object.keys(jrlHandler.journalClassMap_).length;
-test(mapSize > 0, 'JRL journal class map: ' + mapSize + ' entries');
-
-// Resolve class from URI for JRL without "class" field
-var noClassEntry = {"id": 1, "name": "test"};
-var resolvedFromMap = jrlHandler.resolveClassForJrl('file:///path/to/journals/threddCardAuthorizations.jrl', noClassEntry);
-// May or may not resolve depending on whether threddCardAuthorizations is in the map
-test(resolvedFromMap === null || typeof resolvedFromMap === 'string', 'JRL resolveClassForJrl: returns string or null');
-
-// JRL FOAM format (unquoted keys): c({key:"value",num:123})
-var foamJrlLine = 'c({summaryType:"INTERCHANGE VALUE",processDate:1741435200000,id:-4593})';
-var foamJrlEntry = jrlHandler.parseJrlEntry_(foamJrlLine);
-test(foamJrlEntry != null, 'JRL parse: FOAM unquoted-key format parses');
-test(foamJrlEntry && foamJrlEntry.summaryType === 'INTERCHANGE VALUE', 'JRL parse: string value correct');
-test(foamJrlEntry && foamJrlEntry.processDate === 1741435200000, 'JRL parse: number value correct');
-
-// JRL hover on timestamp in FOAM format
-var foamJrlDateHover = jrlHandler.handleHover(foamJrlLine, { line: 0, character: 52 });
-test(foamJrlDateHover != null || true, 'JRL hover: FOAM format timestamp (needs class resolution)');
-
-// JRL getSegmentAt_ with unquoted keys
-var seg = jrlHandler.getSegmentAt_(foamJrlLine, 4);
-test(seg != null && seg.value === 'summaryType' && seg.isKey, 'JRL segment: finds unquoted key');
-
-var segVal = jrlHandler.getSegmentAt_(foamJrlLine, 18);
-test(segVal != null && segVal.isValue, 'JRL segment: finds string value');
-
-// JRL semantic tokens — only emit for verified class values
-// Use a line with a known FOAM class for the test
-var knownClassJrl = 'p({"class":"foam.lang.FObject","id":"test1","name":"Test"})';
-var foamJrlTokens = jrlHandler.handleSemanticTokens(knownClassJrl);
-test(foamJrlTokens.data.length > 0, 'JRL semantic tokens: verified class emits tokens: ' + foamJrlTokens.data.length);
-
-// JRL shortName/alias resolution — uses inline test model
-foam.CLASS({
-  package: 'foam.parse.lsp.test',
-  name: 'JrlTestModel',
-  properties: [
-    { class: 'String', name: 'accountNo', shortName: 'an', aliases: ['acct'], label: 'Account No' },
-    { class: 'String', name: 'referenceId', aliases: ['ref', 'rid'] },
-    { class: 'Long',   name: 'id' }
-  ]
-});
-var shortNameHandler = foam.parse.lsp.handlers.JrlHandler.create({ index: index });
-var jrlTestCls = foam.lookup('foam.parse.lsp.test.JrlTestModel');
-
-var resolvedProp = shortNameHandler.resolveProperty_(jrlTestCls, 'an');
-test(resolvedProp != null, 'JRL resolveProperty: shortName an found');
-test(resolvedProp && resolvedProp.name === 'accountNo', 'JRL resolveProperty: an → accountNo');
-test(resolvedProp && resolvedProp.label === 'Account No', 'JRL resolveProperty: label is Account No');
-
-var aliasResolve = shortNameHandler.resolveProperty_(jrlTestCls, 'ref');
-test(aliasResolve != null && aliasResolve.name === 'referenceId', 'JRL resolveProperty: alias ref → referenceId');
-
-var directResolve = shortNameHandler.resolveProperty_(jrlTestCls, 'id');
-test(directResolve != null && directResolve.name === 'id', 'JRL resolveProperty: direct name id');
-
-var noResolve = shortNameHandler.resolveProperty_(jrlTestCls, 'nonExistent');
-test(noResolve == null, 'JRL resolveProperty: unknown property returns null');
-
-// JRL isJrlFile detection
-test(jrlHandler.isJrlFile('file:///test.jrl') === true, 'isJrlFile: .jrl returns true');
-test(jrlHandler.isJrlFile('file:///test.js') === false, 'isJrlFile: .js returns false');
-
-// ========== JRL Multi-line Entry Parsing ==========
-section('JRL Multi-line');
-
-var multiLineJrl = 'p({\n  "class": "foam.lang.FObject",\n  "id": "test1",\n  "name": "Test Object"\n})';
-var multiFound = jrlHandler.findEntryAtLine_(multiLineJrl, 2);
-test(multiFound != null, 'JRL multi-line: findEntryAtLine_ parses multi-line entry');
-test(multiFound && multiFound.entry['class'] === 'foam.lang.FObject', 'JRL multi-line: extracts class from multi-line');
-test(multiFound && multiFound.entry.name === 'Test Object', 'JRL multi-line: extracts name from multi-line');
-test(multiFound && multiFound.startLine === 0, 'JRL multi-line: startLine is 0');
-test(multiFound && multiFound.endLine === 4, 'JRL multi-line: endLine is 4');
-
-// Multi-line hover on class value
-var multiHover = jrlHandler.handleHover(multiLineJrl, { line: 1, character: 25 }, '');
-test(multiHover != null, 'JRL multi-line: hover on class value works');
-
-// Multi-line hover on property key
-var multiKeyHover = jrlHandler.handleHover(multiLineJrl, { line: 3, character: 4 }, '');
-// "name" key - should be detected as a key
-test(multiKeyHover != null || true, 'JRL multi-line: hover on property key attempted');
-
-// Multi-line single-line still works
-var singleLineJrl = 'p({"class":"foam.lang.FObject","id":"s1"})';
-var singleFound = jrlHandler.findEntryAtLine_(singleLineJrl, 0);
-test(singleFound != null, 'JRL multi-line: single-line still works via findEntryAtLine_');
-
-// ========== JRL Completions ==========
-section('JRL Completions');
-
-var completionJrl = 'p({"class":"foam.parse.lsp.test.JrlTestModel","id":1})';
-var completionResult = jrlHandler.handleCompletion(completionJrl, { line: 0, character: 50 }, '');
-test(completionResult != null, 'JRL completion: returns result');
-test(completionResult && completionResult.items.length > 0, 'JRL completion: has items');
-
-// Should suggest properties from JrlTestModel
-var hasAccountNo = completionResult && completionResult.items.some(function(item) { return item.label === 'accountNo'; });
-test(hasAccountNo, 'JRL completion: suggests accountNo from class');
-
-// Should suggest shortName 'an'
-var hasShortName = completionResult && completionResult.items.some(function(item) { return item.label === 'an'; });
-test(hasShortName, 'JRL completion: suggests shortName an');
-
-// Should NOT suggest already-present 'id' property
-var hasId = completionResult && completionResult.items.some(function(item) { return item.label === 'id'; });
-test(!hasId, 'JRL completion: does not suggest already-present id');
-
-// Multi-line completion
-var multiCompJrl = 'p({\n  "class": "foam.parse.lsp.test.JrlTestModel",\n  "id": 1,\n  \n})';
-var multiCompResult = jrlHandler.handleCompletion(multiCompJrl, { line: 3, character: 2 }, '');
-test(multiCompResult != null && multiCompResult.items.length > 0, 'JRL completion: works on multi-line entry');
-
-// Class name completion
-var classCompJrl = 'p({"class":""})';
-var classCompResult = jrlHandler.handleCompletion(classCompJrl, { line: 0, character: 12 }, '');
-test(classCompResult == null || true, 'JRL completion: class name completion attempted');
-
-// ========== JRL Diagnostics ==========
-section('JRL Diagnostics');
-
-// Unknown class
-var diagUnknownClass = 'p({"class":"com.nonexistent.FakeClass123","id":1})';
-var diags1 = jrlHandler.handleDiagnostics(diagUnknownClass, '');
-test(diags1.length > 0, 'JRL diagnostics: unknown class produces error');
-test(diags1[0] && diags1[0].severity === 1, 'JRL diagnostics: unknown class is severity 1 (error)');
-
-// Unknown property
-var diagUnknownProp = 'p({"class":"foam.parse.lsp.test.JrlTestModel","id":1,"nonExistentProp":"val"})';
-var diags2 = jrlHandler.handleDiagnostics(diagUnknownProp, '');
-test(diags2.length > 0, 'JRL diagnostics: unknown property produces warning');
-test(diags2[0] && diags2[0].severity === 2, 'JRL diagnostics: unknown property is severity 2 (warning)');
-
-// Valid entry — no diagnostics
-var diagValid = 'p({"class":"foam.parse.lsp.test.JrlTestModel","id":1,"accountNo":"123"})';
-var diags3 = jrlHandler.handleDiagnostics(diagValid, '');
-test(diags3.length === 0, 'JRL diagnostics: valid entry produces no diagnostics');
-
-// Multi-line diagnostics
-var diagMulti = 'p({\n  "class": "com.nonexistent.FakeClass456",\n  "id": 1\n})';
-var diags4 = jrlHandler.handleDiagnostics(diagMulti, '');
-test(diags4.length > 0, 'JRL diagnostics: multi-line unknown class detected');
-
-// Comment lines should not produce diagnostics
-var diagComment = '// This is a comment\np({"class":"foam.parse.lsp.test.JrlTestModel","id":1})';
-var diags5 = jrlHandler.handleDiagnostics(diagComment, '');
-test(diags5.length === 0, 'JRL diagnostics: comment lines ignored');
-
-// ========== JRL Nested Class Context ==========
-section('JRL Nested Class');
-
-var nestedJrl = 'p({\n  "class": "foam.parse.lsp.test.JrlTestModel",\n  "id": 1,\n  "nested": {\n    "class": "foam.lang.FObject",\n    "id": "inner"\n  }\n})';
-// Line 5 is inside the nested object with class FObject
-var nestedClass = jrlHandler.resolveNearestClass_(nestedJrl, 5, '', null);
-test(nestedClass === 'foam.lang.FObject', 'JRL nested: resolves inner class at nested depth: ' + nestedClass);
-
-// Line 2 is at top level with class JrlTestModel
-var topClass = jrlHandler.resolveNearestClass_(nestedJrl, 2, '', null);
-test(topClass === 'foam.parse.lsp.test.JrlTestModel', 'JRL nested: resolves outer class at top level: ' + topClass);
-
-// ========== JRL Command Hovers ==========
-section('JRL Command Hovers');
-
-var cmdPLine = 'p({"class":"foam.lang.FObject","id":"t"})';
-var cmdPHover = jrlHandler.handleHover(cmdPLine, { line: 0, character: 0 }, '');
-test(cmdPHover != null, 'JRL cmd hover: p returns hover');
-test(cmdPHover && cmdPHover.contents.value.indexOf('Put') !== -1, 'JRL cmd hover: p mentions Put');
-
-var cmdRLine = 'r({"class":"foam.lang.FObject","id":"t"})';
-var cmdRHover = jrlHandler.handleHover(cmdRLine, { line: 0, character: 0 }, '');
-test(cmdRHover != null, 'JRL cmd hover: r returns hover');
-test(cmdRHover && cmdRHover.contents.value.indexOf('Remove') !== -1, 'JRL cmd hover: r mentions Remove');
-
-var cmdCLine = 'c({"class":"foam.lang.FObject","id":"t"})';
-var cmdCHover = jrlHandler.handleHover(cmdCLine, { line: 0, character: 0 }, '');
-test(cmdCHover != null, 'JRL cmd hover: c returns hover');
-test(cmdCHover && cmdCHover.contents.value.indexOf('Create') !== -1, 'JRL cmd hover: c mentions Create');
-
-var cmdVLine = 'v({"class":"foam.lang.FObject","id":"t"})';
-var cmdVHover = jrlHandler.handleHover(cmdVLine, { line: 0, character: 0 }, '');
-test(cmdVHover != null, 'JRL cmd hover: v returns hover');
-test(cmdVHover && cmdVHover.contents.value.indexOf('Version') !== -1, 'JRL cmd hover: v mentions Version');
-
-// ========== JRL Semantic Tokens (slim) ==========
-section('JRL Semantic Tokens Slim');
-
-// Unknown class should NOT emit tokens
-var unknownClassJrl = 'p({"class":"com.fake.NonExistent","id":1})';
-var unknownTokens = jrlHandler.handleSemanticTokens(unknownClassJrl);
-test(unknownTokens.data.length === 0, 'JRL tokens: unknown class emits zero tokens');
-
-// Empty/comment lines
-var commentOnlyJrl = '// just a comment\n\n// another';
-var commentTokens = jrlHandler.handleSemanticTokens(commentOnlyJrl);
-test(commentTokens.data.length === 0, 'JRL tokens: comment-only lines emit zero tokens');
-
-// ========== Hover UI Format ==========
-section('Hover UI Format');
-
-// Class hover should use code block
-var classHover = hoverHandler.handle(requiresHoverText, { line: 2, character: 20 });
-test(classHover != null && classHover.contents.value.indexOf('```foam') !== -1, 'Hover UI: class hover uses code block');
-test(classHover != null && classHover.contents.value.indexOf('| Property') !== -1, 'Hover UI: class hover has property table');
-
-// Method hover format check — buildMethodHover_ should use JS code block
-var fakeMethod = { name: 'testMethod', args: ['x', 'y'], documentation: 'A test method.' };
-var methodMd = hoverHandler.buildMethodHover_(fakeMethod, 'foam.test.FakeClass');
-test(methodMd.indexOf('```javascript') !== -1, 'Hover UI: method hover uses JS code block');
-test(methodMd.indexOf('foam.test.FakeClass') !== -1, 'Hover UI: method hover shows class name');
-
-// Create hover should use code block with .create()
-var createHover = hoverHandler.handle(requiresHoverText, { line: 6, character: 22 });
-test(createHover != null && createHover.contents.value.indexOf('.create()') !== -1, 'Hover UI: create hover shows .create()');
-
-// ========== this.RequiredClass.create() Completion ==========
-section('RequiredClass Completion');
-
-var reqClassText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse.lsp.test' + Q + ',\n  name: ' + Q + 'ReqTest' + Q + ',\n  requires: [' + Q + 'foam.parse.Suggestion' + Q + '],\n  methods: [\n    function test() {\n      this.Suggestion.\n    }\n  ]\n})';
-var reqClassResult = memberHandler.handle(reqClassText, { line: 6, character: 22 });
-test(reqClassResult != null && reqClassResult.items.length > 0, 'RequiredClass completion: returns items for this.Suggestion.');
-
-var hasCreate = reqClassResult && reqClassResult.items.some(function(item) { return item.label === 'create'; });
-test(hasCreate, 'RequiredClass completion: includes create()');
-
-var hasIsInstance = reqClassResult && reqClassResult.items.some(function(item) { return item.label === 'isInstance'; });
-test(hasIsInstance, 'RequiredClass completion: includes isInstance()');
-
-// ========== Java Block Variable Hover ==========
-section('Java Block Variable Hover');
-
-var javaVarHoverText = 'foam.CLASS({\n  package: ' + Q + 'foam.parse.lsp.test' + Q + ',\n  name: ' + Q + 'VarHoverTest' + Q + ',\n  javaCode: `\n    FObject obj = new FObject();\n    obj.fclone();\n  `\n})';
-// Hover on "fclone" at line 5 — should resolve obj → FObject → fclone method
-var fcloneHover = hoverHandler.handle(javaVarHoverText, { line: 5, character: 8 });
-// This may or may not resolve depending on the variable tracking, but the path should not crash
-test(fcloneHover != null || true, 'Java hover: variable.method() does not crash');
-
-// Hover on "FObject" type name in Java block
-var fobjectHover = hoverHandler.handle(javaVarHoverText, { line: 4, character: 5 });
-test(fobjectHover != null, 'Java hover: type name FObject resolves in Java block');
-
-// ========== Java Block: variable.method() Hover Regression Tests ==========
-section('Java variable.method() Hover');
-
-// Simulate DAONotificationTest.js javaCode block
-var javaMethodText = 'foam.CLASS({\n  package: ' + Q + 'foam.core.notification.test' + Q + ',\n  name: ' + Q + 'DAONotificationTest' + Q + ',\n  javaCode: `\n    Country country = (Country) countryDAO.find("CA");\n    country = (Country) country.fclone();\n    country.setName("Canada Eh!");\n  `\n})';
-
-// Debug: check if backtick block is detected at line 5 (country.fclone line)
-var blockCtx = hoverHandler.analyzer.getBacktickBlockContext(javaMethodText, { line: 5, character: 30 });
-test(blockCtx != null, 'Java var.method: backtick block detected at line 5: ' + JSON.stringify(blockCtx));
-test(blockCtx && blockCtx.blockKey && blockCtx.blockKey.indexOf('java') !== -1, 'Java var.method: block key is java*: ' + (blockCtx ? blockCtx.blockKey : 'null'));
-
-// Debug: check variable type resolution
-var javaModel = hoverHandler.cache.getModelAt('', javaMethodText, 5);
-test(javaModel != null, 'Java var.method: model found at line 5');
-
-if ( blockCtx && javaModel ) {
-  var countryType = hoverHandler.analyzer.resolveJavaVariableType(javaMethodText, { line: 5, character: 30 }, 'country', javaModel, hoverHandler.index);
-  test(countryType != null, 'Java var.method: country resolves to type: ' + countryType);
-}
-
-// fclone hover — Java-only FObject method, resolved via fallback constant map
-var fcloneHover = hoverHandler.handle(javaMethodText, { line: 5, character: 33 }, '');
-test(fcloneHover != null, 'Java var.method: hover on country.fclone() returns result');
-test(fcloneHover && fcloneHover.contents.value.indexOf('fclone') !== -1, 'Java var.method: fclone hover mentions fclone');
-
-// setName should work (it's a getter/setter)
-var setNameHover = hoverHandler.handle(javaMethodText, { line: 6, character: 13 }, '');
-test(setNameHover != null, 'Java var.method: hover on country.setName() returns result');
-
-// x.get() hover — x is always foam.lang.X
-var xGetText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'XTest' + Q + ',\n  javaCode: `\n    DAO dao = (DAO) x.get("countryDAO");\n  `\n})';
-var xGetHover = hoverHandler.handle(xGetText, { line: 4, character: 23 }, '');
-test(xGetHover != null, 'Java x.get: hover on x.get() returns result');
-test(xGetHover && xGetHover.contents.value.indexOf('foam.lang.X') !== -1, 'Java x.get: hover mentions foam.lang.X');
-
-// ========== Java File Method Scanner ==========
-section('Java Method Scanner');
-
-// FoamIndex.getJavaMethods should find Java-only methods from .java files
-var fobjectJavaMethods = index.getJavaMethods('foam.lang.FObject');
-test(fobjectJavaMethods.length > 0, 'Java scanner: FObject has Java-only methods: ' + fobjectJavaMethods.length);
-
-var fcloneFound = fobjectJavaMethods.some(function(m) { return m.name === 'fclone'; });
-test(fcloneFound, 'Java scanner: fclone found in FObject Java methods');
-
-var deepCloneFound = fobjectJavaMethods.some(function(m) { return m.name === 'deepClone'; });
-test(deepCloneFound, 'Java scanner: deepClone found in FObject Java methods');
-
-// Log all Java-only method names for debugging
-var javaMethodNames = fobjectJavaMethods.map(function(m) { return m.name; });
-test(fobjectJavaMethods.length >= 10, 'Java scanner: FObject has at least 10 Java-only methods: ' + javaMethodNames.join(', '));
-
-// fclone should have a signature
-var fcloneMethod = fobjectJavaMethods.find(function(m) { return m.name === 'fclone'; });
-test(fcloneMethod && fcloneMethod.sig.indexOf('FObject') !== -1, 'Java scanner: fclone sig has FObject: ' + (fcloneMethod ? fcloneMethod.sig : ''));
-
-// Java methods should NOT include FOAM axiom methods (they're in getMethods)
-var foamMethods = index.getMethods('foam.lang.FObject');
-var foamMethodNames = {};
-foamMethods.forEach(function(m) { foamMethodNames[m.name] = true; });
-var noDuplicates = fobjectJavaMethods.every(function(m) { return ! foamMethodNames[m.name]; });
-test(noDuplicates, 'Java scanner: no overlap with FOAM axiom methods');
-
-// Inheritance: Country should inherit FObject Java methods
-var countryJavaMethods = index.getJavaMethods('foam.core.auth.Country');
-var countryHasFclone = countryJavaMethods.some(function(m) { return m.name === 'fclone'; });
-test(countryHasFclone, 'Java scanner: Country inherits fclone from FObject');
-
-// Go-to-definition for Java-only methods
-var defHandler = foam.parse.lsp.handlers.DefinitionHandler.create({ index: index, cache: foam.parse.lsp.FileModelCache.create() });
-var fcloneJavaLoc = defHandler.findJavaMethodLocation_('foam.lang.FObject', 'fclone');
-test(fcloneJavaLoc != null, 'Java go-to-def: fclone resolves to a .java file location');
-test(fcloneJavaLoc && fcloneJavaLoc.uri.indexOf('.java') !== -1, 'Java go-to-def: URI is a .java file');
-
-// ========== Java Block: Complex Variable Declarations ==========
-section('Java Complex Declarations');
-
-var complexJavaText = 'foam.CLASS({\n  package: ' + Q + 'test' + Q + ',\n  name: ' + Q + 'ComplexTest' + Q + ',\n  javaCode: ' + '`' + '\n    EmailMessage msg = null;\n    for ( EmailMessage m : messages ) { break; }\n    try { } catch ( Exception e ) { }\n  ' + '`' + '\n})';
-
-try {
-  var complexTokens = semanticHandler.handle(complexJavaText, '');
-  var complexTokenCount = complexTokens.data.length / 5;
-  test(complexTokenCount > 0, 'Complex Java: produces semantic tokens: ' + complexTokenCount);
-} catch (e) {
-  test(false, 'Complex Java: semantic tokens crashed: ' + e.message);
-}
-
-// Test that emailMessages is tracked as a declared variable (via generic type)
-// The semantic tokens should include entries for emailMessages
-var complexLines = complexJavaText.split('\n');
-// Check that the generic declaration line produces variable tokens
-test(complexTokenCount > 5, 'Complex Java: enough tokens for generic + for-each + catch');
-
-// ========== Java Block: Go-to-Definition ==========
-section('Java Go-to-Definition');
-
-var javaDefText = 'foam.CLASS({\n  package: ' + Q + 'test.def' + Q + ',\n  name: ' + Q + 'JavaDefTest' + Q + ',\n  javaImports: [' + Q + 'foam.core.auth.Country' + Q + '],\n  javaCode: `\n    Country c = new Country();\n    c.fclone();\n  `\n})';
-
-var defHandler = foam.parse.lsp.handlers.DefinitionHandler.create({
-  index: index,
-  cache: foam.parse.lsp.FileModelCache.create()
-});
-
-// Go-to-def on "Country" type name in Java block (line 5, char 5)
-var countryDef = defHandler.handle(javaDefText, { line: 5, character: 5 });
-test(countryDef != null, 'Java go-to-def: Country type name resolves');
-test(countryDef && countryDef.uri && countryDef.uri.indexOf('Country') !== -1, 'Java go-to-def: Country navigates to correct file');
-
-// Go-to-def on "fclone" Java-only method (line 6, char 7)
-var fcloneDef = defHandler.handle(javaDefText, { line: 6, character: 7 });
-test(fcloneDef != null, 'Java go-to-def: fclone resolves to .java file');
-test(fcloneDef && fcloneDef.uri && fcloneDef.uri.indexOf('.java') !== -1, 'Java go-to-def: fclone URI is a .java file');
-
-// ========== JRL Go-to-Definition ==========
-section('JRL Go-to-Definition');
-
-// Go-to-def on class value navigates to the class .js file
-var jrlDefText = 'p({"class":"foam.lang.FObject","id":"x"})';
-var jrlClassDef = jrlHandler.handleDefinition(jrlDefText, { line: 0, character: 15 }, '');
-test(jrlClassDef != null, 'JRL go-to-def: class value returns location');
-test(jrlClassDef && jrlClassDef.uri && jrlClassDef.uri.indexOf('FObject') !== -1, 'JRL go-to-def: navigates to FObject file');
-
-// Go-to-def on property key navigates to property in class file
-var jrlPropDefText = 'p({"class":"foam.parse.lsp.test.JrlTestModel","id":1,"accountNo":"123"})';
-var jrlPropDef = jrlHandler.handleDefinition(jrlPropDefText, { line: 0, character: 56 }, '');
-test(jrlPropDef != null || true, 'JRL go-to-def: property key (in test model — may not have file)');
-
-// Go-to-def on unknown class returns null
-var jrlBadDef = 'p({"class":"com.fake.Bad","id":1})';
-var jrlBadResult = jrlHandler.handleDefinition(jrlBadDef, { line: 0, character: 15 }, '');
-test(jrlBadResult == null, 'JRL go-to-def: unknown class returns null');
-
-// ========== JavaParser (FOAM Grammar-based) ==========
-section('JavaParser');
-
-var javaParser = foam.parse.lsp.JavaParser.create();
-
-var sampleJava = [
-  'package foam.test;',
-  '',
-  'import java.util.List;',
-  'import static foo.Bar.BAZ;',
-  '',
-  'public interface MyClass {',
-  '  /** Clone this object. */',
-  '  default MyClass myClone() {',
-  '    return null;',
-  '  }',
-  '',
-  '  public List<String> getItems(int n) throws Exception;',
-  '',
-  '  abstract Map<K,V> diff(Object o);',
-  '}'
-].join('\n');
-
-var parsed = javaParser.parseFile(sampleJava);
-test(parsed['package'] === 'foam.test', 'JavaParser: package extracted: ' + parsed['package']);
-test(parsed.imports.length === 2, 'JavaParser: 2 imports extracted: ' + parsed.imports.length);
-test(parsed.imports[0].name === 'java.util.List', 'JavaParser: first import name');
-test(parsed.imports[1].name === 'foo.Bar.BAZ', 'JavaParser: static import name');
-test(parsed.classes.length === 1, 'JavaParser: 1 class extracted: ' + parsed.classes.length);
-test(parsed.classes[0].name === 'MyClass', 'JavaParser: class name');
-test(parsed.classes[0].kind === 'interface', 'JavaParser: class kind');
-test(parsed.methods.length === 3, 'JavaParser: 3 methods extracted: ' + parsed.methods.length);
-
-var myCloneMethod = parsed.methods.find(function(m) { return m.name === 'myClone'; });
-test(myCloneMethod != null, 'JavaParser: myClone method found');
-test(myCloneMethod && myCloneMethod.returnType === 'MyClass', 'JavaParser: myClone return type');
-test(myCloneMethod && myCloneMethod.doc.indexOf('Clone') !== -1, 'JavaParser: javadoc extracted');
-test(myCloneMethod && myCloneMethod.modifiers.indexOf('default') !== -1, 'JavaParser: default modifier captured');
-
-var getItemsMethod = parsed.methods.find(function(m) { return m.name === 'getItems'; });
-test(getItemsMethod && getItemsMethod.returnType === 'List<String>', 'JavaParser: generic return type: ' + (getItemsMethod ? getItemsMethod.returnType : ''));
-test(getItemsMethod && getItemsMethod.params === 'int n', 'JavaParser: params extracted');
-
-// FOAM-aware: scan a real .java file via the index
-var fobjectMethods2 = index.getJavaMethods('foam.lang.FObject');
-test(fobjectMethods2.length > 0, 'JavaParser via index: FObject methods: ' + fobjectMethods2.length);
-var fcloneMethod2 = fobjectMethods2.find(function(m) { return m.name === 'fclone'; });
-test(fcloneMethod2 && fcloneMethod2.line > 0, 'JavaParser via index: fclone has line number: ' + (fcloneMethod2 ? fcloneMethod2.line : ''));
-
-// ========== Documentation Formatting in Hover ==========
-section('Hover Doc Formatting');
-
-var multiParagraphDoc = '\n    First paragraph line.\n    Continues here.\n\n    Entry points:\n      - one\n      - two\n\n    Final paragraph.\n  ';
-var formatted = hoverHandler.formatDocumentation_(multiParagraphDoc);
-test(formatted.indexOf('First paragraph line.') === 0, 'Doc format: dedents leading indent');
-test(formatted.indexOf('\n\nEntry points:') !== -1, 'Doc format: preserves paragraph breaks');
-test(formatted.indexOf('  - one  ') !== -1 || formatted.indexOf('- one  ') !== -1, 'Doc format: indented list items get hard break');
-test(formatted.indexOf('Final paragraph.') !== -1, 'Doc format: keeps last paragraph');
-
-// Class hover should wrap docs in blockquote
-var docHoverClassId = 'foam.parse.lsp.JavaGrammar';
-if ( index.classExists(docHoverClassId) ) {
-  var docClassHover = hoverHandler.buildClassHover(docHoverClassId);
-  test(docClassHover != null, 'Doc hover: class hover returned');
-  test(docClassHover && docClassHover.contents.value.indexOf('> ') !== -1, 'Doc hover: documentation rendered as quoted block');
-  test(docClassHover && docClassHover.contents.value.indexOf('> ') !== -1, 'Doc hover: blockquote for docs');
-}
-
-// ========== Class Signature Multi-line Format ==========
-section('Hover Signature Format');
-
-// Inline test class with multiple implements
-foam.INTERFACE({ package: 'foam.parse.lsp.test', name: 'IFoo' });
-foam.INTERFACE({ package: 'foam.parse.lsp.test', name: 'IBar' });
-foam.INTERFACE({ package: 'foam.parse.lsp.test', name: 'IBaz' });
-foam.CLASS({
-  package: 'foam.parse.lsp.test',
-  name: 'MultiImplTest',
-  implements: ['foam.parse.lsp.test.IFoo', 'foam.parse.lsp.test.IBar', 'foam.parse.lsp.test.IBaz']
-});
-
-var multiImplHover = hoverHandler.buildClassHover('foam.parse.lsp.test.MultiImplTest');
-test(multiImplHover != null, 'Sig format: hover returned');
-test(multiImplHover && multiImplHover.contents.value.indexOf('implements foam.parse.lsp.test.IFoo, foam.parse.lsp.test.IBar') !== -1,
-  'Sig format: multiple implements on same line, comma-separated');
-
-// Single implement stays on one line
-foam.CLASS({
-  package: 'foam.parse.lsp.test',
-  name: 'SingleImplTest',
-  implements: ['foam.parse.lsp.test.IFoo']
-});
-var singleImplHover = hoverHandler.buildClassHover('foam.parse.lsp.test.SingleImplTest');
-test(singleImplHover && singleImplHover.contents.value.indexOf('implements foam.parse.lsp.test.IFoo') !== -1, 'Sig format: single implement inline');
-
-// === JRL LOADER ===
-
-section('JrlLoader');
-
-var jrlLoader = foam.parse.lsp.JrlLoader.create();
-
-// Test p() puts objects
-var result1 = jrlLoader.loadString('p({"class":"foam.core.theme.Theme","id":"theme1","name":"Test Theme"})');
-test(result1.length === 1, 'JrlLoader: p() collects one object');
-test(result1[0].id === 'theme1', 'JrlLoader: p() preserves id');
-test(result1[0].name === 'Test Theme', 'JrlLoader: p() preserves name');
-test(result1[0]['class'] === 'foam.core.theme.Theme', 'JrlLoader: p() preserves class');
-
-// Test c() creates objects (same as p)
-var result2 = jrlLoader.loadString('c({"class":"foam.core.theme.Theme","id":"theme2","name":"Created"})');
-test(result2.length === 1, 'JrlLoader: c() collects object');
-
-// Test r() removes objects
-var result3 = jrlLoader.loadString(
-  'p({"class":"foam.core.theme.Theme","id":"t1","name":"A"})\n' +
-  'p({"class":"foam.core.theme.Theme","id":"t2","name":"B"})\n' +
-  'r({"class":"foam.core.theme.Theme","id":"t1"})'
-);
-test(result3.length === 1, 'JrlLoader: r() removes object by id');
-test(result3[0].id === 't2', 'JrlLoader: r() keeps non-removed objects');
-
-// Test multiple p() calls
-var result4 = jrlLoader.loadString(
-  'p({"class":"test.A","id":"1","value":"x"})\n' +
-  'p({"class":"test.B","id":"2","value":"y"})\n' +
-  'p({"class":"test.A","id":"3","value":"z"})'
-);
-test(result4.length === 3, 'JrlLoader: multiple p() calls collected');
-
-// Test empty/comment lines
-var result5 = jrlLoader.loadString('// comment\n\np({"id":"1"})\n\n// another comment');
-test(result5.length === 1, 'JrlLoader: skips comments and empty lines');
-
-// Test malformed JRL (should not throw)
-var result6 = jrlLoader.loadString('p({invalid json})');
-test(result6.length === 0, 'JrlLoader: malformed JRL returns empty, no throw');
-
-// Test filterByClass
-var result7 = jrlLoader.loadString(
-  'p({"class":"foam.core.theme.Theme","id":"t1"})\n' +
-  'p({"class":"foam.core.theme.customisation.CSSTokenOverride","id":"o1","source":"primary400","target":"#000"})'
-);
-var themes = jrlLoader.filterByClass(result7, 'foam.core.theme.Theme');
-var overrides = jrlLoader.filterByClass(result7, 'foam.core.theme.customisation.CSSTokenOverride');
-test(themes.length === 1, 'JrlLoader: filterByClass returns matching class');
-test(overrides.length === 1, 'JrlLoader: filterByClass returns other class');
-
-// === CSS TOKEN RESOLUTION ===
-
-section('CSSTokenResolver — value resolution');
-// cssTokenResolver is declared earlier (before HoverHandler init).
-
-// Test resolveTokenValue on base tokens
-var resolvedPrimary = cssTokenResolver.resolveTokenValue('primary400');
-test(resolvedPrimary != null && resolvedPrimary !== '', 'resolveTokenValue: primary400 resolves to a value');
-test(resolvedPrimary && resolvedPrimary.charAt(0) === '#', 'resolveTokenValue: primary400 resolves to hex color');
-
-// Test resolveTokenValue on unknown token
-var resolvedUnknown = cssTokenResolver.resolveTokenValue('totallyFakeToken');
-test(resolvedUnknown === null, 'resolveTokenValue: unknown token returns null');
-
-// Test resolveTokenValue on recursive $-references
-var resolvedBg = cssTokenResolver.resolveTokenValue('backgroundDefault');
-test(resolvedBg != null && resolvedBg.charAt(0) === '#', 'resolveTokenValue: backgroundDefault resolves through $white to hex');
-
-// Test that loadFromJournals still works (uses JrlLoader internally now)
-var resolver2 = foam.parse.lsp.CSSTokenResolver.create();
-resolver2.loadFromRegistry();
-resolver2.loadFromJournals();
-test(Object.keys(resolver2.themeNames_).length >= 0, 'loadFromJournals: runs without error using JrlLoader');
-
-// === TABLE COLUMNS / SEARCH COLUMNS ===
-
-section('tableColumns / searchColumns');
-
-// Register a test model for column validation
-foam.CLASS({
-  package: 'foam.parse.lsp.test',
-  name: 'ColumnTestModel',
-  extends: 'foam.lang.FObject',
-  properties: [
-    { class: 'Long', name: 'id' },
-    { class: 'String', name: 'firstName' },
-    { class: 'String', name: 'lastName' },
-    { class: 'Boolean', name: 'active' }
-  ]
-});
-
-// Completion: cursor inside tableColumns array with partial 'fi' should filter matches
-var colCompText = "foam.CLASS({\n  package: 'foam.parse.lsp.test',\n  name: 'ColumnTestModel',\n  extends: 'foam.lang.FObject',\n  properties: [\n    { class: 'Long', name: 'id' },\n    { class: 'String', name: 'firstName' },\n    { class: 'String', name: 'lastName' },\n    { class: 'Boolean', name: 'active' }\n  ],\n  tableColumns: ['fi']\n})";
-var colCompResult = completionHandler.handle(colCompText, { line: 10, character: 20 });
-var colItems = colCompResult.items || [];
-test(colItems.some(function(it) { return it.label === 'firstName'; }), 'tableColumns: partial "fi" suggests firstName');
-
-// Completion: empty partial should suggest all own properties
-var colCompText2 = "foam.CLASS({\n  package: 'foam.parse.lsp.test',\n  name: 'ColumnTestModel',\n  extends: 'foam.lang.FObject',\n  properties: [\n    { class: 'Long', name: 'id' },\n    { class: 'String', name: 'firstName' },\n    { class: 'String', name: 'lastName' },\n    { class: 'Boolean', name: 'active' }\n  ],\n  tableColumns: ['']\n})";
-var colCompResult2 = completionHandler.handle(colCompText2, { line: 10, character: 19 });
-var colItems2 = colCompResult2.items || [];
-test(colItems2.some(function(it) { return it.label === 'firstName'; }), 'tableColumns: suggests firstName');
-test(colItems2.some(function(it) { return it.label === 'lastName'; }), 'tableColumns: suggests lastName');
-test(colItems2.some(function(it) { return it.label === 'id'; }), 'tableColumns: suggests id');
-
-// Completion: searchColumns too
-var searchCompText = "foam.CLASS({\n  package: 'foam.parse.lsp.test',\n  name: 'ColumnTestModel',\n  extends: 'foam.lang.FObject',\n  properties: [\n    { class: 'Long', name: 'id' },\n    { class: 'String', name: 'firstName' }\n  ],\n  searchColumns: ['']\n})";
-var searchCompResult = completionHandler.handle(searchCompText, { line: 8, character: 19 });
-var searchItems = searchCompResult.items || [];
-test(searchItems.some(function(it) { return it.label === 'firstName'; }), 'searchColumns: suggests firstName');
-
-// tableColumns with requires above — should suggest props, not classes
-var colWithReqText = "foam.CLASS({\n  package: 'test',\n  name: 'ColReqTest',\n  requires: ['foam.u2.DetailView'],\n  properties: [\n    { class: 'String', name: 'myProp' }\n  ],\n  tableColumns: ['']\n})";
-var colWithReqResult = completionHandler.handle(colWithReqText, { line: 7, character: 19 });
-var colWithReqItems = colWithReqResult.items || [];
-test(colWithReqItems.some(function(it) { return it.label === 'myProp'; }), 'tableColumns: suggests props not classes when requires is above');
-test( ! colWithReqItems.some(function(it) { return it.label === 'foam.u2.DetailView'; }), 'tableColumns: does NOT suggest class names');
-
-// Diagnostics: valid column name — no warning
-var validColText = "foam.CLASS({\n  package: 'foam.parse.lsp.test',\n  name: 'ColDiagTest',\n  extends: 'foam.lang.FObject',\n  properties: [\n    { class: 'String', name: 'firstName' }\n  ],\n  tableColumns: ['firstName']\n})";
-var validColDiags = diagHandler.handle(validColText);
-var colWarns = validColDiags.filter(function(d) { return d.message.indexOf('firstName') !== -1 && d.message.indexOf('does not exist') !== -1; });
-test(colWarns.length === 0, 'tableColumns: valid property NOT flagged');
-
-// Diagnostics: invalid column name — should warn
-var invalidColText = "foam.CLASS({\n  package: 'foam.parse.lsp.test',\n  name: 'ColDiagTest2',\n  extends: 'foam.lang.FObject',\n  properties: [\n    { class: 'String', name: 'firstName' }\n  ],\n  tableColumns: ['nonExistent']\n})";
-var invalidColDiags = diagHandler.handle(invalidColText);
-test(invalidColDiags.some(function(d) { return d.message.indexOf('nonExistent') !== -1 && d.message.indexOf('does not exist') !== -1; }), 'tableColumns: invalid property IS flagged');
-
-// Diagnostics: inherited property — should NOT warn
-var inheritedColText = "foam.CLASS({\n  package: 'foam.parse.lsp.test',\n  name: 'ColDiagTest3',\n  extends: 'foam.parse.lsp.test.ColumnTestModel',\n  tableColumns: ['firstName']\n})";
-var inheritedColDiags = diagHandler.handle(inheritedColText);
-var inheritedWarns = inheritedColDiags.filter(function(d) { return d.message.indexOf('firstName') !== -1 && d.message.indexOf('does not exist') !== -1; });
-test(inheritedWarns.length === 0, 'tableColumns: inherited property NOT flagged');
-
-// Diagnostics: searchColumns works too
-var invalidSearchText = "foam.CLASS({\n  package: 'foam.parse.lsp.test',\n  name: 'ColDiagTest4',\n  extends: 'foam.lang.FObject',\n  properties: [\n    { class: 'String', name: 'x' }\n  ],\n  searchColumns: ['bogus']\n})";
-var invalidSearchDiags = diagHandler.handle(invalidSearchText);
-test(invalidSearchDiags.some(function(d) { return d.message.indexOf('bogus') !== -1; }), 'searchColumns: invalid property IS flagged');
-
-// === RAW CSS VALUE DIAGNOSTICS ===
-
-section('Raw CSS Value Diagnostics');
-
-// Hex color in css: template string — should warn
-var hexCssText = "foam.CLASS({\n  package: 'test',\n  name: 'HexTest',\n  css: `\n    ^ { color: #FF0000; }\n  `\n})";
-var hexDiags = diagHandler.handle(hexCssText);
-test(hexDiags.some(function(d) { return /raw color/i.test(d.message) && d.message.indexOf('#FF0000') !== -1; }), 'Raw CSS: hex color in css: flagged');
-
-// rgb() in css: — should warn
-var rgbCssText = "foam.CLASS({\n  package: 'test',\n  name: 'RgbTest',\n  css: `\n    ^ { background-color: rgb(255, 0, 0); }\n  `\n})";
-var rgbDiags = diagHandler.handle(rgbCssText);
-test(rgbDiags.some(function(d) { return /raw color/i.test(d.message) && d.message.indexOf('rgb(') !== -1; }), 'Raw CSS: rgb() in css: flagged');
-
-// $token reference — should NOT warn
-var tokenCssText = "foam.CLASS({\n  package: 'test',\n  name: 'TokenTest',\n  css: `\n    ^ { color: $primary400; }\n  `\n})";
-var tokenDiags = diagHandler.handle(tokenCssText);
-var tokenRawWarns = tokenDiags.filter(function(d) { return /raw color/i.test(d.message); });
-test(tokenRawWarns.length === 0, 'Raw CSS: $token NOT flagged');
-
-// Non-color property — should NOT warn
-var widthCssText = "foam.CLASS({\n  package: 'test',\n  name: 'WidthTest',\n  css: `\n    ^ { width: 100px; height: 50px; }\n  `\n})";
-var widthDiags = diagHandler.handle(widthCssText);
-var widthRawWarns = widthDiags.filter(function(d) { return /raw color/i.test(d.message); });
-test(widthRawWarns.length === 0, 'Raw CSS: width/height NOT flagged');
-
-// Hex color in enum property value — should warn
-var enumCssText = "foam.ENUM({\n  package: 'test',\n  name: 'LogLevel',\n  values: [\n    { name: 'ERROR', color: '#FF0000' }\n  ]\n})";
-var enumDiags = diagHandler.handle(enumCssText);
-test(enumDiags.some(function(d) { return /raw color/i.test(d.message); }), 'Raw CSS: hex in enum color property flagged');
-
-// 3-char hex — should warn
-var hex3CssText = "foam.CLASS({\n  package: 'test',\n  name: 'Hex3Test',\n  css: `\n    ^ { border-color: #F00; }\n  `\n})";
-var hex3Diags = diagHandler.handle(hex3CssText);
-test(hex3Diags.some(function(d) { return /raw color/i.test(d.message); }), 'Raw CSS: 3-char hex flagged');
-
-// === EXPRESSION PARAMETER VALIDATION ===
-
-section('Expression Parameter Validation');
-
-// Register test models for expression chain validation
-foam.CLASS({
-  package: 'foam.parse.lsp.test',
-  name: 'ExprParent',
-  properties: [
-    { class: 'String', name: 'title' },
-    {
-      class: 'FObjectProperty',
-      of: 'foam.parse.lsp.test.ExprChild',
-      name: 'child'
-    },
-    { class: 'Boolean', name: 'isActive' }
-  ]
-});
-
-foam.CLASS({
-  package: 'foam.parse.lsp.test',
-  name: 'ExprChild',
-  properties: [
-    { class: 'String', name: 'label' },
-    { class: 'Int', name: 'count' }
-  ]
-});
-
-// Valid simple property — no warning
-var exprValidText = "foam.CLASS({\n  package: 'foam.parse.lsp.test',\n  name: 'ExprParent',\n  properties: [\n    { class: 'String', name: 'title' },\n    { class: 'FObjectProperty', of: 'foam.parse.lsp.test.ExprChild', name: 'child' },\n    { class: 'Boolean', name: 'isActive' },\n    { name: 'computed', expression: function(title, isActive) { return title + isActive; } }\n  ]\n})";
-var exprValidDiags = diagHandler.handle(exprValidText);
-var exprTitleWarns = exprValidDiags.filter(function(d) { return d.message.indexOf("'title'") !== -1 && d.message.indexOf('does not exist') !== -1; });
-test(exprTitleWarns.length === 0, 'Expression: valid property title NOT flagged');
-
-// Invalid property name — should warn
-var exprInvalidText = "foam.CLASS({\n  package: 'foam.parse.lsp.test',\n  name: 'ExprParent',\n  properties: [\n    { class: 'String', name: 'title' },\n    { name: 'computed', expression: function(title, nonExistent) { return title; } }\n  ]\n})";
-var exprInvalidDiags = diagHandler.handle(exprInvalidText);
-test(exprInvalidDiags.some(function(d) { return d.message.indexOf('nonExistent') !== -1 && d.message.indexOf('does not exist') !== -1; }), 'Expression: invalid property nonExistent IS flagged');
-
-// Slot access suffix $ — should validate base name
-var exprSlotText = "foam.CLASS({\n  package: 'foam.parse.lsp.test',\n  name: 'ExprParent',\n  properties: [\n    { class: 'Boolean', name: 'isActive' },\n    { name: 'computed', expression: function(isActive$) { return isActive$; } }\n  ]\n})";
-var exprSlotDiags = diagHandler.handle(exprSlotText);
-var exprSlotWarns = exprSlotDiags.filter(function(d) { return d.message.indexOf('isActive') !== -1 && d.message.indexOf('does not exist') !== -1; });
-test(exprSlotWarns.length === 0, 'Expression: isActive$ (slot suffix) NOT flagged');
-
-// Deep $ chain — valid path
-var exprChainText = "foam.CLASS({\n  package: 'foam.parse.lsp.test',\n  name: 'ExprParent',\n  properties: [\n    { class: 'FObjectProperty', of: 'foam.parse.lsp.test.ExprChild', name: 'child' },\n    { name: 'computed', expression: function(child$label) { return child$label; } }\n  ]\n})";
-var exprChainDiags = diagHandler.handle(exprChainText);
-var exprChainWarns = exprChainDiags.filter(function(d) { return d.message.indexOf('does not exist') !== -1 && (d.message.indexOf('child') !== -1 || d.message.indexOf('label') !== -1); });
-test(exprChainWarns.length === 0, 'Expression: valid chain child$label NOT flagged');
-
-// Deep $ chain — invalid segment
-var exprBadChainText = "foam.CLASS({\n  package: 'foam.parse.lsp.test',\n  name: 'ExprParent',\n  properties: [\n    { class: 'FObjectProperty', of: 'foam.parse.lsp.test.ExprChild', name: 'child' },\n    { name: 'computed', expression: function(child$bogus) { return child$bogus; } }\n  ]\n})";
-var exprBadChainDiags = diagHandler.handle(exprBadChainText);
-test(exprBadChainDiags.some(function(d) { return d.message.indexOf('bogus') !== -1 && d.message.indexOf('does not exist') !== -1; }), 'Expression: invalid chain segment bogus IS flagged');
-
-// Unresolvable type — should NOT flag further segments (no false positives)
-var exprStringChainText = "foam.CLASS({\n  package: 'test',\n  name: 'StrChainTest',\n  properties: [\n    { class: 'String', name: 'title' },\n    { name: 'computed', expression: function(title$length) { return title$length; } }\n  ]\n})";
-var exprStrDiags = diagHandler.handle(exprStringChainText);
-var exprStrWarns = exprStrDiags.filter(function(d) { return d.message.indexOf('length') !== -1 && d.message.indexOf('does not exist') !== -1; });
-test(exprStrWarns.length === 0, 'Expression: unresolvable chain stops validation (no false positive on String$length)');
-
-// Multi-model file — expression params should NOT bleed across models
-var multiModelText = "foam.CLASS({\n  package: 'test',\n  name: 'ModelA',\n  properties: [\n    { class: 'String', name: 'propA' },\n    { name: 'computed', expression: function(propA) { return propA; } }\n  ]\n})\n\nfoam.CLASS({\n  package: 'test',\n  name: 'ModelB',\n  properties: [\n    { class: 'String', name: 'propB' },\n    { name: 'computed', expression: function(propB) { return propB; } }\n  ]\n})";
-var multiDiags = diagHandler.handle(multiModelText);
-var multiWarns = multiDiags.filter(function(d) { return d.message.indexOf('does not exist') !== -1 && d.message.indexOf('expression') === -1; });
-var propAOnB = multiWarns.filter(function(d) { return d.message.indexOf('propA') !== -1 && d.message.indexOf('ModelB') !== -1; });
-var propBOnA = multiWarns.filter(function(d) { return d.message.indexOf('propB') !== -1 && d.message.indexOf('ModelA') !== -1; });
-test(propAOnB.length === 0, 'Expression: propA NOT flagged against ModelB (multi-model scoping)');
-test(propBOnA.length === 0, 'Expression: propB NOT flagged against ModelA (multi-model scoping)');
-
-// === POM COMPLETIONS ===
-
-section('POM Completions');
-
-var pomText = "foam.POM({\n  na\n})";
-var pomResult = completionHandler.handle(pomText, { line: 1, character: 4 });
-var pomItems = pomResult.items || [];
-test(pomItems.some(function(it) { return it.label === 'name: '; }), 'POM: suggests name');
-
-var pomText2 = "foam.POM({\n  \n})";
-var pomResult2 = completionHandler.handle(pomText2, { line: 1, character: 2 });
-var pomItems2 = pomResult2.items || [];
-test(pomItems2.some(function(it) { return it.label === 'files: '; }), 'POM: suggests files');
-test(pomItems2.some(function(it) { return it.label === 'projects: '; }), 'POM: suggests projects');
-test(pomItems2.some(function(it) { return it.label === 'javaDependencies: '; }), 'POM: suggests javaDependencies');
-
-// === FOAM.CLASS KEY COMPLETIONS ===
-
-section('foam.CLASS Key Completions');
-
-var classText = "foam.CLASS({\n  \n})";
-var classResult = completionHandler.handle(classText, { line: 1, character: 2 });
-var classItems = classResult.items || [];
-test(classItems.some(function(it) { return it.label === 'refines: '; }), 'CLASS key: suggests refines');
-test(classItems.some(function(it) { return it.label === 'label: '; }), 'CLASS key: suggests label');
-test(classItems.some(function(it) { return it.label === 'plural: '; }), 'CLASS key: suggests plural');
-test(classItems.some(function(it) { return it.label === 'ids: '; }), 'CLASS key: suggests ids');
-test(classItems.some(function(it) { return it.label === 'javaCode: '; }), 'CLASS key: suggests javaCode');
-test(classItems.some(function(it) { return it.label === 'cssTokens: '; }), 'CLASS key: suggests cssTokens');
-
-// Existing keys still work
-test(classItems.some(function(it) { return it.label === 'package: '; }), 'CLASS key: still suggests package');
-test(classItems.some(function(it) { return it.label === 'properties: '; }), 'CLASS key: still suggests properties');
-
-// === PROPERTY KEY COMPLETIONS ===
-
-section('Property Key Completions');
-
-var propText = "foam.CLASS({\n  properties: [\n    {\n      \n    }\n  ]\n})";
-var propResult = completionHandler.handle(propText, { line: 3, character: 6 });
-var propItems = propResult.items || [];
-test(propItems.some(function(it) { return it.label === 'class: ' || it.label === 'class'; }), 'Prop key: suggests class');
-test(propItems.some(function(it) { return it.label === 'name: ' || it.label === 'name'; }), 'Prop key: suggests name');
-test(propItems.some(function(it) { return it.label === 'value: '; }), 'Prop key: suggests value');
-test(propItems.some(function(it) { return it.label === 'factory: '; }), 'Prop key: suggests factory');
-test(propItems.some(function(it) { return it.label === 'expression: '; }), 'Prop key: suggests expression');
-test(propItems.some(function(it) { return it.label === 'javaCode: '; }), 'Prop key: suggests javaCode');
-test(propItems.some(function(it) { return it.label === 'javaGetter: '; }), 'Prop key: suggests javaGetter');
-test(propItems.some(function(it) { return it.label === 'view: '; }), 'Prop key: suggests view');
-test(propItems.some(function(it) { return it.label === 'visibility: '; }), 'Prop key: suggests visibility');
-test(propItems.some(function(it) { return it.label === 'tableCellFormatter: '; }), 'Prop key: suggests tableCellFormatter');
-test(propItems.some(function(it) { return it.label === 'label: '; }), 'Prop key: suggests label');
-test(propItems.some(function(it) { return it.label === 'section: '; }), 'Prop key: suggests section');
-
-// === INNER CLASS EXPRESSION SCOPING ===
-
-section('Inner Class Expression Scoping');
-
-// Expression in inner class should validate against inner class properties, not outer
-var innerClassText = "foam.CLASS({\n  package: 'test',\n  name: 'Outer',\n  properties: [\n    { class: 'String', name: 'outerProp' }\n  ],\n  classes: [\n    {\n      name: 'Inner',\n      properties: [\n        { class: 'String', name: 'innerProp' },\n        { name: 'computed', expression: function(innerProp) { return innerProp; } }\n      ]\n    }\n  ]\n})";
-var innerDiags = diagHandler.handle(innerClassText);
-var innerExprWarns = innerDiags.filter(function(d) { return d.message.indexOf('innerProp') !== -1 && d.message.indexOf('does not exist') !== -1; });
-test(innerExprWarns.length === 0, 'Inner class: innerProp expression NOT flagged');
-
-// Expression in inner class referencing outer property should be flagged
-var innerBadText = "foam.CLASS({\n  package: 'test',\n  name: 'Outer2',\n  properties: [\n    { class: 'String', name: 'outerProp' }\n  ],\n  classes: [\n    {\n      name: 'Inner2',\n      properties: [\n        { name: 'computed', expression: function(outerProp) { return outerProp; } }\n      ]\n    }\n  ]\n})";
-var innerBadDiags = diagHandler.handle(innerBadText);
-test(innerBadDiags.some(function(d) { return d.message.indexOf('outerProp') !== -1 && d.message.indexOf('does not exist') !== -1; }), 'Inner class: outerProp expression IS flagged (wrong scope)');
-
-// Outer expression should still work
-var outerExprText = "foam.CLASS({\n  package: 'test',\n  name: 'Outer3',\n  properties: [\n    { class: 'String', name: 'outerProp' },\n    { name: 'computed', expression: function(outerProp) { return outerProp; } }\n  ],\n  classes: [\n    {\n      name: 'Inner3',\n      properties: [\n        { class: 'String', name: 'innerProp' }\n      ]\n    }\n  ]\n})";
-var outerExprDiags = diagHandler.handle(outerExprText);
-var outerExprWarns = outerExprDiags.filter(function(d) { return d.message.indexOf('outerProp') !== -1 && d.message.indexOf('does not exist') !== -1; });
-test(outerExprWarns.length === 0, 'Outer class: outerProp expression NOT flagged');
-
-// === POM VALUE COMPLETIONS ===
-
-section('POM Value Completions');
-
-// flags: '...' should suggest flag values
-var pomFlagsText = "foam.POM({\n  files: [\n    { name: 'MyFile', flags: '' }\n  ]\n})";
-var pomFlagsResult = completionHandler.handle(pomFlagsText, { line: 2, character: 30 });
-var pomFlagsItems = pomFlagsResult.items || [];
-test(pomFlagsItems.some(function(it) { return it.label === 'js|java'; }), 'POM flags: suggests js|java');
-test(pomFlagsItems.some(function(it) { return it.label === 'js'; }), 'POM flags: suggests js');
-test(pomFlagsItems.some(function(it) { return it.label === 'js&test|java&test'; }), 'POM flags: suggests js&test|java&test');
-test(pomFlagsItems.some(function(it) { return it.label === 'web'; }), 'POM flags: suggests web');
-
-// javaDependencies: ['...'] should suggest known deps
-var pomDepsText = "foam.POM({\n  javaDependencies: [\n    ''\n  ]\n})";
-var pomDepsResult = completionHandler.handle(pomDepsText, { line: 2, character: 5 });
-var pomDepsItems = pomDepsResult.items || [];
-test(pomDepsItems.length >= 0, 'POM javaDependencies: returns suggestions (count: ' + pomDepsItems.length + ')');
-
-// flags inside POM should NOT suggest class names
-test( ! pomFlagsItems.some(function(it) { return it.label && it.label.indexOf('foam.') === 0; }), 'POM flags: does NOT suggest class names');
-
-// === CURSOR SENTINEL ===
-section('CursorSentinel');
-
-var sentinelCS = foam.parse.lsp.CursorSentinel.create();
-
-var ins1 = sentinelCS.insertAt('hello world', { line: 0, character: 4 });
-test(ins1.text.indexOf(sentinelCS.CHAR) === 0, 'Sentinel replaces word under cursor at start');
-test(ins1.offset === 0, 'Sentinel offset reported correctly (word replaced)');
-
-var ins1b = sentinelCS.insertAt('hello  world', { line: 0, character: 6 });
-test(ins1b.text === 'hello ' + sentinelCS.CHAR + ' world',
-  'Sentinel inserts at whitespace-only position');
-test(ins1b.offset === 6, 'Sentinel offset at whitespace-only position is correct');
-
-var ins1c = sentinelCS.insertAt('hello world', { line: 0, character: 5 });
-test(ins1c.text === sentinelCS.CHAR + ' world',
-  'Sentinel replaces preceding word when cursor is right after it');
-
-var ins2 = sentinelCS.insertAt('foo bar baz', { line: 0, character: 5 });
-test(ins2.text === 'foo ' + sentinelCS.CHAR + ' baz', 'Sentinel replaces identifier under cursor');
-
-var ins3 = sentinelCS.insertAt('line1\nline2\nline3', { line: 1, character: 3 });
-test(ins3.text.split('\n')[1].indexOf(sentinelCS.CHAR) === 0, 'Sentinel replaces word at line+column');
-
-test(sentinelCS.CHAR.charCodeAt(0) < 32 || sentinelCS.CHAR.charCodeAt(0) > 126,
-  'Sentinel char is non-ASCII-printable so no terminal matches');
-
-test(sentinelCS.removeFrom('ab' + sentinelCS.CHAR + 'cd') === 'abcd', 'removeFrom strips sentinel');
-
-// === GRAMMAR COLLECT SUGGESTIONS AT ===
-section('FoamClassGrammar.collectSuggestionsAt');
-
-var grammarC = foam.parse.lsp.FoamClassGrammar.create({ index: index });
-var sentinelC = foam.parse.lsp.CursorSentinel.create();
-
-var srcExt = "foam.CLASS({\n  package: 'test',\n  name: 'X',\n  extends: ''\n});";
-var insExt = sentinelC.insertAt(srcExt, { line: 3, character: 12 });
-var sugsExt = grammarC.collectSuggestionsAt(insExt.text, insExt.offset);
-test(sugsExt.length > 100, 'extends: suggests many class IDs (' + sugsExt.length + ')');
-test(sugsExt.some(function(s) { return s.text === 'foam.lang.FObject'; }),
-  'extends: suggests FObject');
-
-var srcReq = "foam.CLASS({\n  package: 'test',\n  name: 'X',\n  requires: ['']\n});";
-var insReq = sentinelC.insertAt(srcReq, { line: 3, character: 14 });
-var sugsReq = grammarC.collectSuggestionsAt(insReq.text, insReq.offset);
-test(sugsReq.length > 100, 'requires: suggests many class IDs (' + sugsReq.length + ')');
-
-var srcOf = "foam.CLASS({\n  name: 'X',\n  properties: [\n    { class: 'FObjectProperty', of: '' }\n  ]\n});";
-// Line 3: `    { class: 'FObjectProperty', of: '' }` — opening of-quote is at index 37
-var insOf = sentinelC.insertAt(srcOf, { line: 3, character: 38 });
-var sugsOf = grammarC.collectSuggestionsAt(insOf.text, insOf.offset);
-test(sugsOf.length > 100, 'of: suggests many class IDs (' + sugsOf.length + ')');
-
-var srcCls = "foam.CLASS({\n  package: 'test',\n  name: 'X',\n  properties: [{ class: '' }]\n});";
-var insCls = sentinelC.insertAt(srcCls, { line: 3, character: 25 });
-var sugsCls = grammarC.collectSuggestionsAt(insCls.text, insCls.offset);
-test(sugsCls.some(function(s) { return s.text === 'String'; }),
-  'class: suggests String property type');
-test(sugsCls.some(function(s) { return s.text === 'FObjectProperty'; }),
-  'class: suggests FObjectProperty');
-
-// Valid entry — sentinel forces failure, suggestions still come back
-var srcValid = "foam.CLASS({\n  package: 'test',\n  name: 'X',\n  requires: ['foam.u2.Element']\n});";
-var insValid = sentinelC.insertAt(srcValid, { line: 3, character: 14 });
-var sugsValid = grammarC.collectSuggestionsAt(insValid.text, insValid.offset);
-test(sugsValid.length > 100,
-  'requires: with already-valid entry still returns suggestions via sentinel');
-
-// === COMPLETION (GRAMMAR-DRIVEN CONTEXT) ===
-section('Completion — grammar-driven context detection');
-
-var grammarHandler = foam.parse.lsp.handlers.CompletionHandler.create({ index: index });
-
-var extSrc = "foam.CLASS({\n  package: 'test',\n  name: 'X',\n  extends: ''\n});";
-test(grammarHandler.detectContext_(extSrc, { line: 3, character: 12 }).classRef,
-  'Grammar detects extends: empty string context');
-
-var ofSrcCtx = "foam.CLASS({\n  name: 'X',\n  properties: [\n    { class: 'FObjectProperty', of: '' }\n  ]\n});";
-test(grammarHandler.detectContext_(ofSrcCtx, { line: 3, character: 38 }).classRef,
-  'Grammar detects of: context deep inside property object');
-
-var nonCtxSrc = "foam.CLASS({\n  name: 'X',\n  documentation: ''\n});";
-test( ! grammarHandler.detectContext_(nonCtxSrc, { line: 2, character: 18 }).classRef,
-  'Grammar correctly rejects documentation: as class-ref context');
-
-var extRes = grammarHandler.handle(extSrc, { line: 3, character: 12 });
-test(extRes.items.length > 50, 'extends: completion returns many items (' + extRes.items.length + ')');
-
-// Filter with a partial to ensure specific classes surface despite 200-item cap
-var extFObjSrc = "foam.CLASS({\n  package: 'test',\n  name: 'X',\n  extends: 'foam.lang.F'\n});";
-var extFObjRes = grammarHandler.handle(extFObjSrc, { line: 3, character: 23 });
-test(extFObjRes.items.some(function(i) { return i.label === 'foam.lang.FObject'; }),
-  'extends: partial foam.lang.F surfaces FObject');
-
-var reqSrc2 = "foam.CLASS({\n  package: 'test',\n  name: 'X',\n  requires: ['']\n});";
-var reqRes = grammarHandler.handle(reqSrc2, { line: 3, character: 14 });
-test(reqRes.items.length > 50, 'requires: completion returns many items');
-
-var ofRes = grammarHandler.handle(ofSrcCtx, { line: 3, character: 38 });
-test(ofRes.items.length > 50, 'of: completion returns many items');
-
-var impSrc = "foam.CLASS({\n  name: 'X',\n  implements: ['']\n});";
-var impRes = grammarHandler.handle(impSrc, { line: 2, character: 16 });
-test(impRes.items.length > 50, 'implements: completion returns many items');
-
-// Partial value: regex fallback still fires so suggestions appear.
-var partialSrc = "foam.CLASS({\n  name: 'X',\n  extends: 'foam.u2'\n});";
-var partialRes = grammarHandler.handle(partialSrc, { line: 2, character: 19 });
-test(partialRes.items.length > 0, 'Partial extends value still returns suggestions');
-
-// === GRAMMAR CONTEXT DETECTION ===
-section('Grammar context detection (detectContext_)');
-
-var ctxHandler = foam.parse.lsp.handlers.CompletionHandler.create({ index: index });
-
-var extCtx = ctxHandler.detectContext_(
-  "foam.CLASS({\n  name: 'X',\n  extends: ''\n});",
-  { line: 2, character: 12 });
-test(extCtx.classRef, 'detectContext_: extends → classRef');
-test( ! extCtx.propKey, 'detectContext_: extends is not propKey');
-
-var propObjCtx = ctxHandler.detectContext_(
-  "foam.CLASS({ name: 'X', properties: [{  }] });",
-  { line: 0, character: 39 });
-test(propObjCtx.propKey, 'detectContext_: inside property object → propKey');
-test( ! propObjCtx.classRef, 'detectContext_: property object is not classRef');
-
-var tableCtx = ctxHandler.detectContext_(
-  "foam.CLASS({\n  name: 'X',\n  tableColumns: ['']\n});",
-  { line: 2, character: 18 });
-test(tableCtx.columnName, 'detectContext_: tableColumns value → columnName');
-
-var searchCtx = ctxHandler.detectContext_(
-  "foam.CLASS({\n  name: 'X',\n  searchColumns: ['']\n});",
-  { line: 2, character: 19 });
-test(searchCtx.columnName, 'detectContext_: searchColumns value → columnName');
-
-var docCtx = ctxHandler.detectContext_(
-  "foam.CLASS({\n  name: 'X',\n  documentation: ''\n});",
-  { line: 2, character: 18 });
-test( ! docCtx.classRef && ! docCtx.propKey && ! docCtx.columnName,
-  'detectContext_: documentation is not a structural context');
-
-// === MODELED TYPES ===
-section('CompletionItem and Diagnostic models');
-
-var CI = foam.parse.lsp.CompletionItem;
-var item = CI.create({
-  label: 'foo', kind: CI.KIND_CLASS, detail: 'a class',
-  filterText: 'foo', sortText: '!foo'
-});
-var itemLSP = item.toLSP();
-test(itemLSP.label === 'foo', 'CompletionItem.toLSP preserves label');
-test(itemLSP.kind === 7, 'CompletionItem.toLSP uses class kind 7');
-test(itemLSP['class'] === undefined, 'CompletionItem.toLSP strips FOAM class marker');
-test(itemLSP.insertText === undefined, 'CompletionItem.toLSP omits unset optional fields');
-
-var D = foam.parse.lsp.Diagnostic;
-var diag = D.create({
-  range: { start: { line: 0, character: 0 }, end: { line: 0, character: 5 } },
-  severity: D.WARNING,
-  message: 'test',
-  code: 'TEST-1',
-  fix: { title: 'Fix it', edit: {} }
-});
-var diagLSP = diag.toLSP();
-test(diagLSP.message === 'test', 'Diagnostic.toLSP preserves message');
-test(diagLSP.severity === 2, 'Diagnostic.toLSP uses WARNING=2');
-test(diagLSP.source === 'foam-lsp', 'Diagnostic.toLSP default source');
-test(diagLSP.fix === undefined, 'Diagnostic.toLSP strips fix (code-action metadata)');
-test(diagLSP['class'] === undefined, 'Diagnostic.toLSP strips FOAM class marker');
-
-// === POM GRAMMAR CONTEXT ===
-section('POM grammar context detection');
-
-var pomHandler = foam.parse.lsp.handlers.CompletionHandler.create({ index: index });
-
-var pomFlags = pomHandler.detectContext_(
-  "foam.POM({\n  files: [\n    { name: 'X', flags: '' }\n  ]\n});",
-  { line: 2, character: 25 });
-test(pomFlags.pomFlagValue, 'POM flags: detects pomFlagValue');
-
-var pomFileNameCtx = pomHandler.detectContext_(
-  "foam.POM({\n  files: [\n    { name: '' }\n  ]\n});",
-  { line: 2, character: 13 });
-test(pomFileNameCtx.pomFileName, 'POM files.name: detects pomFileName');
-
-var pomJavaFileCtx = pomHandler.detectContext_(
-  "foam.POM({\n  javaFiles: [\n    { name: '' }\n  ]\n});",
-  { line: 2, character: 13 });
-test(pomJavaFileCtx.pomJavaFileName, 'POM javaFiles.name: detects pomJavaFileName');
-
-var pomProjCtx = pomHandler.detectContext_(
-  "foam.POM({\n  projects: [\n    { name: '' }\n  ]\n});",
-  { line: 2, character: 13 });
-test(pomProjCtx.pomProjectPath, 'POM projects.name: detects pomProjectPath');
-
-var pomDepCtx = pomHandler.detectContext_(
-  "foam.POM({\n  javaDependencies: [\n    ''\n  ]\n});",
-  { line: 2, character: 5 });
-test(pomDepCtx.pomJavaDep, 'POM javaDependencies: detects pomJavaDep');
-
-// === HANDLER OUTPUT SHAPE ===
-section('Handler output — LSP wire format');
-
-var dh = foam.parse.lsp.handlers.DiagnosticsHandler.create({ index: index });
-var badText = "foam.CLASS({\n  name: 'X',\n  extends: 'not.a.real.class.at.all.NoWay'\n});";
-var diags = dh.handle(badText, 'file:///tmp/test.js');
-test(diags.length > 0, 'DiagnosticsHandler: produces diagnostics for unknown class');
-test( ! diags.some(function(d) { return d['class']; }),
-  'DiagnosticsHandler: no FOAM class marker leaks into LSP output');
-test(diags.every(function(d) { return d.range && d.message && d.severity; }),
-  'DiagnosticsHandler: every diagnostic has range/message/severity');
-
-var ch = foam.parse.lsp.handlers.CompletionHandler.create({ index: index });
-var compText = "foam.CLASS({\n  name: 'X',\n  extends: ''\n});";
-var compRes = ch.handle(compText, { line: 2, character: 12 });
-test(compRes.items.length > 0, 'CompletionHandler: produces items');
-test( ! compRes.items.some(function(i) { return i['class']; }),
-  'CompletionHandler: no FOAM class marker leaks into LSP output');
-
-// === ENUM COMPLETION ===
-section('Enum value completion (this.EnumAlias.▊)');
-
-// Find a real FOAM enum for testing
-var enumCandidates = index.getAllClassIds().filter(function(id) {
-  var cls = index.getClass(id);
-  return cls && cls.VALUES && cls.VALUES.length > 0;
-});
-if ( enumCandidates.length === 0 ) {
-  test(false, 'Expected at least one enum in registry');
-} else {
-  var enumId = enumCandidates[0];
-  var enumVals = index.getEnumValues(enumId);
-  var enumShort = enumId.split('.').pop();
-  // Line 3: `this.<EnumShort>.` — cursor right after trailing dot
-  var line3 = "this." + enumShort + ".";
-  var srcEnum = "foam.CLASS({\n  name: 'X',\n  requires: [ '" + enumId + "' ],\n" + line3 + "\n});";
-  var mh = foam.parse.lsp.handlers.MemberCompletionHandler.create({ index: index });
-  var res = mh.handle(srcEnum, { line: 3, character: line3.length });
-  var hasEnum = res.items.some(function(i) {
-    return enumVals.some(function(v) { return v.name === i.label; });
-  });
-  test(hasEnum, 'Enum completion: suggests enum values for required enum (' + enumId + ')');
-  test(res.items.every(function(i) { return i.kind === 13; }),
-    'Enum completion: all items use EnumMember kind (13)');
-}
-
-// === CSS TOKEN RESOLVER — REVERSE LOOKUP ===
-section('CSSTokenResolver.findTokenForValue');
-
-var resolver = foam.parse.lsp.CSSTokenResolver.create();
-resolver.loadFromRegistry();
-resolver.loadFromJournals();
-
-// Pick a known ColorToken and verify round-trip
-var allNames = resolver.getAllTokenNames();
-var colorName = null;
-for ( var i = 0 ; i < allNames.length ; i++ ) {
-  var info = resolver.getTokenInfo(allNames[i]);
-  if ( info && info.type === 'ColorToken' ) {
-    var val = resolver.resolveTokenValue(allNames[i]);
-    if ( val && val.charAt(0) === '#' ) { colorName = allNames[i]; break; }
-  }
-}
-if ( colorName ) {
-  var colorVal = resolver.resolveTokenValue(colorName);
-  test(resolver.findTokenForValue(colorVal) === colorName,
-    'findTokenForValue: exact-hex match returns token name (' + colorName + ')');
-  test(resolver.findTokenForValue(colorVal.toUpperCase()) === colorName,
-    'findTokenForValue: case-insensitive');
-}
-test(resolver.findTokenForValue('#deadbeef') === null,
-  'findTokenForValue: unknown color returns null');
-test(resolver.findTokenForValue(null) === null,
-  'findTokenForValue: null input returns null');
-
-// 3-char hex normalization: '#f00' → '#ff0000'
-var f00 = resolver.findTokenForValue('#ff0000');
-var f00Short = resolver.findTokenForValue('#f00');
-test(f00 === f00Short, 'findTokenForValue: 3-char hex normalizes to 6-char');
-
-// === CREATE CONTEXT (DEEP + STRINGS + COMMENTS) ===
-section('findCreateContext — method-body resilience');
-
-var ca = foam.parse.lsp.CursorAnalyzer.create();
-
-// Deep-nested: cursor 30 lines below the .create(
-var deep = [
-  "foam.CLASS({",
-  "  package: 'test',",
-  "  name: 'X',",
-  "  requires: [ 'foam.u2.Element' ],",
-  "  methods: [",
-  "    function m() {",
-  "      var e = this.Element.create({"
-];
-for ( var pad = 0 ; pad < 25 ; pad++ ) deep.push("        // filler " + pad);
-deep.push("        ");  // cursor line
-deep.push("      });");
-deep.push("    }");
-deep.push("  ]");
-deep.push("});");
-var deepText = deep.join('\n');
-var deepLines = deepText.split('\n');
-var cursorLine = 6 + 25; // last filler → cursor line right after, offset from pad
-var deepCtx = ca.findCreateContext(deepLines, cursorLine, deepText, index);
-test(deepCtx === 'foam.u2.Element',
-  'findCreateContext: resolves 30+ lines below opening .create( (old limit was 20)');
-
-// String with `{` inside shouldn't fool the scanner
-var strText = [
-  "foam.CLASS({",
-  "  requires: [ 'foam.u2.Element' ],",
-  "  methods: [ function m() {",
-  "    var msg = 'hello { weird } string';",
-  "    this.Element.create({",
-  "      foo: 'bar'",
-  "    });",
-  "  } ]",
-  "});"
-].join('\n');
-var strLines = strText.split('\n');
-// Cursor line 5 is `      foo: 'bar'` — inside the create
-var strCtx = ca.findCreateContext(strLines, 5, strText, index);
-test(strCtx === 'foam.u2.Element',
-  'findCreateContext: ignores braces inside string literals');
-
-// === RAW COLOR MESSAGE + REPLACEMENT LOGIC ===
-section('Raw color diagnostic — message + code-action replacement');
-
-var diagHandler2 = foam.parse.lsp.handlers.DiagnosticsHandler.create({
-  index: index,
-  cssTokenResolver: cssTokenResolver
-});
-
-// Pick a ColorToken whose resolved value is a hex (so we can construct a
-// source file that uses exactly that color and expect a matching-token msg).
-var ctNames = cssTokenResolver.getAllTokenNames();
-var hitToken = null, hitHex = null;
-for ( var i = 0 ; i < ctNames.length ; i++ ) {
-  var info = cssTokenResolver.getTokenInfo(ctNames[i]);
-  if ( ! info || info.type !== 'ColorToken' ) continue;
-  var v = cssTokenResolver.resolveTokenValue(ctNames[i]);
-  if ( v && /^#[0-9a-fA-F]{6}$/.test(v) ) { hitToken = ctNames[i]; hitHex = v; break; }
-}
-
-if ( hitToken ) {
-  var hitSrc = "foam.CLASS({\n  package: 'test',\n  name: 'HitColor',\n" +
-               "  css: `\n    ^ { color: " + hitHex + "; }\n  `\n})";
-  var hitDiags = diagHandler2.handle(hitSrc);
-  var withMatch = hitDiags.filter(function(d) {
-    return /raw color/i.test(d.message) && d.message.indexOf(hitHex) !== -1;
-  });
-  test(withMatch.length === 1, 'Raw color with matching token: diagnostic present');
-  test(withMatch[0].message.indexOf("'$" + hitToken + "'") !== -1,
-    'Matching token name appears in diagnostic message (not generic $primary400)');
-}
-
-// Raw color with NO matching token — different phrasing, no false recommendation
-var missSrc = "foam.CLASS({\n  package: 'test',\n  name: 'MissColor',\n" +
-              "  css: `\n    ^ { color: #deadbe; }\n  `\n})";
-var missDiags = diagHandler2.handle(missSrc);
-var missing = missDiags.filter(function(d) { return /raw color/i.test(d.message); });
-test(missing.length === 1, 'Raw color without matching token: diagnostic present');
-test(missing[0].message.indexOf('no matching') !== -1,
-  'No-match message explicitly states there is no matching token');
-test(missing[0].message.indexOf('$primary400') === -1,
-  'No-match message does NOT include a misleading example token');
-
-// === HOVER ON ^selector IN CSS BLOCK ===
-section('Hover — CSS ^selector vs same-named property');
-
-var selSrc = [
-  "foam.CLASS({",
-  "  package: 'test',",
-  "  name: 'SelTest',",
-  "  properties: [ { class: 'Boolean', name: 'centered' } ],",
-  "  css: `",
-  "    ^centered > * { align-self: center; }",
-  "  `",
-  "})"
-].join('\n');
-// Line 5: `    ^centered > * { align-self: center; }`, cursor on `centered`
-var selHover = hoverHandler.handle(selSrc, { line: 5, character: 8 });
-test(selHover != null, 'CSS ^centered: hover returned');
-test(selHover && selHover.contents.value.indexOf('Expands to') !== -1,
-  'CSS ^centered: hover explains selector expansion');
-test(selHover && selHover.contents.value.indexOf('Boolean') === -1,
-  'CSS ^centered: hover does NOT confuse it with the Boolean property');
-test(selHover && selHover.contents.value.indexOf('Not a reference') !== -1,
-  'CSS ^centered: hover clarifies it is not a property reference');
-
-// === DOCUMENT HIGHLIGHT ===
-section('DocumentHighlightHandler');
-
-var dhh = foam.parse.lsp.handlers.DocumentHighlightHandler.create();
-var dhText = "foam.CLASS({\n  properties: [\n    { class: 'String', name: 'foobar' }\n  ],\n  methods: [ function m() { return this.foobar + this.foobar; } ]\n});";
-// Cursor on 'foobar' in the method body — inside `this.foobar`
-var fhHighlights = dhh.handle(dhText, { line: 4, character: 42 });
-test(fhHighlights.length === 3,
-  'documentHighlight: finds all 3 foobar occurrences (' + fhHighlights.length + ')');
-test(fhHighlights.every(function(h) { return h.range.end.character - h.range.start.character === 6; }),
-  'documentHighlight: ranges span exactly the identifier length');
-
-// === RENAME ===
-section('RenameHandler');
-
-var rh = foam.parse.lsp.handlers.RenameHandler.create({ index: index });
-var renameSrc = "foam.CLASS({\n  extends: 'foam.lang.FObject'\n});";
-var prep = rh.prepare(renameSrc, { line: 1, character: 20 });
-test(prep !== null, 'prepareRename: returns range for a known class');
-test(prep && prep.placeholder === 'foam.lang.FObject',
-  'prepareRename: placeholder is the current class id');
-
-var prep2 = rh.prepare("foam.CLASS({\n  documentation: 'hi'\n})", { line: 1, character: 5 });
-test(prep2 === null, 'prepareRename: returns null when cursor is not on a class id');
-
-var sameWe = rh.handle(renameSrc, { line: 1, character: 20 }, 'foam.lang.FObject');
-test(sameWe === null, 'rename: returns null for same-name rename');
-
-// === REFERENCES EXPANDED ===
-section('ReferencesHandler — expanded coverage');
-
-var rfh = foam.parse.lsp.handlers.ReferencesHandler.create({ index: index });
-var fobjRefs = rfh.handle("foam.CLASS({\n  extends: 'foam.lang.FObject'\n});",
-  { line: 1, character: 20 });
-test(fobjRefs.length > 10,
-  'references: FObject has many references (subclasses + users): ' + fobjRefs.length);
-test(fobjRefs.every(function(l) { return l.uri && l.range; }),
-  'references: every location has uri and range');
-
-// References on `name: '...'` should find refs to the declared class
-var refOnNameSrc = "foam.CLASS({\n  package: 'foam.lang',\n  name: 'FObject'\n});";
-var refOnNameResolved = rfh.resolveClassAtCursor_(refOnNameSrc, { line: 2, character: 12 }, 'FObject', 'test://name');
-var refOnNameLocs = rfh.handle(refOnNameSrc, { line: 2, character: 12 }, 'test://name');
-test(refOnNameResolved === 'foam.lang.FObject',
-  'references on name: resolves to full class id (got: ' + refOnNameResolved + ')');
-test(refOnNameLocs.length > 10,
-  'references on name: value resolves via package+name (' + refOnNameLocs.length + ' refs)');
-
-// References on a property name should find the prop in the own class +
-// inheriting subclasses that reference it.
-var propRefSrc = [
-  "foam.CLASS({",
-  "  package: 'foam.lang',",
-  "  name: 'Property',",
-  "  properties: [",
-  "    { class: 'String', name: 'name' }",
-  "  ]",
-  "});"
-].join('\n');
-// Line 4: `    { class: 'String', name: 'name' }` — cursor inside 'name' value
-var propRefLocs = rfh.handle(propRefSrc, { line: 4, character: 30 }, 'test://prop');
-test(propRefLocs.length > 0,
-  'property references: finds refs to `name` property (' + propRefLocs.length + ')');
-test(propRefLocs.every(function(l) { return l.uri && l.range; }),
-  'property references: each location has uri and range');
-
-// False-positive guard: words inside comments/docs must NOT match.
-var falsePosText = [
-  "foam.CLASS({",
-  "  package: 'foam.mlang.sink',",
-  "  name: 'FakeGroupBy',",
-  "  documentation: 'the top groups based on sortOrder and their values',",
-  "  properties: [",
-  "    { class: 'Map', name: 'groups' }",
-  "  ],",
-  "  methods: [",
-  "    function m() {",
-  "      // remaining groups and includeOthers is true",
-  "      /* replace groups with only top N */",
-  "      return this.groups['k'];",
-  "    }",
-  "  ]",
-  "});"
-].join('\n');
-var fpRfh = foam.parse.lsp.handlers.ReferencesHandler.create({ index: index });
-// Cursor on the `groups` property name (line 5 `    { class: 'Map', name: 'groups' }`)
-var fpLocs = fpRfh.handle(falsePosText, { line: 5, character: 33 }, 'test://fp');
-// Expect ONLY the real refs in this file: the quoted definition + this.groups.
-// The documentation prose, line comment, and block comment mentions must be skipped.
-var fpOwnFileLocs = fpLocs.filter(function(l) { return l.uri === 'test://fp'; });
-test(fpOwnFileLocs.length === 0 /* file not on disk, won't scan */ ||
-     fpOwnFileLocs.every(function(l) {
-       // If we scanned, none of the matches should land on the documentation line
-       return l.range.start.line !== 3;
-     }),
-  'property references: documentation prose containing the name is NOT matched');
-
-// === METHOD RETURN-TYPE RESOLUTION ===
-section('FoamIndex.getMethodReturnType');
-
-// Expressions mixin covers every return pattern we care about.
-var EXPR = 'foam.mlang.Expressions';
-test(index.classExists(EXPR), 'foam.mlang.Expressions is loaded');
-
-// `return this.GroupBy.create(...)` → short name via class.requires
-test(index.getMethodReturnType(EXPR, 'GROUP_BY') === 'foam.mlang.sink.GroupBy',
-  'this.X.create pattern resolves to full class id');
-
-// `return this.Count.create()` → short name via class.requires
-test(index.getMethodReturnType(EXPR, 'COUNT') === 'foam.mlang.sink.Count',
-  'parameterless this.X.create resolves');
-
-// `return this._binary_("StartsWith", ...)` → Name via class.requires
-test(index.getMethodReturnType(EXPR, 'STARTS_WITH') === 'foam.mlang.predicate.StartsWith',
-  '_binary_ helper resolves the quoted short name');
-
-// `return this._unary_("Ref", a)` → Name via class.requires
-test(index.getMethodReturnType(EXPR, 'REF') === 'foam.mlang.expr.Ref',
-  '_unary_ helper resolves the quoted short name');
-
-// `return this._nary_("Add", arguments)` → Name via class.requires
-test(index.getMethodReturnType(EXPR, 'ADD') === 'foam.mlang.expr.Add',
-  '_nary_ helper resolves the quoted short name');
-
-// `type:` axiom on DESC → explicit
-test(index.getMethodReturnType(EXPR, 'DESC') === 'foam.mlang.order.Desc',
-  'explicit type: axiom wins (DESC declares type: foam.mlang.order.Comparator)' +
-  ' — resolver returns the concrete returned class where create is visible, else the declared type');
-
-// Methods with no discernible return → null, not a crash
-test(index.getMethodReturnType(EXPR, 'nonexistentMethod') === null,
-  'unknown method returns null');
-
-// TypeTracker integration: `var x = this.GROUP_BY(...)` typed as GroupBy
-var ttSrc = [
-  "foam.CLASS({",
-  "  package: 'test',",
-  "  name: 'TTTest',",
-  "  implements: [ 'foam.mlang.Expressions' ],",
-  "  methods: [",
-  "    function m() {",
-  "      var g = this.GROUP_BY('x', this.COUNT());",
-  "      g",           // cursor sits on this `g` to query types
-  "    }",
-  "  ]",
-  "});"
-].join('\n');
-var ttModel = { package: 'test', name: 'TTTest', extends: EXPR, requires: [ EXPR ] };
-var tt = foam.parse.lsp.TypeTracker.create();
-var types = tt.getVariableTypes(ttSrc, { line: 7, character: 6 }, ttModel, index);
-// Note: without a real compiled model the 'this.GROUP_BY' is still reachable
-// because TypeTracker queries index.getMethodReturnType on the model's classId.
-// We use EXPR as classId directly via a synthetic model:
-var syntheticModel = { package: 'foam.mlang', name: 'Expressions' };
-var types2 = tt.getVariableTypes(ttSrc, { line: 7, character: 6 }, syntheticModel, index);
-test(types2.g === 'foam.mlang.sink.GroupBy' ||
-     // fallback ok if model resolution path differs — the important check is
-     // the return-type helper itself works, covered above
-     index.getMethodReturnType('foam.mlang.Expressions', 'GROUP_BY') === 'foam.mlang.sink.GroupBy',
-  'TypeTracker uses getMethodReturnType for non-create calls');
-
-// === StringFilterView — GROUP_BY hover + .then chain ===
-section('StringFilterView — return-type resolution & .then param');
-
-// Core: implementers of mlang.Expressions should resolve GROUP_BY, COUNT, etc.
-var SFV = 'foam.u2.filter.properties.StringFilterView';
-if ( index.classExists(SFV) ) {
-  test(index.getMethodReturnType(SFV, 'GROUP_BY') === 'foam.mlang.sink.GroupBy',
-    'StringFilterView.GROUP_BY resolves via implements: Expressions');
-  test(index.getMethodReturnType(SFV, 'COUNT') === 'foam.mlang.sink.Count',
-    'StringFilterView.COUNT resolves via implements: Expressions');
-  test(index.getMethodReturnType(SFV, 'STARTS_WITH') === 'foam.mlang.predicate.StartsWith',
-    'StringFilterView.STARTS_WITH resolves via implements: Expressions');
-
-  // Hover on `GROUP_BY` inside a StringFilterView method body should include
-  // the return type. StringFilterView IS in the registry (loaded via pmake).
-  var sfvFs = require('fs');
-  var sfvPath = 'foam3/src/foam/u2/filter/properties/StringFilterView.js';
-  if ( sfvFs.existsSync(sfvPath) ) {
-    var sfvText = sfvFs.readFileSync(sfvPath, 'utf8');
-    // Find `this.GROUP_BY(` — cursor right after the `Y` of GROUP_BY
-    var idx = sfvText.indexOf('this.GROUP_BY(');
-    var ln = 0, col = 0;
-    for ( var k = 0 ; k < idx ; k++ ) {
-      if ( sfvText.charCodeAt(k) === 10 ) { ln++; col = 0; } else col++;
-    }
-    // land cursor inside `GROUP_BY`
-    var cursorChar = col + 'this.GROUP_BY'.length - 2;
-    var h = hoverHandler.handle(sfvText, { line: ln, character: cursorChar }, 'file://' + sfvPath);
-    var hv = h && h.contents && h.contents.value || '';
-    test(hv.indexOf('foam.mlang.sink.GroupBy') !== -1,
-      'Hover on this.GROUP_BY in StringFilterView shows return foam.mlang.sink.GroupBy');
-    // Regression: make sure only ONE "Returns:" line appears, even when the
-    // method axiom declares `type:` AND the code-parse resolves a concrete
-    // class. The parser-resolved concrete wins; the declared-type line is
-    // suppressed by buildMethodHover_.
-    var returnsCount = (hv.match(/Returns:/g) || []).length;
-    test(returnsCount === 1,
-      'Hover on this.GROUP_BY has exactly one `Returns:` line (got ' + returnsCount + ')');
-  }
-}
-
-// `.select(SINK).then((p) => …)` — p typed as SINK's class
-var thenSrc = [
-  "foam.CLASS({",
-  "  package: 'test',",
-  "  name: 'ChainTest',",
-  "  implements: [ 'foam.mlang.Expressions' ],",
-  "  methods: [",
-  "    function m() {",
-  "      this.dao.where(pred)",
-  "        .select(this.GROUP_BY(this.property, this.COUNT(), 21))",
-  "        .then((results) => {",
-  "          this.countByContents = results.groups;",
-  "        });",
-  "    }",
-  "  ]",
-  "});"
-].join('\n');
-var tt2 = foam.parse.lsp.TypeTracker.create();
-var chainModel = { package: 'test', name: 'ChainTest',
-  implements: [ 'foam.mlang.Expressions' ],
-  requires: [] };
-// cursor on `results.groups` (line 9, character 30-ish)
-var chainTypes = tt2.getVariableTypes(thenSrc, { line: 9, character: 30 }, chainModel, index);
-test(chainTypes.results === 'foam.mlang.sink.GroupBy',
-  '.then((results) => …) param typed from preceding .select(this.GROUP_BY(...)) (got: ' + chainTypes.results + ')');
-
-// Same with `.select(this.GroupBy.create({...}))` — direct create form
-var createChain = [
-  "foam.CLASS({",
-  "  name: 'X',",
-  "  requires: [ 'foam.mlang.sink.GroupBy' ],",
-  "  methods: [",
-  "    function m() {",
-  "      this.dao.select(this.GroupBy.create({})).then(function(r) { r.groups; });",
-  "    }",
-  "  ]",
-  "});"
-].join('\n');
-var createModel = { package: 'test', name: 'X',
-  requires: [ 'foam.mlang.sink.GroupBy' ] };
-var createTypes = tt2.getVariableTypes(createChain, { line: 5, character: 65 }, createModel, index);
-test(createTypes.r === 'foam.mlang.sink.GroupBy',
-  '.then(function(r) …) param typed from preceding .select(this.GroupBy.create({}))');
-
-// Arrow function with single-param no-parens: .then(r => r.groups)
-var bareArrow = [
-  "foam.CLASS({",
-  "  name: 'Y',",
-  "  requires: [ 'foam.mlang.sink.Count' ],",
-  "  methods: [",
-  "    function m() {",
-  "      this.dao.select(this.Count.create()).then(r => r.value);",
-  "    }",
-  "  ]",
-  "});"
-].join('\n');
-var bareModel = { package: 'test', name: 'Y', requires: [ 'foam.mlang.sink.Count' ] };
-var bareTypes = tt2.getVariableTypes(bareArrow, { line: 5, character: 55 }, bareModel, index);
-test(bareTypes.r === 'foam.mlang.sink.Count',
-  '.then(r => …) bare-arrow param typed from preceding .select');
-
-// === MESSAGE AXIOM: hover + go-to-definition ===
-section('Message axiom hover + go-to-definition');
-
-if ( index.classExists(SFV) ) {
-  var sfvFs2 = require('fs');
-  var sfvFile = 'foam3/src/foam/u2/filter/properties/StringFilterView.js';
-  var sfvTxt  = sfvFs2.readFileSync(sfvFile, 'utf8');
-
-  // FoamIndex layer
-  var allMsgs = index.getMessages(SFV);
-  test(allMsgs.length >= 7,
-    'getMessages returns StringFilterView messages (' + allMsgs.length + ')');
-  var lm = index.findMessage(SFV, 'LABEL_PLACEHOLDER');
-  test(lm && lm.message === 'Search', 'findMessage returns LABEL_PLACEHOLDER with its text');
-  test(index.findMessage(SFV, 'NOT_A_MESSAGE') === null,
-    'findMessage returns null for unknown names');
-
-  // Hover — cursor on `this.LABEL_PLACEHOLDER` inside render()
-  var hIdx = sfvTxt.indexOf('this.LABEL_PLACEHOLDER');
-  var hLine = 0, hCol = 0;
-  for ( var i = 0 ; i < hIdx ; i++ ) {
-    if ( sfvTxt.charCodeAt(i) === 10 ) { hLine++; hCol = 0; } else hCol++;
-  }
-  // land cursor inside LABEL_PLACEHOLDER (pos = after `this.`)
-  var msgHover = hoverHandler.handle(sfvTxt,
-    { line: hLine, character: hCol + 'this.'.length + 5 }, 'file://' + sfvFile);
-  var mv = msgHover && msgHover.contents && msgHover.contents.value || '';
-  test(mv.indexOf('LABEL_PLACEHOLDER') !== -1,
-    'Message hover: includes the message name');
-  test(mv.indexOf('Search') !== -1,
-    'Message hover: includes the message text');
-
-  // Definition — same cursor jumps to the `{ name: 'LABEL_PLACEHOLDER', … }` entry
-  var defHandler = foam.parse.lsp.handlers.DefinitionHandler.create({ index: index });
-  var msgDef = defHandler.handle(sfvTxt,
-    { line: hLine, character: hCol + 'this.'.length + 5 }, 'file://' + sfvFile);
-  test(msgDef && msgDef.uri && msgDef.uri.indexOf('StringFilterView.js') !== -1,
-    'Message go-to-def: lands in StringFilterView.js');
-  test(msgDef && msgDef.range && typeof msgDef.range.start.line === 'number',
-    'Message go-to-def: has a valid range');
-}
-
-// === GRAMMAR-DRIVEN AXIOM POSITIONS ===
-section('FoamClassGrammar.collectAxiomPositions');
-
-var axiomGrammar = foam.parse.lsp.FoamClassGrammar.create({ index: index });
-
-var axSrc = [
-  "foam.CLASS({",
-  "  package: 'test',",
-  "  name: 'Ax',",
-  "  documentation: 'messages: [ fake ]',",         // must NOT be picked up
-  "  messages: [",                                  // L4
-  "    { name: 'GREETING', message: 'hi' },",       // L5
-  "    { name: 'FAREWELL', message: 'bye' }",       // L6
-  "  ]",
-  "});"
-].join('\n');
-
-var axMap = axiomGrammar.collectAxiomPositions(axSrc);
-test(axMap.message.GREETING && axMap.message.GREETING.line === 5,
-  'Grammar axiom-pos: GREETING message at line 5');
-test(axMap.message.FAREWELL && axMap.message.FAREWELL.line === 6,
-  'Grammar axiom-pos: FAREWELL message at line 6');
-test(! axMap.message.fake,
-  'Grammar axiom-pos: docstring containing messages: [ fake ] is NOT indexed');
-
-// Caching
-var m1 = axiomGrammar.collectAxiomPositions(axSrc);
-var m2 = axiomGrammar.collectAxiomPositions(axSrc);
-test(m1 === m2, 'Grammar axiom-pos: cache hit on identical text');
-
-// Enum values container
-var enumSrc = [
-  "foam.ENUM({",
-  "  package: 'test',",
-  "  name: 'Color',",
-  "  values: [",                                 // L3
-  "    { name: 'RED', label: 'Red' },",          // L4
-  "    { name: 'GREEN', label: 'Green' }",       // L5
-  "  ]",
-  "});"
-].join('\n');
-var enumMap = axiomGrammar.collectAxiomPositions(enumSrc);
-test(enumMap.value.RED && enumMap.value.RED.line === 4,
-  'Grammar axiom-pos: enum value RED at line 4');
-test(enumMap.value.GREEN && enumMap.value.GREEN.line === 5,
-  'Grammar axiom-pos: enum value GREEN at line 5');
-
-// Regression: `foam.mlang.Expressions` was greedy-matched by `foam.mlang.Expr`
-// before we started sorting class-ref sugs by length-descending. Guard that
-// here so the bug can't come back silently.
-var exprCheck = axiomGrammar.collectAxiomPositions(
-  "foam.CLASS({ implements: [ 'foam.mlang.Expressions' ], messages: [ { name: 'FOO', message: 'hi' } ] });"
-);
-test(Object.keys(exprCheck.message).length === 1,
-  'Grammar parses implements: [ foam.mlang.Expressions ] without prefix-match regression');
-
-// StringFilterView cross-check — grammar MUST resolve axiom positions
-// on real files.
-if ( index.classExists(SFV) ) {
-  var sfvFs3 = require('fs');
-  var sfvTxt3 = sfvFs3.readFileSync('foam3/src/foam/u2/filter/properties/StringFilterView.js', 'utf8');
-  var sfvAxMap = axiomGrammar.collectAxiomPositions(sfvTxt3);
-
-  // Progressive isolation: test increasingly minimal versions of SFV until
-  // grammar produces results — identifies which construct breaks it.
-
-  var msgKeys = Object.keys(sfvAxMap.message);
-  test(msgKeys.length >= 7,
-    'Grammar axiom-pos on StringFilterView: all 7 messages indexed (got ' + msgKeys.length + ')');
-  test(sfvAxMap.message.LABEL_PLACEHOLDER && sfvAxMap.message.LABEL_PLACEHOLDER.line > 90,
-    'Grammar axiom-pos: LABEL_PLACEHOLDER at expected line');
-}
-
-// === POM completion: grammar-driven key + value suggestions ===
-section('POM completion — grammar drives every position');
-
-var pomCh = foam.parse.lsp.handlers.CompletionHandler.create({
-  index: index, analyzer: analyzer, cssTokenResolver: cssTokenResolver
-});
-
-var pomSrc = [
-  "foam.POM({",                                          // L0
-  "  name: 'lsp',",                                      // L1
-  "  files: [",                                          // L2
-  "    {  },",                                           // L3 — empty file object
-  "    { name: '', flags: '' }",                         // L4 — empty value strings
-  "  ]",                                                  // L5
-  "});"                                                   // L6
-].join('\n');
-
-function pomLabels(pos) {
-  var items = (pomCh.handle(pomSrc, pos) || {}).items || [];
-  return items.map(function(it) { return it.label || it.insertText; });
-}
-
-// Top-level POM body: keys like files, javaFiles, projects, name, journalFiles
-var topLabels = pomLabels({ line: 0, character: 10 });
-test(topLabels.indexOf('files: ') !== -1,
-  'POM top-level: suggests files: after foam.POM({');
-test(topLabels.indexOf('projects: ') !== -1,
-  'POM top-level: suggests projects:');
-
-// Empty file object `{ ▊ }` — grammar must emit pomKey sugs for name/flags
-var emptyObjLabels = pomLabels({ line: 3, character: 6 });
-test(emptyObjLabels.indexOf('name: ') !== -1,
-  'POM empty file object: suggests name:');
-test(emptyObjLabels.indexOf('flags: ') !== -1,
-  'POM empty file object: suggests flags:');
-
-// Inside name: '▊' — grammar context marker triggers file-name suggestions
-var nameValLabels = pomLabels({ line: 4, character: 13 });
-test(nameValLabels.length > 0,
-  'POM name value: at least one file-name suggestion');
-test(nameValLabels.some(function(l) { return /^[A-Z]\w+$/.test(l); }),
-  'POM name value: suggestions look like file names (PascalCase)');
-
-// Inside flags: '▊' — grammar context marker triggers flag values
-var flagsLabels = pomLabels({ line: 4, character: 24 });
-test(flagsLabels.indexOf('js') !== -1,
-  'POM flags value: suggests js');
-test(flagsLabels.indexOf('java') !== -1,
-  'POM flags value: suggests java');
-
-// === Migration coverage: grammar emits 'property' and 'method' positions ===
-section('Grammar: property / method axiom positions');
-
-var propMethodSrc = [
-  "foam.CLASS({",
-  "  package: 'test',",                              // L1
-  "  name: 'PM',",                                   // L2
-  "  properties: [",                                 // L3
-  "    { class: 'String', name: 'firstName' },",     // L4
-  "    { class: 'Int',    name: 'age' }",            // L5
-  "  ],",
-  "  methods: [",                                    // L7
-  "    function greet() { return 'hi'; },",          // L8
-  "    { name: 'farewell', code: function() { return 'bye'; } }",  // L9
-  "  ]",
-  "});"
-].join('\n');
-var pmMap = axiomGrammar.collectAxiomPositions(propMethodSrc);
-test(pmMap.property && pmMap.property.firstName && pmMap.property.firstName.line === 4,
-  'Grammar axiom-pos: property firstName at line 4');
-test(pmMap.property && pmMap.property.age && pmMap.property.age.line === 5,
-  'Grammar axiom-pos: property age at line 5');
-test(pmMap.method && pmMap.method.greet && pmMap.method.greet.line === 8,
-  'Grammar axiom-pos: method greet at line 8 (bare function form)');
-test(pmMap.method && pmMap.method.farewell && pmMap.method.farewell.line === 9,
-  'Grammar axiom-pos: method farewell at line 9 (object form)');
-
-// === Migration coverage: buildLocationAtProperty uses the grammar path ===
-section('DefinitionHandler — grammar-based property navigation');
-
-// Create a temp file we can point the handler at.
-var tmpFs = require('fs');
-var tmpOs = require('os');
-var tmpPath2 = require('path');
-var tmpFile = tmpPath2.join(tmpOs.tmpdir(), 'lsp-grammar-propnav.js');
-tmpFs.writeFileSync(tmpFile,
-  "foam.CLASS({\n" +
-  "  package: 'test',\n" +
-  "  name: 'PropNav',\n" +
-  "  properties: [\n" +
-  "    { class: 'String', name: 'aProp' }\n" +
-  "  ]\n" +
-  "});\n");
-try {
-  var propLoc = defHandler.buildLocationAtProperty(tmpFile, 'aProp');
-  test(propLoc && propLoc.uri && propLoc.uri.indexOf('lsp-grammar-propnav.js') !== -1,
-    'buildLocationAtProperty returns a URI for aProp');
-  test(propLoc && propLoc.range && propLoc.range.start.line === 4,
-    'buildLocationAtProperty: aProp is on line 4 (grammar-resolved)');
-} finally {
-  try { tmpFs.unlinkSync(tmpFile); } catch ( e ) {}
-}
-
-// === Migration coverage: buildLocationAtMethod uses the grammar path ===
-section('DefinitionHandler — grammar-based method navigation');
-
-var tmpFile2 = tmpPath2.join(tmpOs.tmpdir(), 'lsp-grammar-methodnav.js');
-tmpFs.writeFileSync(tmpFile2,
-  "foam.CLASS({\n" +
-  "  package: 'test',\n" +
-  "  name: 'MethodNav',\n" +
-  "  methods: [\n" +
-  "    function computeValue() { return 42; }\n" +
-  "  ]\n" +
-  "});\n");
-try {
-  var methodLoc = defHandler.buildLocationAtMethod(tmpFile2, 'test.MethodNav', 'computeValue');
-  test(methodLoc && methodLoc.uri && methodLoc.uri.indexOf('lsp-grammar-methodnav.js') !== -1,
-    'buildLocationAtMethod returns a URI for computeValue');
-  test(methodLoc && methodLoc.range && methodLoc.range.start.line === 4,
-    'buildLocationAtMethod: computeValue is on line 4 (grammar-resolved)');
-} finally {
-  try { tmpFs.unlinkSync(tmpFile2); } catch ( e ) {}
-}
-
-// === Migration coverage: LIB + POM eval recovery ===
-section('FileModelCache: foam.LIB captured via parseFileModels');
-
-var cacheInst = foam.parse.lsp.FileModelCache.create();
-var libFileSrc =
-  "foam.CLASS({ package: 'test', name: 'Together' });\n" +
-  "foam.LIB({\n" +
-  "  name: 'test.MyLib',\n" +
-  "  methods: [ function doIt(x) { return x; } ]\n" +
-  "});\n";
-var libModels = cacheInst.parseFileModels(libFileSrc);
-var libCaught = libModels.some(function(m) { return m.type_ === 'LIB' && m.name === 'test.MyLib'; });
-test(libCaught, 'parseFileModels captures foam.LIB with correct name');
-var clsCaught = libModels.some(function(m) {
-  return m.type_ !== 'LIB' && m.package === 'test' && m.name === 'Together';
-});
-test(clsCaught, 'parseFileModels still captures sibling foam.CLASS in the same file');
-
-// Syntax-error fallback still finds LIB (Phase 4: LIB added to evalIndividualBlocks_ regex)
-var brokenLibSrc =
-  "foam.CLASS({ package: 'test', name: 'BrokenSibling' });\n" +
-  "this is not valid JS + syntax\n" +
-  "foam.LIB({\n" +
-  "  name: 'test.RecoveredLib',\n" +
-  "  methods: [ function ok() { return 1; } ]\n" +
-  "});\n";
-var brokenModels = cacheInst.parseFileModels(brokenLibSrc);
-var recovered = brokenModels.some(function(m) {
-  return m.type_ === 'LIB' && m.name === 'test.RecoveredLib';
-});
-test(recovered, 'evalIndividualBlocks_ fallback recovers foam.LIB from a file with a syntax error');
-
-// === Migration coverage: FoamIndex POM eval parsers ===
-section('FoamIndex — POM eval parsers');
-
-var simplePom = "foam.POM({ name: 'p', projects: [ { name: 'test/pom', flags: 'test' } ] });";
-var parsedProjects = index.parsePomProjects_(simplePom);
-test(parsedProjects && parsedProjects.length === 1 && parsedProjects[0].name === 'test/pom',
-  'parsePomProjects_: eval returns the projects array');
-test(parsedProjects && parsedProjects[0].flags === 'test',
-  'parsePomProjects_: flags preserved through eval');
-
-var simpleFilesPom = "foam.POM({ name: 'p', files: [ { name: 'Foo', flags: 'js' } ] });";
-var parsedFiles = index.parsePomFiles_(simpleFilesPom);
-test(parsedFiles && parsedFiles.length === 1 && parsedFiles[0].name === 'Foo',
-  'parsePomFiles_: eval returns the files array');
-
-// === Migration coverage: LIB + class indexed in the same pass ===
-section('FoamIndex — unified eval-based file indexing');
-
-// foam.Color LIB is in colorlib.js alongside no other classes; index was built
-// at LSP boot so both should be present.
-test(index.getLibEntry('foam.Color') && index.getLibEntry('foam.Color').methods.indexOf('adjustAlpha') !== -1,
-  'Unified indexing: foam.Color captured via eval-intercept');
-// At least one RELATIONSHIP class must be indexed — these are only captured by eval
-// (regex approach misses them because the name is synthesized from sourceModel+targetModel).
-var relCount = 0;
-var allIds = Object.keys(index.fileIndex_ || {});
-for ( var rIdx = 0 ; rIdx < allIds.length ; rIdx++ ) {
-  if ( /Relationship$/.test(allIds[rIdx]) ) relCount++;
-}
-test(relCount >= 1,
-  'Unified indexing: RELATIONSHIP classes are indexed (got ' + relCount + ')');
-
-// === MESSAGE + CONSTANT REFERENCES ===
-section('ReferencesHandler — message & constant axioms');
-
-if ( index.classExists(SFV) ) {
-  var sfvTxt4 = require('fs').readFileSync(
-    'foam3/src/foam/u2/filter/properties/StringFilterView.js', 'utf8');
-
-  // Cursor on `name: 'LABEL_PLACEHOLDER'` inside messages: [...]
-  var lpIdx = sfvTxt4.indexOf("name: 'LABEL_PLACEHOLDER'");
-  if ( lpIdx !== -1 ) {
-    var ln = 0, col = 0;
-    for ( var i = 0 ; i < lpIdx ; i++ ) {
-      if ( sfvTxt4.charCodeAt(i) === 10 ) { ln++; col = 0; } else col++;
-    }
-    var lpRefs = rfh.handle(sfvTxt4,
-      { line: ln, character: col + "name: '".length + 3 },
-      'file://sfv-msg');
-    test(lpRefs.length >= 1,
-      'Message references: finds uses of LABEL_PLACEHOLDER (' + lpRefs.length + ')');
-    test(lpRefs.every(function(l) { return l.uri && l.range; }),
-      'Message references: every location has uri and range');
-  }
-}
-
-// Recognition helpers — decouple from registry so they test cleanly.
-test(rfh.isOwnMessageName_({ messages: [{ name: 'HI' }] }, 'HI'),
-  'axiomReferences: recognizes message name from model.messages');
-test(rfh.isOwnConstantName_({ constants: [{ name: 'FOO' }] }, 'FOO'),
-  'axiomReferences: recognizes constant name from array form');
-test(rfh.isOwnConstantName_({ constants: { BAR: 1 } }, 'BAR'),
-  'axiomReferences: recognizes constant name from object-map form');
-test(! rfh.isOwnConstantName_({}, 'BAR'),
-  'axiomReferences: model without constants returns false');
-test(! rfh.isOwnMessageName_({}, 'HI'),
-  'axiomReferences: model without messages returns false');
-
-// === JRL TRIPLE-QUOTED SERVICE SCRIPT / CLIENT ===
-section('JRL triple-quote completion (serviceScript + client)');
-
-var jrlH = foam.parse.lsp.handlers.JrlHandler.create({ index: index });
-jrlH.buildJournalClassMap();
-
-// detectTripleQuoteContext_
-var jrlTq = [
-  'p({',
-  '  "class": "foam.core.boot.CSpec",',
-  '  "name": "xDAO",',
-  '  "serviceScript": """',
-  '    return new foam.dao.EasyDAO.Builder(x).build();',
-  '  """,',
-  '  "client": """',
-  '    { "class": "foam.dao.EasyDAO", "of": "" }',
-  '  """',
-  '})'
-].join('\n');
-
-// Cursor inside serviceScript (line 4, char 15 — inside "foam.dao.Easy")
-var ssCtx = jrlH.detectTripleQuoteContext_(jrlTq, { line: 4, character: 20 });
-test(ssCtx && ssCtx.key === 'serviceScript',
-  'detectTripleQuoteContext_: identifies serviceScript');
-
-// Cursor inside client JSON (line 7, char 20)
-var clCtx = jrlH.detectTripleQuoteContext_(jrlTq, { line: 7, character: 20 });
-test(clCtx && clCtx.key === 'client',
-  'detectTripleQuoteContext_: identifies client');
-
-// Outside triple-quote (line 1, inside "name")
-var outCtx = jrlH.detectTripleQuoteContext_(jrlTq, { line: 2, character: 12 });
-test(outCtx === null,
-  'detectTripleQuoteContext_: returns null outside triple-quote');
-
-// serviceScript completion — dotted prefix yields class-id matches
-var ssSrc = [
-  'p({',
-  '  "serviceScript": """',
-  '    return foam.dao.',
-  '  """',
-  '})'
-].join('\n');
-var ssRes = jrlH.handleCompletion(ssSrc, { line: 2, character: 20 });
-test(ssRes.items.length > 5,
-  'serviceScript completion: dotted foam.dao. prefix yields class-id matches (' + ssRes.items.length + ')');
-test(ssRes.items.every(function(i) { return i.label.indexOf('foam.dao.') === 0; }),
-  'serviceScript completion: all suggestions start with foam.dao.');
-
-// serviceScript completion — member-access on a resolved receiver derives
-// its suggestions from the registry (no hardcoded names). Uses
-// `foam.dao.EasyDAO` since it's universally present.
-if ( index.classExists('foam.dao.EasyDAO') ) {
-  var memberSrc = [
-    'p({',
-    '  "serviceScript": """',
-    '    foam.dao.EasyDAO.set',
-    '  """',
-    '})'
-  ].join('\n');
-  var memberRes = jrlH.handleCompletion(memberSrc, { line: 2, character: 23 });
-  // We expect at least one setter surfaced from the class's own properties.
-  test(memberRes.items.some(function(i) {
-    return /^set\w+/.test(i.label);
-  }), 'serviceScript completion: member-access on EasyDAO surfaces registry-derived setters');
-}
-
-// client completion — delegation to nested JRL completion. The inner
-// JSON gets treated as a JRL entry; `"class": "…"` should suggest classes.
-var clientSrc = [
-  'p({',
-  '  "client": """{ "class": "" }"""',
-  '})'
-].join('\n');
-// Line 1 `  "client": """{ "class": "" }"""` — cursor at the empty class value (char 28)
-var clientRes = jrlH.handleCompletion(clientSrc, { line: 1, character: 28 });
-test(Array.isArray(clientRes.items),
-  'client completion: returns an items array (delegated to JRL completion)');
-
-// === JRL EMBEDDED BLOCK VARIANTS (triple + escaped) ===
-section('JRL embedded block — triple-quote + escaped double-quote');
-
-var jrlH2 = foam.parse.lsp.handlers.JrlHandler.create({ index: index });
-jrlH2.buildJournalClassMap();
-
-// Escaped double-quoted client: "client": "{\"of\":\"com.paytic.Transaction\"}"
-var escSrc = [
-  'p({',
-  '  "class": "foam.core.boot.CSpec",',
-  '  "name": "txDAO",',
-  '  "client": "{\\"of\\":\\"com.paytic.Transaction\\"}"',
-  '})'
-].join('\n');
-// Cursor inside the empty space after `"of":"` (line 3)
-var escLine = '  "client": "{\\"of\\":\\"com.paytic.Transaction\\"}"';
-// Test detection fires for escaped form
-var escCtx = jrlH2.detectEmbeddedBlockContext_(escSrc, { line: 3, character: 30 });
-test(escCtx && escCtx.key === 'client' && escCtx.escaped === true,
-  'detectEmbeddedBlockContext_: detects escaped client form');
-test(escCtx && escCtx.content.indexOf('"of":"com.paytic.Transaction"') !== -1,
-  'detectEmbeddedBlockContext_: unescapes the content correctly');
-
-// Triple-quote client still works through the unified detector
-var tripleClient = [
-  'p({',
-  '  "client": """',
-  '    { "of": "com.paytic.Transaction" }',
-  '  """',
-  '})'
-].join('\n');
-var triCtx = jrlH2.detectEmbeddedBlockContext_(tripleClient, { line: 2, character: 20 });
-test(triCtx && triCtx.key === 'client' && triCtx.escaped === false,
-  'detectEmbeddedBlockContext_: still detects triple-quote client form');
-
-// Triple-quote serviceScript detection
-var tripleSS = [
-  'p({',
-  '  "serviceScript": """',
-  '    return new foam.dao.EasyDAO.Builder(x).build();',
-  '  """',
-  '})'
-].join('\n');
-var ssCtx2 = jrlH2.detectEmbeddedBlockContext_(tripleSS, { line: 2, character: 20 });
-test(ssCtx2 && ssCtx2.key === 'serviceScript',
-  'detectEmbeddedBlockContext_: detects triple-quote serviceScript form');
-
-// === JRL TEXTMATE HIGHLIGHTS (Java + JSON injections) ===
-section('JRL grammar injections — Java & JSON');
-
-var jrlGrammarPath = 'foam3/tools/lsp/editors/vscode/syntaxes/foam-jrl.tmLanguage.json';
-var jrlGrammar = JSON.parse(require('fs').readFileSync(jrlGrammarPath, 'utf8'));
-
-test(!! jrlGrammar.repository['java-block-triple'],
-  'JRL grammar: java-block-triple injection exists');
-test(!! jrlGrammar.repository['java-block-backtick'],
-  'JRL grammar: java-block-backtick injection exists');
-test(!! jrlGrammar.repository['java-block-single-quoted'],
-  'JRL grammar: java-block-single-quoted injection exists');
-test(!! jrlGrammar.repository['json-block-triple'],
-  'JRL grammar: json-block-triple injection exists (client """…""")');
-test(!! jrlGrammar.repository['json-block-backtick'],
-  'JRL grammar: json-block-backtick injection exists');
-
-// The `object` repository entry (reachable from top-level patterns) includes
-// the new JSON injections.
-var objPatterns = JSON.stringify(jrlGrammar.repository.object.patterns);
-test(objPatterns.indexOf('json-block-triple') !== -1,
-  'JRL grammar: object patterns reference json-block-triple');
-test(objPatterns.indexOf('json-block-backtick') !== -1,
-  'JRL grammar: object patterns reference json-block-backtick');
-
-// === JRL EMBEDDED-BLOCK HOVERS (full coverage) ===
-section('JRL embedded-block hovers — serviceScript + client, both forms');
-
-var jrlH3 = foam.parse.lsp.handlers.JrlHandler.create({ index: index });
-jrlH3.buildJournalClassMap();
-
-// Hover on a FOAM class id inside triple-quoted serviceScript.
-// Use a class guaranteed to be in the registry (foam.dao.EasyDAO).
-if ( index.classExists('foam.dao.EasyDAO') ) {
-  var ssH = [
-    'p({',
-    '  "class": "foam.core.boot.CSpec",',
-    '  "name": "x",',
-    '  "serviceScript": """',
-    '    return new foam.dao.EasyDAO.Builder(x).build();',
-    '  """',
-    '})'
-  ].join('\n');
-  // Cursor on `foam.dao.EasyDAO` — line 4, around char 26
-  var h1 = jrlH3.handleHover(ssH, { line: 4, character: 26 });
-  test(h1 && h1.contents && h1.contents.value &&
-       h1.contents.value.indexOf('foam.dao.EasyDAO') !== -1,
-    'Hover on class id in triple-quoted serviceScript resolves via registry');
-}
-
-// Hover on a class id inside triple-quoted client JSON
-if ( index.classExists('foam.dao.EasyDAO') ) {
-  var cliH = [
-    'p({',
-    '  "name": "x",',
-    '  "client": """',
-    '    { "class": "foam.dao.EasyDAO", "of": "foam.lang.FObject" }',
-    '  """',
-    '})'
-  ].join('\n');
-  // Cursor on the "foam.dao.EasyDAO" value — line 3, around char 24
-  var h2 = jrlH3.handleHover(cliH, { line: 3, character: 24 });
-  test(h2 && h2.contents && h2.contents.value &&
-       h2.contents.value.indexOf('foam.dao.EasyDAO') !== -1,
-    'Hover on class id in triple-quoted client JSON resolves via registry');
-}
-
-// Hover on a class id inside ESCAPED-double-quote client
-if ( index.classExists('foam.dao.EasyDAO') ) {
-  var escH = [
-    'p({',
-    '  "name": "x",',
-    '  "client": "{\\"class\\":\\"foam.dao.EasyDAO\\"}"',
-    '})'
-  ].join('\n');
-  // Cursor on EasyDAO — the escaped line is:
-  //   "client": "{\"class\":\"foam.dao.EasyDAO\"}"
-  // Position inside `foam.dao.EasyDAO` — around char 36 on line 2
-  var h3 = jrlH3.handleHover(escH, { line: 2, character: 36 });
-  test(h3 && h3.contents && h3.contents.value &&
-       h3.contents.value.indexOf('foam.dao.EasyDAO') !== -1,
-    'Hover on class id in escaped-double-quote client resolves via registry');
-}
-
-// Completion at a dotted prefix works in both forms of serviceScript,
-// regardless of whether the body happens to look Java-ish or JS-ish.
-// Simple one-liner (JS-like)
-var simpleScript = [
-  'p({',
-  '  "serviceScript": """',
-  '    return foam.dao.',
-  '  """',
-  '})'
-].join('\n');
-var c1 = jrlH3.handleCompletion(simpleScript, { line: 2, character: 20 });
-test(c1.items.length > 5,
-  'serviceScript completion: simple `return foam.dao.` yields class-id matches');
-
-// Java-like builder chain
-var complexScript = [
-  'p({',
-  '  "serviceScript": """',
-  '    return new foam.dao.EasyDAO.Builder(x).setOf(com.paytic.',
-  '  """',
-  '})'
-].join('\n');
-var c2 = jrlH3.handleCompletion(complexScript, { line: 2, character: 62 });
-test(c2.items.length >= 0 /* zero is acceptable when no classes match com.paytic in this env */,
-  'serviceScript completion: Java-style builder chain still returns an items array');
-
-// Escaped-double-quote serviceScript (rarer but valid)
-var escScript = [
-  'p({',
-  '  "serviceScript": "return x.get(\\"transactionDAO\\");"',
-  '})'
-].join('\n');
-var escCtxSS = jrlH3.detectEmbeddedBlockContext_(escScript, { line: 1, character: 30 });
-test(escCtxSS && escCtxSS.key === 'serviceScript' && escCtxSS.escaped === true,
-  'detectEmbeddedBlockContext_: detects escaped-double-quote serviceScript');
-
-// === EMBEDDED-BLOCK: dot-after-class completion + method hover ===
-section('JRL embedded-block: typing dot after known class should surface setters (not classes)');
-
-if ( index.classExists('foam.dao.EasyDAO') ) {
-  // Typing `foam.dao.EasyDAO.Builder(x).` — cursor right after the trailing
-  // dot. Expected: setters/getters from foam.dao.EasyDAO, NOT the full
-  // class-id list (which is what the user reported seeing).
-  var dotSrc = [
-    'p({',
-    '  "serviceScript": """',
-    '    return new foam.dao.EasyDAO.Builder(x).',
-    '  """',
-    '})'
-  ].join('\n');
-  // Line 2 is `    return new foam.dao.EasyDAO.Builder(x).`; cursor at EOL (char 44)
-  var line2Len = dotSrc.split('\n')[2].length;
-  var dotRes = jrlH3.handleCompletion(dotSrc, { line: 2, character: line2Len });
-  // At least one setter must be present, and NOT all suggestions should be full class IDs.
-  var hasSetter = dotRes.items.some(function(i) {
-    return /^set\w/.test(i.label);
-  });
-  var allClassIds = dotRes.items.every(function(i) {
-    return i.label.indexOf('.') !== -1;
-  });
-  test(hasSetter,
-    'Dot after Builder(x).: surfaces at least one setter (' +
-    dotRes.items.slice(0, 5).map(function(i) { return i.label; }).join(',') + ')');
-  test(! allClassIds,
-    'Dot after Builder(x).: suggestions are not just full class IDs');
-
-  // Hover on a setter name like `.setPm(` — cursor on 'setPm' itself.
-  var setH = [
-    'p({',
-    '  "serviceScript": """',
-    '    return new foam.dao.EasyDAO.Builder(x).setPm(true).build();',
-    '  """',
-    '})'
-  ].join('\n');
-  // Line 2: `    return new foam.dao.EasyDAO.Builder(x).setPm(true).build();`
-  //  chars  0123456789012345678901234567890123456789012345 — `setPm` starts at 43
-  var setLine = setH.split('\n')[2];
-  var setPmIdx = setLine.indexOf('setPm');
-  var hSet = jrlH3.handleHover(setH, { line: 2, character: setPmIdx + 2 });
-  // If EasyDAO doesn't have a `pm` property we can't resolve `setPm` as a
-  // setter — but the resolver should still return something useful for
-  // setters that DO map to a property. Try `setOf` which is universal.
-  var setOfIdx = setLine.indexOf('setOf');
-  if ( setOfIdx === -1 ) {
-    // Different test with setOf explicitly
-    var setOfSrc = [
-      'p({',
-      '  "serviceScript": """',
-      '    return new foam.dao.EasyDAO.Builder(x).setOf(foam.lang.FObject.getOwnClassInfo());',
-      '  """',
-      '})'
-    ].join('\n');
-    var setOfLine = setOfSrc.split('\n')[2];
-    setOfIdx = setOfLine.indexOf('setOf');
-    var hOf = jrlH3.handleHover(setOfSrc, { line: 2, character: setOfIdx + 2 });
-    test(hOf && hOf.contents && hOf.contents.value &&
-         /setOf|of/i.test(hOf.contents.value),
-      'Hover on .setOf( (maps to `of` property) inside serviceScript shows member info');
-  } else {
-    var hOf = jrlH3.handleHover(setH, { line: 2, character: setOfIdx + 2 });
-    test(hOf && hOf.contents && hOf.contents.value &&
-         /setOf|of/i.test(hOf.contents.value),
-      'Hover on .setOf( inside serviceScript shows member info');
-  }
-}
-
-// === EMBEDDED-BLOCK SEMANTIC TOKENS ===
-section('JRL embedded-block semantic tokens — Java + client JSON');
-
-// Java serviceScript with a dotted FOAM class id inside.
-var embedJava = [
-  'p({',
-  '  "class": "foam.core.boot.CSpec",',
-  '  "id": "myService",',
-  '  "serviceScript": """',
-  '    return new foam.dao.EasyDAO.Builder(x).setOf(foam.lang.FObject.getOwnClassInfo()).build();',
-  '  """',
-  '})'
-].join('\n');
-var javaTokens = jrlH3.handleSemanticTokens(embedJava);
-// data format: [dL, dC, len, type, mods, …]. Expect at least one type=0
-// token (class reference) somewhere inside the serviceScript body.
-function hasTokenOfType(data, type) {
-  for ( var i = 3 ; i < data.length ; i += 5 ) {
-    if ( data[i] === type ) return true;
-  }
-  return false;
-}
-test(hasTokenOfType(javaTokens.data, 0),
-  'Embedded Java block emits type (0) tokens for registered class IDs');
-
-// Verify class:"..." at top level is tagged as class (type=1).
-test(hasTokenOfType(javaTokens.data, 1),
-  'Top-level "class":"…" value still emitted as class (1) token');
-
-// JSON client block — verified "class":"…" inside the nested JSON.
-var embedClient = [
-  'p({',
-  '  "class": "foam.core.boot.CSpec",',
-  '  "id": "myCSpec",',
-  '  "client": """',
-  '    {',
-  '      "class": "foam.dao.EasyDAO",',
-  '      "of": "foam.lang.FObject"',
-  '    }',
-  '  """',
-  '})'
-].join('\n');
-var clientTokens = jrlH3.handleSemanticTokens(embedClient);
-// Expect TWO class tokens: top-level foam.core.boot.CSpec AND nested foam.dao.EasyDAO.
-var classTokenCount = 0;
-for ( var i = 3 ; i < clientTokens.data.length ; i += 5 ) {
-  if ( clientTokens.data[i] === 1 ) classTokenCount++;
-}
-test(classTokenCount >= 2,
-  'Nested "class":"…" inside triple-quoted client block emits class token (got ' + classTokenCount + ')');
-
-// Escaped-in-double-quote client form.
-var embedClientEsc = [
-  'p({',
-  '  "class": "foam.core.boot.CSpec",',
-  '  "id": "myCSpec",',
-  '  "client": "{\\"class\\":\\"foam.dao.EasyDAO\\",\\"of\\":\\"foam.lang.FObject\\"}"',
-  '})'
-].join('\n');
-var escTokens = jrlH3.handleSemanticTokens(embedClientEsc);
-var escClassCount = 0;
-for ( var i = 3 ; i < escTokens.data.length ; i += 5 ) {
-  if ( escTokens.data[i] === 1 ) escClassCount++;
-}
-test(escClassCount >= 2,
-  'Escaped "class":"…" inside escaped-double-quote client block emits class token (got ' + escClassCount + ')');
-
-// Unknown class inside serviceScript should NOT emit a type token.
-var embedUnknown = [
-  'p({',
-  '  "class": "foam.core.boot.CSpec",',
-  '  "serviceScript": """',
-  '    return totally.not.a.real.Class.foo();',
-  '  """',
-  '})'
-].join('\n');
-var unkTokens = jrlH3.handleSemanticTokens(embedUnknown);
-// Should only have the top-level class token (type=1), no type=0 from the unknown ID.
-var unkTypeCount = 0;
-for ( var i = 3 ; i < unkTokens.data.length ; i += 5 ) {
-  if ( unkTokens.data[i] === 0 ) unkTypeCount++;
-}
-test(unkTypeCount === 0,
-  'Unknown dotted identifier in serviceScript does not emit type token (got ' + unkTypeCount + ')');
-
-// === ZED TREE-SITTER — JSON INJECTION FOR client BLOCKS ===
-section('Zed tree-sitter grammar: JSON injection for client blocks');
-var fs_ = require('fs');
-var path_ = require('path');
-var zedInj = fs_.readFileSync(path_.join(__dirname, '../lsp/editors/zed-foam3/languages/jrl/injections.scm'), 'utf8');
-test(/injection\.language\s+"json"/.test(zedInj),
-  'Zed JRL injections.scm declares JSON injection');
-test(/#eq\?\s+@_key\s+"client"/.test(zedInj),
-  'Zed JRL injections.scm matches `client` key for JSON injection');
-// VS Code grammar parity.
-var vscodeJrl = JSON.parse(fs_.readFileSync(path_.join(__dirname, '../lsp/editors/vscode/syntaxes/foam-jrl.tmLanguage.json'), 'utf8'));
-test(!! vscodeJrl.repository['json-block-triple'] && !! vscodeJrl.repository['json-block-backtick'],
-  'VS Code foam-jrl grammar has JSON injections for client triple/backtick');
-
-// === MULTI-LINE BUILDER CHAIN (real-world services.jrl shape) ===
-section('Multi-line builder chain — hover + enum completion');
-var mlsrc = [
-  'p({',
-  '  "class":"foam.core.boot.CSpec",',
-  '  "name":"transactionDAO",',
-  '  "serviceScript": """',
-  '    return new foam.dao.EasyDAO.Builder(x)',
-  '      .setPm(true)',
-  '      .setSeqNo(true)',
-  '      .setJournalType(foam.dao.JournalType.SINGLE_JOURNAL)',
-  '      .setOf(foam.lang.FObject.getOwnClassInfo())',
-  '      .build();',
-  '  """',
-  '})'
-].join('\n');
-
-// Hover on .setSeqNo — receiver is on line 4 (the return new ... Builder(x)),
-// but .setSeqNo is on line 6. Walk-back must cross line boundaries.
-var linesMl = mlsrc.split('\n');
-var seqLine = 6; // .setSeqNo
-var seqCol = linesMl[seqLine].indexOf('setSeqNo') + 2;
-var seqH = jrlH3.handleHover(mlsrc, { line: seqLine, character: seqCol });
-test(seqH && seqH.contents && seqH.contents.value && /setSeqNo|property|seqNo/i.test(seqH.contents.value),
-  'Multi-line: hover on .setSeqNo resolves to EasyDAO setter (receiver on prior line)');
-
-// Hover on .setJournalType (line 7, two chained setters + Builder above)
-var jtLine = 7;
-var jtCol = linesMl[jtLine].indexOf('setJournalType') + 2;
-var jtH = jrlH3.handleHover(mlsrc, { line: jtLine, character: jtCol });
-test(jtH && jtH.contents && jtH.contents.value && /setJournalType|journalType|property/i.test(jtH.contents.value),
-  'Multi-line: hover on .setJournalType resolves to EasyDAO setter (chain with multiple prior setters)');
-
-// Hover on SINGLE_JOURNAL — should be enum value hover.
-var enumLine = 7;
-var enumCol = linesMl[enumLine].indexOf('SINGLE_JOURNAL') + 2;
-var enumH = jrlH3.handleHover(mlsrc, { line: enumLine, character: enumCol });
-if ( index.classExists('foam.dao.JournalType') ) {
-  test(enumH && enumH.contents && enumH.contents.value && /JournalType\.SINGLE_JOURNAL|enum value/i.test(enumH.contents.value),
-    'Hover on foam.dao.JournalType.SINGLE_JOURNAL shows enum value info');
-}
-
-// Completion at `foam.dao.JournalType.` — must surface enum values FIRST,
-// not all classes starting with foam.dao.JournalType.
-if ( index.classExists('foam.dao.JournalType') ) {
-  var enumCompSrc = [
-    'p({',
-    '  "serviceScript": """',
-    '    return new foam.dao.EasyDAO.Builder(x).setJournalType(foam.dao.JournalType.',
-    '  """',
-    '})'
-  ].join('\n');
-  var enumCompLine = 2;
-  var enumCompCol = enumCompSrc.split('\n')[enumCompLine].length;
-  var enumComp = jrlH3.handleCompletion(enumCompSrc, { line: enumCompLine, character: enumCompCol });
-  var hasSingleJournal = enumComp.items.some(function(it) { return it.label === 'SINGLE_JOURNAL'; });
-  test(hasSingleJournal,
-    'Completion after `foam.dao.JournalType.` offers enum values (SINGLE_JOURNAL found: ' +
-    enumComp.items.slice(0, 4).map(function(i) { return i.label; }).join(',') + ')');
-  // First item must be an enum value (kind 20), not a class
-  var firstIsEnum = enumComp.items.length > 0 && enumComp.items[0].kind === 20;
-  test(firstIsEnum,
-    'First completion item is an enum value (kind=20), not a class');
-}
-
-// === CHAINED BUILDER SETTER HOVERS ===
-section('Chained builder setter hovers — walk back through .a(x).b(y) chains');
-var chained = [
-  'p({',
-  '  "class":"foam.core.boot.CSpec",',
-  '  "id":"myDAO",',
-  '  "serviceScript":"""',
-  '    return new foam.dao.EasyDAO.Builder(x).setPm(true).setOf(foam.lang.FObject.getOwnClassInfo()).build();',
-  '  """',
-  '})'
-].join('\n');
-var chainedLine = chained.split('\n')[4];
-// setOf comes AFTER setPm(true). in the chain. Walk-back must skip the
-// intermediate call to resolve to foam.dao.EasyDAO.
-['setPm', 'setOf'].forEach(function(name) {
-  var idx = chainedLine.indexOf(name);
-  var hover = jrlH3.handleHover(chained, { line: 4, character: idx + 2 });
-  test(hover && hover.contents && hover.contents.value &&
-       /void ' + name + '|property/i.test(hover.contents.value.replace(/[|]/g, '')),
-    'Hover on .' + name + '( in chained builder — resolves to EasyDAO setter');
-});
-
-// === REAL services.jrl sanity check ===
-section('Real services.jrl hover sanity');
-var realJrlPath = require('path').resolve(__dirname, '../../../journals/services.jrl');
-if ( require('fs').existsSync(realJrlPath) ) {
-  var realText = require('fs').readFileSync(realJrlPath, 'utf8');
-  var realLines = realText.split('\n');
-
-  // Find the first .setPm( occurrence and hover on it.
-  for ( var rl = 0 ; rl < realLines.length ; rl++ ) {
-    var m = realLines[rl].indexOf('.setPm(');
-    if ( m === -1 ) continue;
-    var h = jrlH3.handleHover(realText, { line: rl, character: m + 3 });
-    test(h && h.contents && h.contents.value && /setPm|pm/i.test(h.contents.value),
-      'Real services.jrl: hover on .setPm at line ' + rl + ' resolves (setPm or pm in output)');
-    break;
-  }
-
-  // First .setOf( across the entire file.
-  for ( var rl2 = 0 ; rl2 < realLines.length ; rl2++ ) {
-    var m2 = realLines[rl2].indexOf('.setOf(');
-    if ( m2 === -1 ) continue;
-    var h2 = jrlH3.handleHover(realText, { line: rl2, character: m2 + 3 });
-    test(h2 && h2.contents && h2.contents.value && /setOf|of /i.test(h2.contents.value),
-      'Real services.jrl: hover on .setOf at line ' + rl2 + ' resolves');
-    break;
-  }
-
-  // First `foam.dao.JournalType.SINGLE_JOURNAL` reference.
-  for ( var rl3 = 0 ; rl3 < realLines.length ; rl3++ ) {
-    var m3 = realLines[rl3].indexOf('JournalType.SINGLE_JOURNAL');
-    if ( m3 === -1 ) continue;
-    // Hover on SINGLE_JOURNAL
-    var sj = realLines[rl3].indexOf('SINGLE_JOURNAL');
-    var h3 = jrlH3.handleHover(realText, { line: rl3, character: sj + 2 });
-    test(h3 && h3.contents && h3.contents.value && /SINGLE_JOURNAL|enum value|ordinal/i.test(h3.contents.value),
-      'Real services.jrl: hover on SINGLE_JOURNAL enum value resolves');
-    break;
-  }
-}
-
-// === SAVE → TARGETED REANALYZE ===
-section('Targeted reanalyze: getAffectedFiles covers the dependency closure');
-
-// FObject is the mother class — every FOAM class should be affected.
-// We just want to sanity-check the API and ordering. Using a mid-level
-// class keeps the set reasonable.
-var startId = 'foam.dao.EasyDAO';
-if ( index.classExists(startId) ) {
-  var affected = index.getAffectedFiles([startId]);
-  test(Array.isArray(affected),
-    'getAffectedFiles returns an array');
-  // The saved file's own path should be in the set.
-  var selfPath = index.getFilePath(startId);
-  test(selfPath && affected.indexOf(selfPath) !== -1,
-    'Affected set includes the saved file itself');
-  // It should NOT include every file in the workspace — narrower than full scan.
-  test(affected.length < index.getAllClassIds().length / 2,
-    'Affected set (' + affected.length + ') is a small fraction of total classes (' +
-    index.getAllClassIds().length + ')');
-
-  // A subclass's file should be in the set if any subclass exists.
-  var subs = index.getSubclasses(startId);
-  if ( subs.length > 0 ) {
-    var subPath = index.getFilePath(subs[0]);
-    if ( subPath ) {
-      test(affected.indexOf(subPath) !== -1,
-        'Affected set includes subclass file ' + subPath);
-    }
-  }
-}
-
-// analyzeFiles runs diagnostics on the supplied files only
-var analyzer = foam.parse.lsp.handlers.WorkspaceAnalyzer.create({ index: index });
-var anyFilePath = startId && index.getFilePath(startId);
-if ( anyFilePath ) {
-  var singleRes = analyzer.analyzeFiles([anyFilePath]);
-  test(singleRes.filesScanned === 1,
-    'analyzeFiles({[path]}) scans exactly one file');
-  test(typeof singleRes.fileResults === 'object',
-    'analyzeFiles returns fileResults map');
-}
-
-// === LSP #4999 Fix 1: property-type completion inserts full path (except foam.lang.*) ===
-section('CompletionHandler — property-type full-path insertion (issue #4999)');
-
-// Empty value: class: '▊' — suggestions should include both a foam.lang.* short
-// name (e.g., 'String') and a non-lang full path (e.g., 'foam.u2.ViewSpec').
-var classEmptyText = 'foam.CLASS({\n  properties: [\n    { class: ' + Q + Q + ' }\n  ]\n})';
-var classEmptyPos = { line: 2, character: classEmptyText.split('\n')[2].indexOf(Q) + 1 };
-var classEmptyRes = completionHandler.handle(classEmptyText, classEmptyPos);
-
-function findItem(items, pred) {
-  for ( var i = 0 ; i < items.length ; i++ ) if ( pred(items[i]) ) return items[i];
-  return null;
-}
-function insertedText(item) {
-  return (item && item.textEdit && item.textEdit.newText) || (item && item.insertText) || item && item.label;
-}
-
-var stringItem = findItem(classEmptyRes.items, function(i) { return i.label === 'String'; });
-test(stringItem && insertedText(stringItem) === 'String',
-  'foam.lang.String: label and inserted text both short (got ' + insertedText(stringItem) + ')');
-
-var viewSpecItem = findItem(classEmptyRes.items, function(i) { return i.label === 'ViewSpec'; });
-test(viewSpecItem && insertedText(viewSpecItem) === 'foam.u2.ViewSpec',
-  'foam.u2.ViewSpec: inserted text is full path (got ' + insertedText(viewSpecItem) + ')');
-
-// Partial value: class: 'foam.u2.V▊' — replacement should still be full path.
-var classPartialText = 'foam.CLASS({\n  properties: [\n    { class: ' + Q + 'foam.u2.V' + Q + ' }\n  ]\n})';
-var classPartialPos = { line: 2, character: classPartialText.split('\n')[2].indexOf('V') + 1 };
-var classPartialRes = completionHandler.handle(classPartialText, classPartialPos);
-var viewSpecPartial = findItem(classPartialRes.items, function(i) { return i.label === 'ViewSpec'; });
-test(viewSpecPartial && insertedText(viewSpecPartial) === 'foam.u2.ViewSpec',
-  'Partial foam.u2.V: ViewSpec still replaces with full path (got ' + insertedText(viewSpecPartial) + ')');
-
-// Grammar-level check: the propType sug carries the correct insert text so
-// sug-driven paths stay consistent with the contextFallback path.
-var sentinelForProp = foam.parse.lsp.CursorSentinel.create();
-var propSentinelSrc = 'foam.CLASS({ properties: [ { class: ' + Q + Q + ' } ] })';
-var propSentinelPos = { line: 0, character: propSentinelSrc.indexOf(Q) + 1 };
-var propIns = sentinelForProp.insertAt(propSentinelSrc, propSentinelPos);
-var propTypeSugs = grammar.collectSuggestionsAt(propIns.text, propIns.offset);
-// Scan every sug's text so we don't depend on which specific prop types the
-// grammar's cursor-window happens to return at this offset.
-var langShorts = {};
-index.getPropertyTypes().forEach(function(t) {
-  if ( t.id && t.id.indexOf('foam.lang.') === 0 ) langShorts[t.name] = true;
-});
-var shortLeaks = propTypeSugs.filter(function(s) {
-  return s.category === 'property' && s.text.indexOf('.') === -1 && ! langShorts[s.text];
-});
-test(shortLeaks.length === 0,
-  'Grammar sug: no non-lang property types emit short-name insert text (leaks=' + shortLeaks.length + ')');
-var fullPathSugs = propTypeSugs.filter(function(s) {
-  return s.category === 'property' && s.text.indexOf('.') !== -1;
-});
-test(fullPathSugs.length > 0,
-  'Grammar sug: at least one non-lang property type emits a full-path insert text (' + fullPathSugs.length + ')');
-
-// === LSP #4999 Fix 2: view: '...' offers view-class suggestions ===
-section('CompletionHandler — view: class-ref suggestions (issue #4999)');
-
-// Empty string value for view: — should list class ids including views.
-var viewEmptyText = 'foam.CLASS({\n  properties: [\n    { name: ' + Q + 'x' + Q + ', view: ' + Q + Q + ' }\n  ]\n})';
-var viewEmptyPos = { line: 2, character: viewEmptyText.split('\n')[2].lastIndexOf(Q) };
-var viewEmptyRes = completionHandler.handle(viewEmptyText, viewEmptyPos);
-test(viewEmptyRes.items.length > 0,
-  'view: empty string gets class suggestions (' + viewEmptyRes.items.length + ' items)');
-test(viewEmptyRes.items.some(function(i) { return /^foam\.u2\./.test(i.label); }),
-  'view: suggestions include foam.u2.* classes');
-
-// Partial value — view: 'foam.u2.' — suggestions should still include views.
-var viewPartialText = 'foam.CLASS({\n  properties: [\n    { name: ' + Q + 'x' + Q + ', view: ' + Q + 'foam.u2.' + Q + ' }\n  ]\n})';
-var viewPartialPos = { line: 2, character: viewPartialText.split('\n')[2].lastIndexOf(Q) };
-var viewPartialRes = completionHandler.handle(viewPartialText, viewPartialPos);
-test(viewPartialRes.items.some(function(i) { return /^foam\.u2\./.test(i.label); }),
-  'view: partial foam.u2. still offers view classes (' + viewPartialRes.items.length + ' items)');
-
-// === LSP #4999 Fix 3: exports: [...] suggests model axiom names ===
-section('CompletionHandler — exports axiom names (issue #4999)');
-
-// Model with a requires: [...] block above exports: [...] — this is the
-// shape that previously produced bogus foam.comics.* suggestions because
-// the 10-line requires lookback hijacked the classRef fallback.
-var exportText =
-  'foam.CLASS({\n' +
-  "  package: 'test',\n" +
-  "  name: 'MyExporter',\n" +
-  "  requires: [\n" +
-  "    'foam.u2.View',\n" +
-  "    'foam.u2.ViewSpec'\n" +
-  "  ],\n" +
-  "  exports: [\n" +
-  "    'as asController',\n" +
-  "    " + Q + Q + "\n" +
-  "  ],\n" +
-  "  properties: [\n" +
-  "    { name: 'alpha' },\n" +
-  "    { name: 'beta' }\n" +
-  "  ],\n" +
-  "  methods: [\n" +
-  "    function refresh() {}\n" +
-  "  ],\n" +
-  "  actions: [\n" +
-  "    { name: 'reload' }\n" +
-  "  ],\n" +
-  "  listeners: [\n" +
-  "    { name: 'onUpdate' }\n" +
-  "  ]\n" +
-  "})";
-
-var exportLines = exportText.split('\n');
-// Target line 9 (1-indexed 10) which is the empty '' slot inside exports.
-var exportLineIdx = -1;
-for ( var li = 0 ; li < exportLines.length ; li++ ) {
-  if ( exportLines[li].indexOf("    ''") === 0 ) { exportLineIdx = li; break; }
-}
-test(exportLineIdx !== -1, 'Exports test setup: found empty-string line at index ' + exportLineIdx);
-var exportPos = { line: exportLineIdx, character: exportLines[exportLineIdx].indexOf(Q) + 1 };
-var exportRes = completionHandler.handle(exportText, exportPos);
-
-test(exportRes.items.length > 0,
-  'exports: empty slot returns completions (' + exportRes.items.length + ' items)');
-test(exportRes.items.every(function(i) { return i.label.indexOf('foam.') !== 0; }),
-  'exports: no class-id (foam.*) leakage from the requires block above');
-test(exportRes.items.some(function(i) { return i.label === 'alpha'; }),
-  'exports: own property "alpha" is suggested');
-test(exportRes.items.some(function(i) { return i.label === 'beta'; }),
-  'exports: own property "beta" is suggested');
-test(exportRes.items.some(function(i) { return i.label === 'refresh'; }),
-  'exports: own method "refresh" is suggested');
-test(exportRes.items.some(function(i) { return i.label === 'reload'; }),
-  'exports: own action "reload" is suggested');
-test(exportRes.items.some(function(i) { return i.label === 'onUpdate'; }),
-  'exports: own listener "onUpdate" is suggested');
-
-// Grammar-level: the exportName context marker fires at an empty-string slot.
-var exportCtx = completionHandler.detectContext_(exportText, exportPos);
-test(exportCtx.exportName === true, 'detectContext_: exportName flag set inside exports array');
-test(exportCtx.classRef === false, 'detectContext_: classRef flag NOT set inside exports array');
-
-// === SUMMARY ===
-
-section('SUMMARY');
-console.error(passes + ' passed, ' + failures + ' failed');
-process.exit(failures > 0 ? 1 : 0);
+var h = require('./lsp/_harness');
+
+// Each require() runs its category's tests against the shared harness. Order
+// is roughly "building blocks first" so a failure in the grammar surfaces
+// before the downstream handlers that depend on it.
+require('./lsp/foamIndex');
+require('./lsp/grammar');
+require('./lsp/utilities');
+require('./lsp/completion');
+require('./lsp/hover');
+require('./lsp/diagnostics');
+require('./lsp/navigation');
+require('./lsp/java');
+require('./lsp/jrl');
+
+h.section('SUMMARY');
+console.error(h.counters.passes + ' passed, ' + h.counters.failures + ' failed');
+process.exit(h.counters.failures > 0 ? 1 : 0);
